% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{AgeAndLengthBasedCatchCurvesCalcs}
\alias{AgeAndLengthBasedCatchCurvesCalcs}
\title{Calculations for length and age and lengths-based catch curves}
\usage{
AgeAndLengthBasedCatchCurvesCalcs(
  params,
  DistnType,
  lbnd,
  ubnd,
  midpt,
  MaxAge,
  NatMortType,
  NatMort,
  NatMort_sd,
  TimeStep,
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  CVSizeAtAge,
  CatchCurveType,
  MLL,
  SelectivityType,
  SelectivityAtLen,
  RetenAtLen,
  DiscMort
)
}
\arguments{
\item{params}{vector of model parameters in log space (params) to be estimated}

\item{DistnType}{1 = Multinomial, 2 = Dirichlet multinomial}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{MaxAge}{maximum age considered in model}

\item{NatMortType}{1 = fixed, 2 = estimated with prior to allow for uncertainty}

\item{NatMort}{natural mortality}

\item{NatMort_sd}{standard deviation for natural mortality in log space. Set to NA if NatMortType = 1}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{GrowthCurveType}{1 = von Bertalanffy, 2 = Schnute}

\item{GrowthParams}{c(Linf, vbK, CVSizeAtAge) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge),
both sexes von Bertalanffy, or c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b), both sexes Schnute}

\item{RefnceAges}{reference ages for Schnute function (set to NA if growth based on another function)}

\item{CVSizeAtAge}{coefficient of variation (CV), common across all mean lengths at age}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params,  3=logistic selectivity + retention params}

\item{SelectivityAtLen}{selectivity at length}

\item{RetenAtLen}{retention at length}

\item{DiscMort}{proportion of fish that die following catch and release}
}
\value{
gear selectivity at length (SelAtLength), selectivity of landings at length (SelLandAtLength),
 mean size at age from growth curve (MeanSizeAtAge), retention probability at length (RetAtLength),
mid points of length classes (midpt), mean length after 1 year of growth given initial length of midpt
(MeanEndingLength), associated change in length after 1 year of growth (TimestepGrowthSizeInc), length distribution
of 1+ year old recruits (RecLenDist), retained catch proportions at each length for females, males and sexes combined
(ExpRetCatchPropInLenClass, ExpRetCatchPropInLenClass_Fem, ExpRetCatchPropInLenClass_Mal), expected retained catches at integer
age classes for females, males and sexes combined (RetCatchAtIntAge, RetCatchAtIntAge_Fem, RetCatchAtIntAge_Mal), retained
catch proportions at integer age classes for females, males and sexes combined (ExpRetCatchPropAtIntAge, ExpRetCatchPropAtIntAge_Fem, ExpRetCatchPropAtIntAge_Mal),
relative catches at length and age (Catch), expected catch proportions at length given integer age class for females, males and sexes combined
(ExpRetCatchPropLengthGivenIntAge, ExpRetCatchPropLengthGivenIntAge_Fem, ExpRetCatchPropLengthGivenIntAge_Mal), penalty values
for L50 and L95 selectivity parameters (L50_Pen, L95_Pen), growth model type, 1=single sex von Bertalanffy, 2=2 sex
von Bertalanffy, 3=1 sex Schnute, 4=2 sex Schnute (GrowthModelType), expected total (retained plus discarded) catches at
integer age and length (TotCatchAtIntAgeLen, TotCatchAtIntAgeLen_Fem, TotCatchAtIntAgeLen_Mal), retained catches at
length (RetCatchAtIntAgeLen, RetCatchAtIntAgeLen_Fem, RetCatchAtIntAgeLen_Mal), discarded catches at length (DiscCatchAtIntAgeLen,
DiscCatchAtIntAgeLen_Fem, DiscCatchAtIntAgeLen_Mal), total catches (discarded plus retained) at length (TotCatchAtLen,
TotCatchAtLen_Fem, TotCatchAtLen_Mal), retained catches at length (RetCatchAtLen, RetCatchAtLen_Fem, RetCatchAtLen_Mal),
discarded catches at length (DiscCatchAtLen, DiscCatchAtLen_Fem, RetCatchAtLen_Mal), fishing mortality
at length, for sexes combined or separate (FAtLen, FAtLen_Fem, FAtLen_Mal), total mortality at length, for sexes combined
or separate (ZAtLen, ZAtLen_Fem, ZAtLen_Mal), fishing mortality associated with landings at length, for sexes combined or
separate (FAtLen, FAtLen_Fem, FAtLen_Mal), fishing mortality associated with capture and retention at length, for sexes combined or
separate (FAtLenReten, FAtLenReten_Fem, FAtLenReten_Mal), fishing mortality associated with capture and discarding at
length, for sexes combined or separate (FAtLenDisc, FAtLenDisc_Fem, FAtLenDisc_Mal)
}
\description{
This function undertakes a range of calculations required by the length and age and lengths-based catch curves
}
\keyword{internal}
