% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcAsymptoticLengthFromSchnuteParams}
\alias{CalcAsymptoticLengthFromSchnuteParams}
\title{Calculate asymptotic length from Schnute growth parameters}
\usage{
CalcAsymptoticLengthFromSchnuteParams(
  GrowthModelType,
  Res,
  GrowthParams,
  RefnceAges
)
}
\arguments{
\item{GrowthModelType}{1=length-based von Bertalanffy, 2=length-based Schnute,
3=age and length based von Bertalanffy, 4=age and length based Schnute}

\item{Res}{Schnute parameters (output from GetGrowthModelType function)}

\item{GrowthParams}{Schnute growth parameters for 2 sex catch curve models}

\item{RefnceAges}{Schnute growth reference ages for 2 sex catch curve models}
}
\value{
Linf asymptotic length parameter
}
\description{
Calculate asymptotic length from Schnute growth parameters, for use in penalty function
for selectivity parameters
}
\keyword{internal}
