% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcCatchAgePropGivenLen}
\alias{CalcCatchAgePropGivenLen}
\title{Calculate catch proportions within length classes at age}
\usage{
CalcCatchAgePropGivenLen(
  nTimeSteps,
  nLenCl,
  midpt,
  ExpRetCatchPropLengthGivenDecAge_Fem,
  ExpRetCatchPropAtDecAge_Fem,
  ExpRetCatchPropLengthGivenDecAge_Mal,
  ExpRetCatchPropAtDecAge_Mal,
  ExpDiscCatchPropLengthGivenDecAge_Fem,
  ExpDiscCatchPropAtDecAge_Fem,
  ExpDiscCatchPropLengthGivenDecAge_Mal,
  ExpDiscCatchPropAtDecAge_Mal,
  MLL,
  RetenParams
)
}
\arguments{
\item{nTimeSteps}{number of model timesteps}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{ExpRetCatchPropLengthGivenDecAge_Fem}{expected female retained catches at length given decimal age}

\item{ExpRetCatchPropAtDecAge_Fem}{expected female retained catches at decimal age}

\item{ExpRetCatchPropLengthGivenDecAge_Mal}{expected male retained catches at length given decimal age}

\item{ExpRetCatchPropAtDecAge_Mal}{expected male retained catches at decimal age}

\item{ExpDiscCatchPropLengthGivenDecAge_Fem}{expected female discarded catches at length given decimal age}

\item{ExpDiscCatchPropAtDecAge_Fem}{expected female discarded catches at decimal age}

\item{ExpDiscCatchPropLengthGivenDecAge_Mal}{expected male discarded catches at length given decimal age}

\item{ExpDiscCatchPropAtDecAge_Mal}{expected male discarded catches at decimal age}

\item{MLL}{minimum legal length}

\item{RetenParams}{retention parameters}
}
\value{
ExpRetCatchPropDecAgeGivenLength_Fem, ExpRetCatchPropDecAgeGivenLength_Mal, ExpDiscCatchPropDecAgeGivenLength_Fem,
ExpDiscCatchPropDecAgeGivenLength_Mal
}
\description{
Calculate retained and discarded catch proportions within each of the length classes, for each decimal age
}
\keyword{internal}
