% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcCatchAgePropGivenLen_DynSimMod_ALB}
\alias{CalcCatchAgePropGivenLen_DynSimMod_ALB}
\title{Calculate expected annual retained and discarded catch proportions at length given age, for dynamic simulation model}
\usage{
CalcCatchAgePropGivenLen_DynSimMod_ALB(
  nYears,
  nTimeSteps,
  nLenCl,
  midpt,
  ExpRetCatchPropLengthGivenDecAge_Fem,
  ExpRetCatchPropAtDecAge_Fem,
  ExpRetCatchPropLengthGivenDecAge_Mal,
  ExpRetCatchPropAtDecAge_Mal,
  ExpDiscCatchPropLengthGivenDecAge_Fem,
  ExpDiscCatchPropAtDecAge_Fem,
  ExpDiscCatchPropLengthGivenDecAge_Mal,
  ExpDiscCatchPropAtDecAge_Mal,
  ret_L50
)
}
\arguments{
\item{nYears}{number of length classes}

\item{nTimeSteps}{number of mode time steps per year}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{ExpRetCatchPropLengthGivenDecAge_Fem}{Expected female retained catch proportions at length given age}

\item{ExpRetCatchPropAtDecAge_Fem}{Expected female retained catch proportions at age}

\item{ExpRetCatchPropLengthGivenDecAge_Mal}{Expected male retained catch proportions at length given age}

\item{ExpRetCatchPropAtDecAge_Mal}{Expected male retained catch proportions at age}

\item{ExpDiscCatchPropLengthGivenDecAge_Fem}{Expected female discarded catch proportions at length given age}

\item{ExpDiscCatchPropAtDecAge_Fem}{Expected female discarded catch proportions at age}

\item{ExpDiscCatchPropLengthGivenDecAge_Mal}{Expected male discarded catch proportions at length given age}

\item{ExpDiscCatchPropAtDecAge_Mal}{Expected male discarded catch proportions at age}

\item{ret_L50}{length at 50 percent retention}
}
\value{
ExpRetCatchPropDecAgeGivenLength_Fem, ExpRetCatchPropDecAgeGivenLength_Mal,
ExpDiscCatchPropDecAgeGivenLength_Fem, ExpDiscCatchPropDecAgeGivenLength_Mal
}
\description{
Calculates expected annual retained and discarded proportions at length given age. Outputs used for generating random
catches at age from a from a dynamic simulation model.
}
\keyword{internal}
