% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcCatchLenPropGivenAge}
\alias{CalcCatchLenPropGivenAge}
\title{Calculate catch proportions within length classes at age}
\usage{
CalcCatchLenPropGivenAge(
  nTimeSteps,
  nLenCl,
  midpt,
  RetCatchAtDecAgeLen_Fem,
  RetCatchAtDecAgeLen_Mal,
  DiscCatchAtDecAgeLen_Fem,
  DiscCatchAtDecAgeLen_Mal
)
}
\arguments{
\item{nTimeSteps}{number of model timesteps}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{RetCatchAtDecAgeLen_Fem}{expected female retained catches at length and decimal age}

\item{RetCatchAtDecAgeLen_Mal}{expected male retained catches at length and decimal age}

\item{DiscCatchAtDecAgeLen_Fem}{expected female discarded catches at length and decimal age}

\item{DiscCatchAtDecAgeLen_Mal}{expected male discarded catches at length and decimal age}
}
\value{
ExpRetCatchPropLengthGivenDecAge_Fem, ExpRetCatchPropLengthGivenDecAge_Mal, ExpDiscCatchPropLengthGivenDecAge_Fem,
ExpDiscCatchPropLengthGivenDecAge_Mal
}
\description{
Calculate retained and discarded catch proportions within each of the length classes, for each decimal age
}
\keyword{internal}
