% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcCatchLenPropGivenAge_DynSimMod_ALB}
\alias{CalcCatchLenPropGivenAge_DynSimMod_ALB}
\title{Calculate expected annual retained and discarded catch proportions at length, given age, for dynamic simulation model}
\usage{
CalcCatchLenPropGivenAge_DynSimMod_ALB(
  nTimeSteps,
  nLenCl,
  midpt,
  RetCatch_Fem_yal,
  RetCatch_Mal_yal,
  DiscCatch_Fem_yal,
  DiscCatch_Mal_yal
)
}
\arguments{
\item{nTimeSteps}{number of mode time steps per year}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{RetCatch_Fem_yal}{expected retained catch numbers of females in each year, at age and and length class}

\item{RetCatch_Mal_yal}{expected retained catch numbers of males in each year, at age and and length class}

\item{DiscCatch_Fem_yal}{expected discarded catch numbers of females in each year, at age and and length class}

\item{DiscCatch_Mal_yal}{expected discarded catch numbers of males in each year, at age and and length class}
}
\value{
ExpRetCatchPropLengthGivenDecAge_Fem, ExpRetCatchPropLengthGivenDecAge_Mal,
ExpDiscCatchPropLengthGivenDecAge_Fem, ExpDiscCatchPropLengthGivenDecAge_Mal
}
\description{
Calculates expected annual retained and discarded catch proportions at length, given age.
Outputs are used for generating random catches at length and age from a from a dynamic simulation model.
}
\keyword{internal}
