% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcDirMultNLL_LogisticCatchCurve}
\alias{CalcDirMultNLL_LogisticCatchCurve}
\title{Return Dirchlet multinomial negative log-likelihood for an age-based catch curve}
\usage{
CalcDirMultNLL_LogisticCatchCurve(
  params,
  nSexes,
  Ages,
  ObsAgeFreq,
  ExpPropAtAge,
  F_Pen
)
}
\arguments{
\item{params}{model parameters log(c(FMort, SelA50, SelDelta) for multinomial NLL or
log(c(FMort, SelA50, SelDelta, theta) for Dirichlet multinomial NLL}

\item{nSexes}{number of sexes}

\item{Ages}{ages considered in the analysis}

\item{ObsAgeFreq}{observed age frequency data}

\item{ExpPropAtAge}{expected proportions at age, given parameter values}

\item{F_Pen}{value of penalty function for fishing mortality}
}
\value{
negative log-likelihood (NLL) for Dirichlet multinomial distribution
}
\description{
This function returns the Dirichlet multinomial negative log-likelihood, given observed age
frequency data and expected proportions at age, given parameters, for a catch curve with
age-based, logistic selectivity.
}
\keyword{internal}
