% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcDirMultNLLMargLengthComposition}
\alias{CalcDirMultNLLMargLengthComposition}
\title{Get NLL for marginal length composition data, assuming Dirichlet multinomial distribution}
\usage{
CalcDirMultNLLMargLengthComposition(
  ObsRetCatchFreqAtLen,
  ExpRetCatchPropInLenClass,
  DM_theta
)
}
\arguments{
\item{ObsRetCatchFreqAtLen}{observed frequencies of fish in length classes}

\item{DM_theta}{parameter theta from Dirichlet multinomial distribution}

\item{ExpCatchPropInLenClass}{observed catch proportions in length classes}
}
\value{
negative log-likelihood,assuming Dirichlet multinomial distribution (Length_NLL)
}
\description{
Get NLL for marginal length composition data, assuming Dirichlet multinomial distribution
}
\keyword{internal}
