% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcEquilibriumRecruitment}
\alias{CalcEquilibriumRecruitment}
\title{Calculate relative equilibrium recruitment}
\usage{
CalcEquilibriumRecruitment(
  SRrel_Type,
  Steepness,
  FishSpBiom,
  UnfishSpBiom,
  Eq_FertRate
)
}
\arguments{
\item{SRrel_Type}{stock recruitment relationship 1=Beverton-Holt, 2=Ricker}

\item{Steepness}{steepness of the stock-recruitmnet relationship}

\item{FishSpBiom}{fished spawning biomass per recruit (females for gonochoristic species, combined sex for hermaphroditic species)}

\item{UnfishSpBiom}{fished spawning biomass per recruit (females for gonochoristic species, combined sex for hermaphroditic species)}

\item{Eq_FertRate}{(NA or from 0.2-1) NA = no effect, 0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects}
}
\value{
Eq_Rec
}
\description{
This function calculates expected relative equilibrium recruitment from stock-recruitment
relationship, given steepness, and fished and unfished spawning biomass per recruitment
}
\keyword{internal}
