% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcExpRetCatchPropLengthGivenIntAge}
\alias{CalcExpRetCatchPropLengthGivenIntAge}
\title{Calculate expected retained catch proportions in length classes, within integer age classes}
\usage{
CalcExpRetCatchPropLengthGivenIntAge(
  MinAge,
  MaxAge,
  midpt,
  RetCatchAtIntAge,
  RetCatchAtIntAgeLen
)
}
\arguments{
\item{MinAge}{minimum age}

\item{MaxAge}{maximum age considered by model}

\item{RetCatchAtIntAge}{expected retained catches at integer age classes}

\item{RetCatchAtIntAgeLen}{expected retained catches at integer age classes and length classes}
}
\value{
ExpRetCatchPropLengthGivenIntAge
}
\description{
Calculate expected retained catch proportions in length classes, within integer age classes, used in age and
length based catch curve analyses
}
\keyword{internal}
