% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcFishingMortalityAtAge}
\alias{CalcFishingMortalityAtAge}
\title{Calculate fishing mortality at age (discards and landings)}
\usage{
CalcFishingMortalityAtAge(
  FMort,
  DiscMort,
  FemSelDiscAtAge,
  MalSelDiscAtAge,
  FemSelLandAtAge,
  MalSelLandAtAge
)
}
\arguments{
\item{FMort}{fully-selected fishing mortality}

\item{DiscMort}{proportion of discarded fish that die}

\item{FemSelDiscAtAge}{selectivity of discarding at age for females}

\item{MalSelDiscAtAge}{selectivity of discarding at age for males}

\item{FemSelLandAtAge}{selectivity of landings at age for females}

\item{MalSelLandAtAge}{selectivity of landings at age for males}
}
\value{
FemDiscFAtAge, MalDiscFAtAge, FemLandFAtAge, MalLandFAtAge, FemFAtAge, MalFAtAge
}
\description{
Calculate fishing mortality at age associated with discarding, landings, and
combined.
}
\keyword{internal}
