% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcFishingMortalityAtLen}
\alias{CalcFishingMortalityAtLen}
\title{Calculate fishing mortality at length (discards and landings)}
\usage{
CalcFishingMortalityAtLen(
  FMort,
  DiscMort,
  FemSelDiscAtLen,
  MalSelDiscAtLen,
  FemSelLandAtLen,
  MalSelLandAtLen
)
}
\arguments{
\item{FMort}{fully-selected fishing mortality}

\item{DiscMort}{proportion of discarded fish that die}

\item{FemSelDiscAtLen}{selectivity of discarding at length for females}

\item{MalSelDiscAtLen}{selectivity of discarding at length for males}

\item{FemSelLandAtLen}{selectivity of landings at length for females}

\item{MalSelLandAtLen}{selectivity of landings at length for males}
}
\value{
FemDiscFAtLen, MalDiscFAtLen, FemLandFAtLen, MalLandFAtLen, FemFAtLen, MalFAtLen
}
\description{
Calculate fishing mortality at the mid point of each length class, associated with discarding, landings, and
combined.
}
\keyword{internal}
