% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcGillnetSelectivity}
\alias{CalcGillnetSelectivity}
\title{Calculate gillnet selectivity}
\usage{
CalcGillnetSelectivity(theta1, theta2, MeshSize_mm, nMeshes, nLenCl, midpt)
}
\arguments{
\item{theta1}{parameter of Kirkwood and Walker (1986) model}

\item{theta2}{parameter of Kirkwood and Walker (1986) model}

\item{MeshSize_mm}{vector of gillnet mesh sizes, in mm}

\item{nMeshes}{number of gillnet meshes}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}
}
\value{
Selectivity at length for each mesh (SelAtLengthForMesh), overall length-based
selectivity (SelAtLength)
}
\description{
Calculate gillnet selectivity at length, by mesh and overall (assuming equal fishing intensity
among meshes) using the method of Kirkwood and Walker (1986), given model parameter values
(theta1 and theta2)
}
\examples{
MaxLen = 1000
LenInc = 20
lbnd = seq(0,MaxLen - LenInc, LenInc)
midpt = lbnd + (LenInc/2)
nLenCl = length(midpt)
MeshSize_mm = c(115, 127)
theta1 = 6
theta2 = 3000
nMeshes = length(MeshSize_mm)
CalcGillnetSelectivity(theta1, theta2, MeshSize_mm, nMeshes, nLenCl, midpt)
}
