% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcLengthAfterGrowthForTimetep}
\alias{CalcLengthAfterGrowthForTimetep}
\title{Calculate expected length after year of growth}
\usage{
CalcLengthAfterGrowthForTimetep(
  GrowthCurveType,
  TimeStep,
  GrowthParamsForSex,
  RefnceAgesForSex,
  midpt
)
}
\arguments{
\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParamsForSex}{c(Linf,vbK) von Bertalanffy, or c(t1,t2,y1,y2,a,b) Schnute}

\item{RefnceAgesForSex}{c(t1,t2)}

\item{midpt}{mid points of length classes}
}
\value{
Expected lengths (ExpLen)
}
\description{
Calculate expected length after year of growth, given initial length (from von Bertalanffy or Schnute function)
}
\keyword{internal}
