% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcLengthAtAge}
\alias{CalcLengthAtAge}
\title{Calculates length at age for each sex for age-based per recruit analysis}
\usage{
CalcLengthAtAge(Linf, vbK, tzero, EstLenAtAge, Ages)
}
\arguments{
\item{Linf}{von Bertalanffy growth parameter}

\item{vbK}{von Bertalanffy growth parameter}

\item{tzero}{von Bertalanffy growth parameter}

\item{EstLenAtAge}{lengths at age, inputted as vector}

\item{Ages}{}
}
\value{
FemLenAtAge

MalLenAtAge
}
\description{
This function calculates mean length at age for each sex, based on von Bertalanffy growth curve or
allows for these to be used as vectors inputted by user, for age-based per recruit analysis
}
\keyword{internal}
