% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcLogisticSelectivity}
\alias{CalcLogisticSelectivity}
\title{Calculate logistic length-based selectivity}
\usage{
CalcLogisticSelectivity(L50, L95, nLenCl, midpt)
}
\arguments{
\item{L50}{length at which 50 percent of fish are selected into the fishery}

\item{L95}{length at which 95 percent of fish are selected into the fishery}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}
}
\value{
Selectivity at length (SelAtLength)
}
\description{
Calculate logistic length-based selectivity (asymptotic curve)
}
\keyword{internal}
