% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcMeanSizeAtAge}
\alias{CalcMeanSizeAtAge}
\title{Calculates mean size at age for each sex for length-based per recruit analysis}
\usage{
CalcMeanSizeAtAge(GrowthCurveType, MaxModelAge, Ages, GrowthParams, RefnceAges)
}
\arguments{
\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute, 3=annual increment input}

\item{MaxModelAge}{maximum age considered by model}

\item{Ages}{ages considered by model}

\item{GrowthParams}{growth parameters of either von Bertalanffy, Schnute model or inputs from tagging analysis}

\item{RefnceAges}{reference ages for Schnute model, set to NA if using von Bertalanffy model}
}
\value{
MeanSizeAtAge
}
\description{
This function calculates mean size at age for each sex applying the von Bertalanffy growth function or Schnute function,
for length-based per recruit analysis,
}
\keyword{internal}
