% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcMultNLLMargLengthComposition}
\alias{CalcMultNLLMargLengthComposition}
\title{Get NLL for marginal length composition data, assuming multinomial distribution}
\usage{
CalcMultNLLMargLengthComposition(
  ObsRetCatchFreqAtLen,
  ExpRetCatchPropInLenClass
)
}
\arguments{
\item{ObsRetCatchFreqAtLen}{observed frequencies of fish in length classes}

\item{ExpCatchPropInLenClass}{observed catch proportions in length classes}
}
\value{
negative log-likelihood, assuming multinomial distribution (Length_NLL)
}
\description{
Get NLL for marginal length composition data, assuming multinomial distribution
}
\keyword{internal}
