% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcNLLMargLengthComposition}
\alias{CalcNLLMargLengthComposition}
\title{Get NLL for marginal length composition data}
\usage{
CalcNLLMargLengthComposition(ObsCatchFreqAtLen, ExpCatchPropInLenClass)
}
\arguments{
\item{ObsCatchFreqAtLen}{observed frequencies of fish in length classes}

\item{ExpCatchPropInLenClass}{observed catch proportions in length classes}
}
\value{
negative log-likelihood (Length_NLL)
}
\description{
Get NLL for marginal length composition data
}
\keyword{internal}
