% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcObjFunc_AgeAndLengthBasedCatchCurve}
\alias{CalcObjFunc_AgeAndLengthBasedCatchCurve}
\title{Get NLL for age and length-based catch curve}
\usage{
CalcObjFunc_AgeAndLengthBasedCatchCurve(params)
}
\arguments{
\item{estimated}{parameters, including fishing mortality, growth and logistic selectivity parameters}
}
\value{
negative log-likelihood (NLL)
}
\description{
Get NLL for age and length-based catch curve
}
