% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcObjFunc_LengthBasedCatchCurve}
\alias{CalcObjFunc_LengthBasedCatchCurve}
\title{Get NLL for length based catch curve}
\usage{
CalcObjFunc_LengthBasedCatchCurve(params)
}
\arguments{
\item{params}{estimated parameters, including fishing mortality and logistic selectivity parameters when estimated}
}
\value{
negative log-likelihood (NLL)
}
\description{
Get NLL for length based catch curve
}
