% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPerRecruitFishLenAndFishWtStats}
\alias{CalcPerRecruitFishLenAndFishWtStats}
\title{Calculates expected lengths and associated statistics from length-based per recruit analysis}
\usage{
CalcPerRecruitFishLenAndFishWtStats(
  lbnd,
  midpt,
  ubnd,
  FemCatchNumLen,
  MalCatchNumLen,
  CombSexCatchNumLen,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtLen
)
}
\arguments{
\item{midpt}{length class mid-points}

\item{FemCatchNumLen}{female per recruit catch numbers at length}

\item{MalCatchNumLen}{male per recruit catch numbers at length}

\item{CombSexCatchNumLen}{}
}
\value{
PerRecRandFishLen

PerRecMeanCatchLen

PerRecLenFreq

MeanCatchLen.95ci

CatchLen.60qntl

CatchLen.95qntl

PerRecRandFishWt

PerRecMeanCatchWt

MeanCatchWt.95ci

CatchWt.60qntl

CatchWt.95qntl
}
\description{
This function outputs random fish lengths, calculated from length-based per recruit analysis
with specified fishing mortality (PerRecFishLen), expected mean length (PerRecLenFreq),
95 percent confidence limits for mean length (MeanCatchLen.95ci), and 60 and 95 percent
prediction intervals for mean length (CatchLen.60qntl, CatchLen.95qntl). Values are calculated
for females, males and combined sexes.
}
\keyword{internal}
