% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPopnSexRatioAndFertRate}
\alias{CalcPopnSexRatioAndFertRate}
\title{Calculate population sex ratio and fertilisation rate}
\usage{
CalcPopnSexRatioAndFertRate(
  UnfishMalSurvAtAge,
  FishMalSurvAtAge,
  UnfishFemSurvAtAge,
  FishFemSurvAtAge,
  MalPropMatAtAge,
  FemPropMatAtAge,
  EggFertParam
)
}
\arguments{
\item{UnfishMalSurvAtAge}{unfished male survival at age (in numbers)}

\item{FishMalSurvAtAge}{fished male survival at age (in numbers)}

\item{UnfishFemSurvAtAge}{unfished female survival at age (in numbers)}

\item{FishFemSurvAtAge}{fished female survival at age (in numbers)}

\item{MalPropMatAtAge}{proportion of males mature at age}

\item{FemPropMatAtAge}{proportion of females mature at age}

\item{EggFertParam}{egg fertilisation parameters (NA, or ~0.2-1)}
}
\value{
UnfishMalToFemRatio, FishMalToFemRatio, Eq_FertRate
}
\description{
This function calculates the population sex ratio, by number, based on mature fish,
and expected egg fertilisation rate, based on value (NA or between 0.2-1) for the
parameter, EggFertParam. This function is of the same form as a Beverton-Holt stock
recruitment relationship, with a value of 0.2 for EggFertParam resulting in a direct
effect of current sex ratio, relative to the unfished level, on egg fertilisation rate,
vs no effect if EggFertParam = 1. The value of EggFertParam is set based on understanding
of fish spawning behaviour, using a higher value of EggFertParam for group spawning vs lower
value for pair spawning.
}
\keyword{internal}
