% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPopnSexRatioAndFertRate_LB}
\alias{CalcPopnSexRatioAndFertRate_LB}
\title{Calculate population sex ratio and fertilisation rate (length-based per recruit analysis)}
\usage{
CalcPopnSexRatioAndFertRate_LB(
  Unfish_MalNPerRecLen,
  Fish_MalNPerRecLen,
  Unfish_FemNPerRecLen,
  Fish_FemNPerRecLen,
  MalPropMatAtLen,
  FemPropMatAtLen,
  EggFertParam
)
}
\arguments{
\item{Unfish_MalNPerRecLen}{unfished male survival at length (in numbers)}

\item{Fish_MalNPerRecLen}{fished male survival at length (in numbers)}

\item{Unfish_FemNPerRecLen}{unfished female survival at length (in numbers)}

\item{Fish_FemNPerRecLen}{fished female survival at length (in numbers)}

\item{MalPropMatAtLen}{proportion of males mature at length}

\item{FemPropMatAtLen}{proportion of females mature at length}

\item{EggFertParam}{egg fertilisation parameters (NA, or ~0.2-1)}
}
\value{
UnfishMalToFemRatio, FishMalToFemRatio, MalDeplRatio, Eq_FertRate
}
\description{
This function calculates the population sex ratio, by number, based on mature fish,
and expected egg fertilisation rate, based on value (NA or between 0.2-1) for the
parameter, EggFertParam. This function is of the same form as a Beverton-Holt stock
recruitment relationship, with a value of 0.2 for EggFertParam resulting in a direct
effect of current sex ratio, relative to the unfished level, on egg fertilisation rate,
vs no effect if EggFertParam = 1. The value of EggFertParam is set based on understanding
of fish spawning behaviour, using a higher value of EggFertParam for group spawning vs lower
value for pair spawning.
}
\keyword{internal}
