% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPropFemAtAge}
\alias{CalcPropFemAtAge}
\title{Calculate proportion of fish at age that are females}
\usage{
CalcPropFemAtAge(
  FinalSex_A50,
  FinalSex_A95,
  EstSexRatioAtAge,
  ReprodPattern,
  Ages
)
}
\arguments{
\item{FinalSex_A50}{logistic sex change parameter}

\item{FinalSex_A95}{logistic sex change parameter}

\item{ReprodPattern}{1=gonochoristic, 2=protogynous, 3=protandrous}

\item{Ages}{specified ages}

\item{FinalSex_Pmax}{logistic sex change parameter}
}
\value{
PropFemAtAge
}
\description{
This function calculates proportion of fish at age that are females,
based on reproductive pattern (gonochoristic or hermaphroditic), and
sex change parameters
}
\keyword{internal}
