% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPropFemAtLength}
\alias{CalcPropFemAtLength}
\title{Calculate proportion of fish at length that are females}
\usage{
CalcPropFemAtLength(
  FinalSex_L50,
  FinalSex_L95,
  FinalSex_Pmax,
  ReprodPattern,
  midpt
)
}
\arguments{
\item{FinalSex_L50}{logistic sex change parameter}

\item{FinalSex_L95}{logistic sex change parameter}

\item{FinalSex_Pmax}{logistic sex change parameter}

\item{ReprodPattern}{1=gonochoristic, 2=protogynous, 3=protandrous}

\item{midpt}{length class mid-points}
}
\value{
PropFemAtLen
}
\description{
This function calculates proportion of fish at length that are females,
based on reproductive pattern (gonochoristic or hermaphroditic), and
sex change parameters
}
\keyword{internal}
