% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPropMatureAtAge}
\alias{CalcPropMatureAtAge}
\title{Calculate proportion of fish at age that are mature}
\usage{
CalcPropMatureAtAge(mat_A50, mat_A95, EstMatAtAge, Ages)
}
\arguments{
\item{mat_A50}{logistic maturity parameter}

\item{mat_A95}{logistic maturity parameter}

\item{EstMatAtAge}{input maturity schedules (could be set to NA)}

\item{Ages}{ages}
}
\value{
FemPropMatAtAge; MalPropMatAtAge
}
\description{
This function calculates proportion of fish at age that are mature,
based on input maturity parameters, or from set input maturity schedules
}
\keyword{internal}
