% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPropMatureAtLength}
\alias{CalcPropMatureAtLength}
\title{Calculate proportion of fish at length that are mature}
\usage{
CalcPropMatureAtLength(mat_L50, mat_L95, EstMatAtLen, midpt)
}
\arguments{
\item{mat_L50}{logistic maturity parameter}

\item{mat_L95}{logistic maturity parameter}

\item{EstMatAtLen}{input maturity schedules (could be set to NA)}

\item{midpt}{length class mid-points}
}
\value{
FemPropMatAtLen

MalPropMatAtLen
}
\description{
This function calculates proportion of fish at length that are mature,
based on input maturity parameters, or from set input maturity schedules
}
\keyword{internal}
