% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcRecruitmentAge}
\alias{CalcRecruitmentAge}
\title{Determine age at full recruitment into the fishery (assuming knife-edge selection)}
\usage{
CalcRecruitmentAge(RecAssump, Ages, ObsAgeFreq)
}
\arguments{
\item{RecAssump}{0=age at peak frequency, 1=age at peak frequency + 1}

\item{Ages}{ages for analysis}

\item{ObsAgeFreq}{observed frequency at age}
}
\value{
age at full recruitment into the fishery (RecAge)
}
\description{
This function determines the age at full recruitment, based on the age at
peak frequency of fish. If RecAssump=0, RecAge is set to the age at
peak frequency, but if RecAssump=1, RecAge is set to one year above the age at
peak frequency
}
\examples{
# simulate data
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = Res$CatchSample
CalcRecruitmentAge(RecAssump=0, Ages, ObsAgeFreq)
}
\keyword{internal}
