% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Calcs_LenCovertCatchCurve_Schnute}
\alias{Calcs_LenCovertCatchCurve_Schnute}
\title{Calculate values for length-converted catch curve with Schnute growth}
\usage{
Calcs_LenCovertCatchCurve_Schnute(
  GrowthParams,
  RefnceAges,
  ObsRetCatchFreqAtLen,
  MinFreq,
  lbnd,
  midpt,
  ubnd,
  MaxAge
)
}
\arguments{
\item{GrowthParams}{c(y1,y2,a,b) Schnute growth curve parameters}

\item{RefnceAges}{c(t1,t2) Schnute growth curve reference ages}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{MinFreq}{minimum frequency to include}

\item{lbnd}{lower bounds of length classes}

\item{midpt}{mid points of length classes}

\item{ubnd}{upper bounds of length classes}

\item{MaxAge}{maximum age of species to be considered by the model}
}
\value{
length class with peak frequency (PeakLencl), last length class included in analysis (LastLenCl),
observed data to which catch curve is fitted (ObsCatchFreqAtLen2), relatives age at midpoints of length classes
(Age_midptlencl), natural logarithms of changes in numbers with respect to time taken to grow through
each length class (ln_n_dt)
}
\description{
Return outputs associated with length-converted catch curve (see Pauly, 1990), with growth described using the
Schnute (1981) growth model, required to fit the model
}
\keyword{internal}
