% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Calcs_PaulyLenConvertCatchCurve}
\alias{Calcs_PaulyLenConvertCatchCurve}
\title{Calculate values for length-converted catch curve with von Bertalanffy growth}
\usage{
Calcs_PaulyLenConvertCatchCurve(
  GrowthParams,
  ObsRetCatchFreqAtLen,
  MinFreq,
  lbnd,
  midpt,
  ubnd
)
}
\arguments{
\item{GrowthParams}{c(Linf,vbK) von Bertalanffy growth parameters}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{MinFreq}{minimum frequency to include}

\item{lbnd}{lower bounds of length classes}

\item{midpt}{mid points of length classes}

\item{ubnd}{upper bounds of length classes}
}
\value{
length class with peak frequency (PeakLencl), last length class included in analysis (LastLenCl),
observed data to which catch curve is fitted (ObsCatchFreqAtLen2), relatives age at midpoints of length classes
(Age_midptlencl), natural logarithms of changes in numbers with respect to time taken to grow through
each length class (ln_n_dt)
}
\description{
Return outputs associated with length-converted catch curve (Pauly, 1990), required to fit the model
}
\keyword{internal}
