% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcSelectivityAndRetentionAtAge}
\alias{CalcSelectivityAndRetentionAtAge}
\title{Calculate gear selectivity and retention at age}
\usage{
CalcSelectivityAndRetentionAtAge(
  EstSelAtAge,
  EstRetenAtAge,
  Ages,
  sel_A50,
  sel_A95,
  ret_A50,
  ret_A95,
  ret_Pmax
)
}
\arguments{
\item{EstSelAtAge}{input vector for gear selectivity at age (may be set to NA)}

\item{EstRetenAtAge}{input vector for fish retention at age (may be set to NA)}

\item{Ages}{ages for analysis calculated from max age and model timestep}

\item{sel_A50}{logistic gear selectivity curve parameter}

\item{sel_A95}{logistic gear selectivity curve parameter}

\item{ret_A50}{logistic gear retention curve parameter}

\item{ret_A95}{logistic gear retention curve parameter}

\item{ret_Pmax}{logistic gear retention curve parameter}
}
\value{
FemGearSelAtAge, MalGearSelAtAge, FemRetProbAtAge, MalRetProbAtAge
}
\description{
Calculate gear selectivity and retention at age, given input parameters or vectors
}
\keyword{internal}
