% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcSelectivityPenalties}
\alias{CalcSelectivityPenalties}
\title{Calculate penalties for selectivity and reset parameters values when necessary}
\usage{
CalcSelectivityPenalties(
  Res,
  GrowthCurveType,
  GrowthModelType,
  GrowthParams,
  RefnceAges,
  SelectivityType,
  SelParams
)
}
\arguments{
\item{Res}{list of outputs from GetGrowthInputsForLengthTransitionMatrices function}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthModelType}{# 1=1 sex, von Bertalanffy, 2=2 sexes, von Bertalanffy, 3=1 sex, Schnute, 4=2 sexes, Schnute}

\item{GrowthParams}{input growth parameters}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{SelParams}{selectivity parameters L50, L95-L50}

\item{Linf}{input growth reference ages}
}
\value{
Linf asymptotic length parameter
}
\description{
This function ensures that the length at 95% selectivity does not exceed Linf, and
that this length is greater than the length at 50% selectivity by at least 2 mm
}
\keyword{internal}
