% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcSizeDistOfRecruits}
\alias{CalcSizeDistOfRecruits}
\title{Calculate size distribution of 1+ year old recruits}
\usage{
CalcSizeDistOfRecruits(MeanSizeAtAge, CVSizeAtAge, lbnd, ubnd, midpt, nLenCl)
}
\arguments{
\item{MeanSizeAtAge}{mean length at age from growth curve}

\item{CVSizeAtAge}{coefficient of variation (CV), common across all mean lengths at age}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}
}
\value{
Selectivity at length (SelAtLength)
}
\description{
Calculate size distribution of 1+ year old recruits, given mean length
at age 1 and specified CV for lengths at age, assuming a normal distribution
}
\keyword{internal}
