% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcSurvivalAtAge}
\alias{CalcSurvivalAtAge}
\title{Calculate unfished and fished survival of fish at age}
\usage{
CalcSurvivalAtAge(
  nTimeSteps,
  InitRatioFem,
  NatMort,
  FMort,
  FemZAtAge,
  MalZAtAge,
  ReprodPattern,
  PropFemAtAge
)
}
\arguments{
\item{nTimeSteps}{number of model timesteps}

\item{InitRatioFem}{sex ratio at birth}

\item{NatMort}{natural mortality}

\item{FemZAtAge}{female total mortality at age}

\item{MalZAtAge}{male total mortality at age}

\item{ReprodPattern}{1=gonochoristic, 2=protogynous, 3=protandrous}

\item{Ages}{PropFemAtAge proportion of fish at age that are female}
}
\value{
UnfishFemSurvAtAge, UnfishMalSurvAtAge, FishFemSurvAtAge, FishMalSurvAtAge
}
\description{
Calculate unfished and fished survival of fish at age. Rountine also allows for sex change
for  hermaphroditic species.
}
\keyword{internal}
