% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcWeightAtAge}
\alias{CalcWeightAtAge}
\title{Calculates weight at age for each sex for age-based per recruit analysis}
\usage{
CalcWeightAtAge(
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtAge,
  FemLenAtAge,
  MalLenAtAge
)
}
\arguments{
\item{lenwt_a}{weight-length parameter}

\item{ln_lenwt_a}{weight-length parameter}

\item{lenwt_b}{weight-length parameter}

\item{WLrel_Type}{1=power, 2=log-log}

\item{EstWtAtAge}{user-specified weights at ages}

\item{FemLenAtAge}{female lengths at each age (from growth curve)}

\item{MalWtAtAge}{male lengths at each age (from growth curve)}
}
\value{
FemWtAtAge

MalWtAtAge
}
\description{
This function calculates mean weight at age for each sex, based on power curve,
log-log relationship, or allows for these to be used as vectors inputted by user,
for age-based per recruit analysis
}
\keyword{internal}
