% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcWeightAtLengthForPerRecruitAnalysis}
\alias{CalcWeightAtLengthForPerRecruitAnalysis}
\title{Calculates weight at length for each sex for length-based per recruit analysis}
\usage{
CalcWeightAtLengthForPerRecruitAnalysis(
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtLen,
  midpt
)
}
\arguments{
\item{lenwt_a}{weight-length parameter}

\item{ln_lenwt_a}{weight-length parameter}

\item{lenwt_b}{weight-length parameter}

\item{WLrel_Type}{1=power, 2=log-log}

\item{EstWtAtLen}{user-specified weights at lengths}

\item{midpt}{length class mid-points}
}
\value{
FemWtAtLen

MalWtAtLen
}
\description{
This function calculates mean size at age for each sex, based on power curve,
log-log relationship, or allows for these to be used as vectors inputted by user,
for length-based per recruit analysis
}
\keyword{internal}
