% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{ci_exp}
\alias{ci_exp}
\title{Get estimate and 95 percent confidence limits after exponentiating log of parameter estimate and standard error}
\usage{
ci_exp(par, se)
}
\arguments{
\item{par}{parameter estimate in log space}

\item{se}{standard error estimate in log space}
}
\value{
point estimate and 95 percent confidence limits in normal space
}
\description{
Use for outputting parameter estimates from fitted models
}
\keyword{internal}
