% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{ConvertObsDataFromDecAgesToIntegerAges}
\alias{ConvertObsDataFromDecAgesToIntegerAges}
\title{Convert observed length and age data from decimal ages to integer age classes}
\usage{
ConvertObsDataFromDecAgesToIntegerAges(
  TimeStep,
  MaxAge,
  nLenCl,
  ObsRetCatchFreqAtLengthAndAge
)
}
\arguments{
\item{TimeStep}{model timestep}

\item{MaxAge}{maximum age considered by model}

\item{nLenCl}{number of length classes}

\item{ObsRetCatchFreqAtLengthAndAge}{observed decimal age and length frequency data}
}
\value{
ObsCatchFreqAtLengthAndIntAge
}
\description{
Convert observed length and age data from decimal ages to integer age classes, used in age and
length based catch curve analyses
}
\keyword{internal}
