% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GenerateCatchFreqData}
\alias{GenerateCatchFreqData}
\title{Generate random observed catch frequency data}
\usage{
GenerateCatchFreqData(
  TimeStep,
  nTimeSteps,
  midpt,
  nLenCl,
  MLL,
  RetenParams,
  SampleSize_Fem,
  SampleSize_Mal,
  SampleSize,
  DiscSampleSize_Fem,
  DiscSampleSize_Mal,
  DiscSampleSize,
  ObsRetCatchFreqAtLen_Fem,
  ExpRetCatchPropDecAgeGivenLength_Fem,
  ObsRetCatchFreqAtLen_Mal,
  ExpRetCatchPropDecAgeGivenLength_Mal,
  ObsDiscCatchFreqAtLen_Fem,
  ExpDiscCatchPropDecAgeGivenLength_Fem,
  ObsDiscCatchFreqAtLen_Mal,
  ExpDiscCatchPropDecAgeGivenLength_Mal
)
}
\arguments{
\item{TimeStep}{model timestep}

\item{nTimeSteps}{number of model timesteps}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}

\item{MLL}{minimum legal length}

\item{RetenParams}{fish retention parameters}

\item{SampleSize_Fem}{specified sample size for female retained fish}

\item{SampleSize_Mal}{specified sample size for male retained fish}

\item{SampleSize}{specified sample sizes for retained and discarded fish}

\item{DiscSampleSize_Fem}{specified sample size for female discarded fish}

\item{DiscSampleSize_Mal}{specified sample size for male  discarded fish}

\item{DiscSampleSize}{specified sample size for discarded fish}

\item{ObsRetCatchFreqAtLen_Fem}{randomly generated female retained catch frequency at length}

\item{ExpRetCatchPropDecAgeGivenLength_Fem}{expected female retained catch prop at decimal age given length}

\item{ObsRetCatchFreqAtLen_Mal}{randomly generated male retained catch frequency at length}

\item{ExpRetCatchPropDecAgeGivenLength_Mal}{expected male retained catch prop at decimal age given length}

\item{ObsDiscCatchFreqAtLen_Fem}{randomly generated female discarded catch frequency at length}

\item{ExpDiscCatchPropDecAgeGivenLength_Fem}{expected female discarded catch prop at decimal age given length}

\item{ObsDiscCatchFreqAtLen_Mal}{randomly generated female discarded catch frequency at length}

\item{ExpDiscCatchPropDecAgeGivenLength_Mal}{expected male discarded catch prop at decimal age given length}
}
\value{
ObsRetCatchFreqAtLengthAndDecAge_Fem, ObsRetCatchFreqAtLengthAndDecAge_Mal, ObsRetCatchFreqAtLengthAndDecAge,
ObsDiscCatchFreqAtLengthAndDecAge_Fem, ObsDiscCatchFreqAtLengthAndDecAge_Mal, ObsDiscCatchFreqAtLengthAndDecAge,
ObsDecAgeRetCatch_Fem, ObsDecAgeRetCatch_Mal, ObsDecAgeRetCatch, ObsDecAgeDiscCatch_Fem, ObsDecAgeDiscCatch_Mal,
ObsDecAgeDiscCatch, ObsAgeClRetCatch_Fem, ObsAgeClRetCatch_Mal, ObsAgeClRetCatch, ObsAgeClDiscCatch_Fem,
ObsAgeClDiscCatch_Mal, ObsAgeClDiscCatch, ObsLenClRetCatchMidPt_Fem, ObsLenClRetCatchMidPt_Mal, ObsLenClRetCatchMidPt,
ObsLenClDiscCatchMidPt_Fem, ObsLenClDiscCatchMidPt_Mal, ObsLenClDiscCatchMidPt
}
\description{
Generate random observed catch frequencies for each length class in each age class, and marginal length and age frequencies
}
\keyword{internal}
