% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GenerateRandomAgeFreqData_DynSimMod}
\alias{GenerateRandomAgeFreqData_DynSimMod}
\title{Generate random age data from age-based dynamic simulation model}
\usage{
GenerateRandomAgeFreqData_DynSimMod(
  nYears,
  SimAnnSampSize,
  MaxModelAge,
  Catch_Fem,
  Catch_Mal,
  Catch_CombSex
)
}
\arguments{
\item{nYears}{number of years of age data}

\item{SimAnnSampSize}{specified annual sample size (both sexes)}

\item{MaxModelAge}{maximum age considered by model}

\item{Catch_Fem}{number of females in annual retained catches in each length class}

\item{Catch_Mal}{number of females in annual retained catches in each length class}

\item{Catch_CombSex}{number of females and males in annual retained catches in each length class}
}
\value{
MeanLengthStats
}
\description{
Generates random length data, and associated age frequencies,
for each year, from dynamic age-based simulation model
}
\keyword{internal}
