% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GenerateRandomLengthFreqData_DynSimMod}
\alias{GenerateRandomLengthFreqData_DynSimMod}
\title{Generate random length data from dynamic simulation model}
\usage{
GenerateRandomLengthFreqData_DynSimMod(
  nYears,
  SimAnnSampSize,
  lbnd,
  midpt,
  ubnd,
  Catch_Fem,
  Catch_Mal,
  Catch_CombSex
)
}
\arguments{
\item{nYears}{number of years of length data}

\item{SimAnnSampSize}{specified annual sample size (both sexes)}

\item{lbnd}{lower bounds of length classes}

\item{midpt}{mid points of length classes}

\item{ubnd}{upper bounds of length classes}

\item{Catch_Fem}{number of females in annual retained catches in each length class}

\item{Catch_Mal}{number of females and males in annual retained catches in each length class}
}
\value{
MeanLengthStats
}
\description{
Generates random length data, and associated length frequencies,
for each year, from dynamics simulation model
}
\keyword{internal}
