% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Get_Relative_Value_Per_Recruit_LB}
\alias{Get_Relative_Value_Per_Recruit_LB}
\title{Get outputs from length-based value per recruit analysis}
\usage{
Get_Relative_Value_Per_Recruit_LB(
  MaxModelAge,
  TimeStep,
  lbnd,
  ubnd,
  midpt,
  nLenCl,
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  CVSizeAtAge,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtLen,
  ReprodScale,
  ReprodPattern,
  InitRatioFem,
  FinalSex_L50,
  FinalSex_L95,
  EstSexRatioAtLen,
  EggFertParam,
  mat_L50,
  mat_L95,
  EstMatAtLen,
  sel_L50,
  sel_L95,
  EstGearSelAtLen,
  ret_Pmax,
  ret_L50,
  ret_L95,
  EstRetenAtLen,
  DiscMort,
  Steepness,
  SRrel_Type,
  NatMort,
  Current_F,
  MinFishWtOfVal,
  ValScaleFact
)
}
\arguments{
\item{MaxModelAge}{maximum age considered by model}

\item{TimeStep}{model time step (in y)}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points bounds of length classes}

\item{nLenCl}{number of length classes}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{growth parameters of either von Bertalanffy or Schnute model}

\item{RefnceAges}{reference ages for Schnute model, set to NA if using von Bertalanffy model}

\item{CVSizeAtAge}{coefficient of variation for size at age}

\item{lenwt_a}{weight-length parameter}

\item{ln_lenwt_a}{weight-length parameter}

\item{lenwt_b}{weight-length parameter}

\item{WLrel_Type}{1=power, 2=log-log}

\item{EstWtAtLen}{user-specified weights at lengths}

\item{ReprodScale}{1=default (standard calculations for spawning biomass), use alternative number to specify reproductive scaling with female mass factor (i.e. BOFFF effects)}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{proportion of fish that are females at hatching}

\item{FinalSex_L50}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_L95}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{EstSexRatioAtLen}{NA  # sex ratio at length, inputted as vector}

\item{EggFertParam}{<- NA # (NA or from ~0.2-1) NA = no effect, 0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects}

\item{mat_L50}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{mat_L95}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{EstMatAtLen}{vector of proportion mature at length (set to NA if using age at maturity parameters)}

\item{sel_L50}{logistic parameter for gear selectivity curve}

\item{sel_L95}{logistic parameter for gear selectivity curve}

\item{EstGearSelAtLen}{gear selectivity curve inputted as vector}

\item{ret_Pmax}{logistic parameter for fish retention curve}

\item{ret_L50}{logistic parameter for fish retention curve}

\item{ret_L95}{logistic parameter for fish retention curve}

\item{EstRetenAtLen}{retention curve inputted as vector}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{SRrel_Type}{1 = Beverton-Holt, 2=Ricker stock-recruitment relationship}

\item{NatMort}{natural mortality}

\item{FMort}{fishing mortality}
}
\value{
relative value of fish vs length to fishers (RelValAtLen), relative value per recruit to fishers
of catch for specified fishing mortality (RelValPerRecAtFMort), maximum value per recruit to fishers
across range of fishing mortalities (MaxValPerRec), level of fishing mortality required to achieve MaxValPerRec.
}
\description{
This function provides outputs associated with a 'value per recruit' analysis, calculated based on
length-based per recruit analysis (with a stock recruitment relationship to account
for potential impacts of fishing on recruitment). The relative value of fish at different lengths
is specified according to a logistic curve, with specified input parameter values.
}
\examples{
MaxModelAge <- 20 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be appropriate to use a smaller time step)
MaxLen = 800
LenInc = 20
lbnd = seq(0,MaxLen - LenInc, LenInc)
ubnd = lbnd + LenInc
midpt = lbnd + (LenInc/2)
nLenCl = length(midpt)
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf <- c(550, 500) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.2, 0.2) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
GrowthParams = data.frame(Linf=Linf, vbK=vbK, tzero=tzero)
RefnceAges = NA
# GrowthParams c(Linf, vbK, tzero) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK, tzero=tzero),
#' both sexes von Bertalanffy, or c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b), both sexes Schnute
CVSizeAtAge = c(0.05,0.05)
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtLen <- data.frame(EstFemWtAtLen=NA,
                         EstMalWtAtLen=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), use alternative number to specify reproductive scaling with female mass factor (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at age zero
FinalSex_L50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_L95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
EstSexRatioAtLen <- NA  # sex ratio at length inputted as vector
EggFertParam <- NA # (NA or from ~0.2-1) NA = no effect, ~0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects
mat_L50 <- c(250, 250) # females, males - Logistic length (mm) at maturity relationship parameters
mat_L95 <- c(300, 300) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtLen <- data.frame(EstFemMatAtLen=NA,
                          EstMalMatAtLen=NA) # maturity at age (from age 0), inputted as values in data frame
sel_L50 <- c(250, 250) # females, males - Logistic age selectivity relationship parameters
sel_L95 <- c(300, 300) # females, males - Logistic age selectivity relationship parameters
EstGearSelAtLen <- data.frame(EstFemGearSelAtLen=NA,
                              EstMalGearSelAtLen=NA)
ret_Pmax <- c(1.0, 1.0) # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_L50 <- c(250, 250) # females, males - Logistic age fish retention at age parameters
ret_L95 <- c(300, 300) # females, males - Logistic age fish retention at age parameters
EstRetenAtLen <- data.frame(EstFemRetenAtLen=NA,
                            EstMalRetenAtLen=NA)
DiscMort <- 0.25 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort = 4.22 / MaxModelAge # natural mortality  (year-1)
Current_F <- 0.2
MinFishWtOfVal = 100 # size of fish below which fish has no value
ValScaleFact = 1.2 # amount by which value scales according to size, above mininum size of value
FittedRes = Get_Relative_Value_Per_Recruit_LB(MaxModelAge, TimeStep, lbnd, ubnd, midpt, nLenCl, GrowthCurveType, GrowthParams, RefnceAges,
                                              CVSizeAtAge, lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtLen, ReprodScale, ReprodPattern,
                                              InitRatioFem, FinalSex_L50, FinalSex_L95, EstSexRatioAtLen, EggFertParam, mat_L50, mat_L95,
                                              EstMatAtLen, sel_L50, sel_L95, EstGearSelAtLen, ret_Pmax, ret_L50, ret_L95, EstRetenAtLen,
                                              DiscMort, Steepness, SRrel_Type, NatMort, Current_F, MinFishWtOfVal, ValScaleFact)
}
