% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Get_xaxis_scale}
\alias{Get_xaxis_scale}
\title{Get maximum and minimum x values and x interval}
\usage{
Get_xaxis_scale(x_data)
}
\arguments{
\item{x_data}{x axis data for plot}
}
\value{
list object with minimum and maximum x axis values and x axis interval
}
\description{
Used for setting plot defaults
}
\keyword{internal}
