% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Get_yaxis_scale}
\alias{Get_yaxis_scale}
\title{Get maximum and minimum y values and y interval}
\usage{
Get_yaxis_scale(y_data)
}
\arguments{
\item{y_data}{y axis data for plot}
}
\value{
list object with minimum and maximum y axis values and y axis interval
}
\description{
Used for setting plot defaults
}
\keyword{internal}
