% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetAgeAndLengthBasedCatchCurveResults}
\alias{GetAgeAndLengthBasedCatchCurveResults}
\title{Get results for a fitted age and length-based catch curve}
\usage{
GetAgeAndLengthBasedCatchCurveResults(
  params,
  RefnceAges,
  MLL,
  DistnType,
  nSexes,
  GrowthCurveType,
  SelectivityType,
  ObsRetCatchFreqAtLen,
  ObsRetCatchFreqAtLengthAndAge,
  lbnd,
  ubnd,
  midpt,
  SelectivityAtLen,
  DiscMort,
  MaxAge,
  NatMortType,
  NatMort,
  NatMort_sd,
  TimeStep
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated}

\item{RefnceAges}{Reference ages for Schnute growth curve (set to NA for von Bertalanffy growth curve)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{DistnType}{1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)}

\item{nSexes}{1 or 2 (for estimating growth parameters)}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{ObsRetCatchFreqAtLengthAndAge}{observed frequencies in length and age classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityAtLen}{selectivity at length}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{MaxAge}{maximum age considered in model}

\item{NatMortType}{1 = fixed, 2 = estimated with prior to allow for uncertainty}

\item{NatMort}{natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space}

\item{NatMort_sd}{standard deviation for natural mortality in log space. Set to NA if NatMortType = 1}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence)
sample size (SampleSize), growth parameter estimates with lower and upper 95 percent
confidence limits (ParamEst), point estimates for estimated parameters (params)
and variance-covariance matrix (vcov.Params), gear selectivity at length (SelAtLength), retention
at length (RetAtLength), selectivity of landings at length (SelLandAtLength),
growth curve (MeanSizeAtAge), midpoint of each length class (midpt), mean length after 1 year from
growth curve, given initial length (MeanEndingLength), mean change in length after 1 year,
from initial length - note, assuming normal a distribution allows for possibility of negative growth
if above asyptotic length (TimestepGrowthSizeInc), length distribution of 1+ year old recruits (RecLenDist),
expected retained catches at length and integer age class (RetCatchAtIntAgeLen), expected catches at length (ExpCatchAtLen),
catch proportions at length (ExpCatchPropInLenClass), expected retained catches at integer age classes (RetCatchAtIntAge),
catch proportions at integer age class of females, males and sexes combined (ExpRetCatchPropAtIntAge, ExpRetCatchPropAtIntAge_Fem,
ExpRetCatchPropAtIntAge_Mal), catch proportions at length and integer age class for females, males and sexes combined
(ExpRetCatchPropLengthGivenIntAge, ExpRetCatchPropLengthGivenIntAge_Fem, ExpRetCatchPropLengthGivenIntAge_Mal),
observed catch frequency at length data (ObsRetCatchFreqAtLen), observed catch frequency at length and age
(ObsRetCatchFreqAtLengthAndAge), approximate CV at length corresponding to maximum age (CV_LenAtMaxAge, FemCV_LenAtMaxAge,
MalCV_LenAtMaxAge)
}
\description{
This function fits a length-based catch curve with length-based selectivity, based on
either 1) specified length-based selectivity inputted as a vector or 2) estimated asymptotic logistic
selectivity curve, and estimated von Bertalalnffy growth curve. The model is fitted to a
sample of fish length and age data, by minimising the overall negative log-likelihood,
including the NLL associated with the marginal length composition and a conditional age at length NLL,
given the parameters (selectivity, growth and mortality) and data, using nlminb.
It provides several statistical outputs including convergence statistics, parameter estimates
and associated 95 percent confidence limits and associated variance-covariance matrix, calculated using
the MASS package. Note, model is not suited to hermaphroditic species
with sexually divergent growth.
}
\examples{
# Simulate data
set.seed(123)
SampleSize=5000
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
RefnceAges = NA
GrowthParams = c(Linf, vbK, CVSizeAtAge)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# 2 sexes, von Bertalanffy
# DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# GrowthCurveType = 1 # 1 = von Bertalanffy
# Linf = c(700,850)
# vbK = c(0.25,0.2)
# CVSizeAtAge = c(0.05,0.05)
# RefnceAges = NA
# GrowthParams = data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge)
# Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
#                          SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# # get data - 2 sexes
# ObsRetCatchFreqAtLen <- data.frame(matrix(nrow = 2, ncol = length(midpt))) # 2 sex
# colnames(ObsRetCatchFreqAtLen) <- midpt
# ObsRetCatchFreqAtLen[1,] = Res$ObsRetCatchFreqAtLen_Fem
# ObsRetCatchFreqAtLen[2,] = Res$ObsRetCatchFreqAtLen_Mal
# ObsRetCatchFreqAtLengthAndAge = array(c(unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem), unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Mal)),
#                                       c(nTimeSteps, length(midpt), 2), dimnames=list(rownames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem),
#                                                                                      colnames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem)))
# get params - 1 sex
nSexes = 1
InitFishMort = 0.3 # specify starting parameters
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
# get params - 2 sexes
# nSexes = 2
# InitFishMort = 0.3 # specify starting parameters
# InitL50 = 320
# InitDelta = 50 # L95-L50
# InitLinf = c(800,800)
# InitvbK = c(0.25,0.25)
# InitCVSizeAtAge = 0.05
# InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
# params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)

# Example with specified selectivity vector
# Simulate data
SelectivityType=1 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = 1 / (1 + exp(-log(19)*(midpt-400)/(500-400)))
SelParams = c(NA, NA) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
# DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
GrowthParams = c(Linf, vbK, CVSizeAtAge)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# get params - 1 sex
nSexes = 1
InitFishMort = 0.3 # specify starting parameters
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)

# Example estimating sex-specific growth parameters for a species with sexually-dimorphic growth,
# fitting the model to combined sex data (as sex not recorded)
# Simulate data
set.seed(123)
SampleSize=1000
MaxAge = 40
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = NatMort
MaxLen = 1000
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
lbnd = seq(0,MaxLen - LenInc, LenInc)
midpt = lbnd + (LenInc/2)
SelectivityAtLen = NA
SelParams = c(350, 50) # L50, L95-L50 for gear selectivity
RetenParams = NA # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = c(600,800) # sex difference in Linfinity
vbK = c(0.2,0.2)
CVSizeAtAge = c(0.05,0.05)
GrowthParams = data.frame(Linf=Linf, vbK=vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex

# fit catch curve model
nSexes = 2
InitFishMort = 0.1 # specify starting parameters
InitL50 = 300
InitDelta = 50 # L95-L50
InitLinf = c(550,750) # make sure starting values reasonable well separated, to aid optimisation
InitvbK = c(0.3,0.3)
InitCVSizeAtAge = 0.04
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)

# Examples for when data are available in 'raw' form, i.e. as lengths and ages for individual fish
# Simulate data
set.seed(123)
SampleSize=5000
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
RefnceAges = NA
GrowthParams = c(Linf, vbK, CVSizeAtAge)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# get individual fish data
FishAges = Res$ObsDecAgeRetCatch # ages for individual fish
FishLengths = Res$ObsRandLenRetCatch # length for individual fish
lbnd = Res$lbnd
midpt = Res$midpt
ubnd = Res$ubnd
LenInc = ubnd[1] - lbnd[1]
# single sex - get length frequency
lbns = trunc(FishLengths/LenInc)*LenInc
ObsRetCatchFreqAtLen  = as.vector(table(factor(lbns, levels=lbnd)))
# single sex - get frequencies in each length and age class
ObsRetCatchFreqAtLengthAndAge <- data.frame(matrix(nrow = MaxAge, ncol = length(lbnd)))
colnames(ObsRetCatchFreqAtLengthAndAge) <- lbnd
ObsRetCatchFreqAtLengthAndAge = as.matrix(table(factor(FishAges, levels=1:MaxAge),
                                                factor(lbns, levels=c(lbnd))))
# fit model
nSexes  = 1
InitFishMort = 0.3 # specify starting parameters
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)
# 2 sexes, using data for individual fish
# Simulate data
set.seed(123)
SampleSize=5000
MaxAge = 30
MaxObsAge = 25
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxObsAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(400, 100) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# 2 sexes, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = c(700,850)
vbK = c(0.25,0.2)
CVSizeAtAge = c(0.05,0.05)
RefnceAges = NA
GrowthParams = data.frame(Linf=Linf, vbK=vbK)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# get individual fish data
FishAges = c(Res$ObsDecAgeRetCatch_Fem, Res$ObsDecAgeRetCatch_Mal) # ages for individual fish
FishLengths = c(Res$ObsRandLenRetCatch_Fem, Res$ObsRandLenRetCatch_Mal) # length for individual fish
lbns = trunc(FishLengths/LenInc)*LenInc
FishSex = c(rep(1,length(Res$ObsDecAgeRetCatch_Fem)), rep(2, length(Res$ObsDecAgeRetCatch_Mal)))
lbnd = Res$lbnd
midpt = Res$midpt
ubnd = Res$ubnd
LenInc = ubnd[1] - lbnd[1]
# 2 sexes - get length frequency
ObsRetCatchFreqAtLen <- data.frame(matrix(nrow = 2, ncol = length(midpt))) # 2 sex
colnames(ObsRetCatchFreqAtLen) <- midpt
ObsRetCatchFreqAtLen = as.matrix(table(factor(FishSex, levels=c(1:2)),
                                       factor(lbns, levels=c(lbnd))))
# 2 sexes - get frequencies in each length and age class
ObsRetCatchFreqAtLengthAndAge = array(c(unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem), unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Mal)),
                                      c(nTimeSteps, length(midpt), 2), dimnames=list(rownames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem),
                                                                                     colnames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem)))
# fit model
# get params - 2 sexes
nSexes = 2
InitFishMort = 0.15 # specify starting parameters
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = c(650,800)
InitvbK = c(0.25,0.25)
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)

# Fit single (or combined) sex model using Schnute growth curve
# Simulate data
set.seed(123)
SampleSize=1000 # sample size for retained catches (and same number for released fish, if an MLL is specified)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# 1 sex, Schnute
GrowthCurveType = 2 # 1 = von Bertalanffy, 2 = Schnute
t1 = 0 # growth - Schnute
t2 = 25 # growth - Schnute
y1 = 0 # growth - Schnute
y2 = 1000 # growth - Schnute
a = 0.1 # growth - Schnute
b = 2 # growth - Schnute
GrowthParams = c(y1, y2, a, b) # ensure t1 and y1 are fixed at zero (not possible to estimate with this model)
RefnceAges = c(t1,t2)
CVSizeAtAge = 0.08
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# fit model
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen
ObsRetCatchFreqAtLengthAndAge = Res$ObsRetCatchFreqAtLengthAndDecAge
lbnd = Res$lbnd
midpt = Res$midpt
ubnd = Res$ubnd
nSexes = 1
GrowthCurveType=2 # Schnute
InitFishMort = 0.2 # specify starting parameters
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
InitL50 = 400
InitDelta = 20 # L95-L50
RefnceAges = c(0,20)
y2=900; a=0.15; b=1.5; InitCVSizeAtAge = 0.1
params = c(InitFishMort_logit, log(InitL50), log(InitDelta), log(y2), a, b, log(InitCVSizeAtAge))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)
# if wanting to obtain estimated lengths-at-ages over larger age range,
# extrapolating beyond MaxAge (using WAFishBiology package)
library(WAFishBiology)
MaxModelAge <- 50
t1=0 # RefnceAges[1]
t2=RefnceAges[2]
y2=FittedRes$ParamEst[4,1]
a=FittedRes$ParamEst[5,1]
b=FittedRes$ParamEst[6,1]
EstLenAtAge <- rep(0,MaxModelAge)
for (Age in 1:MaxModelAge) {
  EstLenAtAge[Age]=SchnuteGrowthfunction(Age, t1=0, t2, y1=0, y2, a, b)
}
plot(1:MaxModelAge, EstLenAtAge, "o")

# Estimate M (with informative prior) as well as other parameters
# Simulate data
set.seed(123)
SampleSize=1000
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 2 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = 0.31 # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
RefnceAges = NA
GrowthParams = c(Linf, vbK, CVSizeAtAge)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# get params - 1 sex
nSexes = 1
InitFishMort = 0.3 # specify starting parameters
InitNatMort = 0.2
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(InitNatMort), log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)
FittedRes$ParamEst
# plot M prior vs posterior
post_median <- FittedRes$ParamEst[2,1]
post_CI_lower <- FittedRes$ParamEst[2,2]
post_CI_upper <- FittedRes$ParamEst[2,3]
post_meanlog <- log(post_median)
post_sdlog <- (log(post_CI_upper) - log(post_CI_lower)) / (2 * qnorm(0.975))
# plot M prior
par(mfrow=c(1,1),mar=c(5,4,2,2))
x <- seq(0, max(NatMort, post_CI_upper)*2, length.out = 1000)
y <- dlnorm(x, meanlog = log(NatMort), sdlog = NatMort_sd)
plot(x, y, ylim=c(0,1.3*max(y)), type = "l", lwd = 2, col = "blue", xlab = "M", ylab = "Probability",
     main = "M prior vs posterior", bty='n', las=1, cex=0.8)
abline(v = NatMort, col = "blue", lty = 2)
# plot M posterior
lines(x, dlnorm(x, meanlog = post_meanlog, sdlog = post_sdlog),
      lwd = 2, col = "red")
abline(v = post_median, col = "red", lty = 2)
abline(v = c(post_CI_lower, post_CI_upper), col = "darkgreen", lty = 3)
legend("topright",
       legend = c("Prior density", "Prior median", "Posterior density", "Posterior median", "Posterior 95\% CI"),
       col = c("blue", "blue", "red", "red", "darkgreen"),
       lty = c(1,2,1,2,3), lwd = c(2,1,2,1,1), bty='n', cex=0.8)

# Fit model assuming marginal length distribution conforms to a Dirichlet multinomial distribution
# Simulate data
set.seed(123)
SampleSize=1000
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 2 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
RefnceAges = NA
GrowthParams = c(Linf, vbK, CVSizeAtAge)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# Now simulate marginal length-frequency data from Dirchlet multinomial distribution.
nSampEvents = 50
nFishPerSampEvent = 20
theta_val = 0.3
midpt = Res$midpt
ExpPropAtLen = Res$ModelDiag$ExpRetCatchPropAtLen_Fem
# plot(midpt, Res$ModelDiag$ExpRetCatchPropAtLen_Fem, "l")
# lines(midpt, Res$ModelDiag$ExpRetCatchPropAtLen_Mal, col="blue") # check that same for both sexes (as single sex model)
res=SimLenFreqDat_DirMultDistn_EqMod(nSampEvents, nFishPerSampEvent, theta_val, midpt, ExpPropAtLen)
plot(res$midpt, Res$ObsRetCatchFreqAtLen, "o")
lines(res$midpt, res$simLenFreq, "o", col="blue")
# Fit length-based catch curve, assuming Dirichlet multinomial distribution
ObsRetCatchFreqAtLen = res$simLenFreq
nSexes = 1
InitFishMort = 0.3 # specify starting parameters
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
InitTheta = 0.3 # specify starting parameters
InitTheta_logit = log(InitTheta/(1-InitTheta)) # logit transform
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)), InitTheta_logit)
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)
}
