% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetDiscardCatchStats}
\alias{GetDiscardCatchStats}
\title{Get discarded catch summary statistics}
\usage{
GetDiscardCatchStats(
  midpt,
  MLL,
  RetenParams,
  DiscCatchAtLen,
  TotCatchAtLen,
  SampleSize
)
}
\arguments{
\item{midpt}{mid points of length classes}

\item{MLL}{minimum legal length}

\item{RetenParams}{parameters for fish retention}

\item{DiscCatchAtLen}{discarded catch at length}

\item{TotCatchAtLen}{total catch at length}

\item{SampleSize}{sample sizes}
}
\value{
PropReleased, PropReleased_sd, MeanLenReleased, MeanLenReleased_sd
}
\description{
Calculate prop discarded and mean length of discarded fish
}
\keyword{internal}
