% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetDiscCatchFreqAtLenAndIntAge}
\alias{GetDiscCatchFreqAtLenAndIntAge}
\title{Get discarded catch frequencies at length in each integer age, and in each age class}
\usage{
GetDiscCatchFreqAtLenAndIntAge(
  TimeStep,
  MaxAge,
  midpt,
  nLenCl,
  ObsAgeClDiscCatch_Fem,
  ObsAgeClDiscCatch_Mal,
  ObsLenClDiscCatchMidPt_Fem,
  ObsLenClDiscCatchMidPt_Mal
)
}
\arguments{
\item{TimeStep}{model time step}

\item{MaxAge}{maximum age}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}

\item{ObsAgeClDiscCatch_Fem}{generated female catch in each integer age class}

\item{ObsAgeClDiscCatch_Mal}{generated male catch in each integer age class}

\item{ObsLenClDiscCatchMidPt_Fem}{generated female catch in each length class}

\item{ObsLenClDiscCatchMidPt_Mal}{generated male catch in each length class}
}
\value{
ObsDiscCatchFreqAtLengthAndIntAge_Fem, ObsDiscCatchFreqAtLengthAndIntAge_Mal, ObsDiscCatchFreqAtLengthAndIntAge
}
\description{
Get discarded catch frequencies in each of the length classes at each integer age, and in each age class
}
\keyword{internal}
