% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetExpCatchResults}
\alias{GetExpCatchResults}
\title{Get expected discarded and retained catches for females, males and combined sexes}
\usage{
GetExpCatchResults(
  MaxAge,
  TimeStep,
  nTimeSteps,
  midpt,
  nLenCl,
  SelectivityType,
  SelParams,
  SelectivityAtLen,
  RetenParams,
  MLL,
  NatMort,
  FishMort,
  DiscMort,
  InitRecNumber,
  RecLenDist,
  LTM_Fem,
  LTM_Mal
)
}
\arguments{
\item{nTimeSteps}{number of model time steps}

\item{midpt}{midpoints of length classes}

\item{nLenCl}{number of length classes}

\item{CatchResults_Fem}{outputs from rcpp function to calculate expected equilibrium catches for females}

\item{CatchResults_Mal}{outputs from rcpp function to calculate expected equilibrium catches for males}
}
\value{
Results
}
\description{
Get expected discarded and retained catches for females, males and combined sexes, from rccp functions
}
\keyword{internal}
