% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetGrowthAndSelectivityParams}
\alias{GetGrowthAndSelectivityParams}
\title{Get growth parameter values for catch curve}
\usage{
GetGrowthAndSelectivityParams(
  params,
  GrowthParams,
  RefnceAges,
  CatchCurveType,
  SelectivityType
)
}
\arguments{
\item{params}{estimated model parameters (varies, depending on growth curve type, selectivity type and catch curve type)}

\item{GrowthParams}{c(Linf, vbK, CVSizeAtAge) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge),
both sexes von Bertalanffy, or c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b), both sexes Schnute}

\item{RefnceAges}{reference ages for Schnute function (set to NA if growth based on another function)}

\item{CatchCurveType}{1=length based, 2=age and length based}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}
}
\value{
values of growth parameters given specified catch curve, growth and selectivity model option
}
\description{
Get growth parameter values for catch curve given specified catch curve, growth and selectivity model options
}
\keyword{internal}
