% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetGrowthModelType}
\alias{GetGrowthModelType}
\title{Allocate number for alternative length based catch curve models}
\usage{
GetGrowthModelType(GrowthCurveType, GrowthParams)
}
\arguments{
\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{(von Bertalanffy or Schnute growth parameters)}
}
\value{
GrowthModelType 1=length-based von Bertalanffy, 2=length-based Schnute,
3=age and length based von Bertalanffy, 4=age and length based Schnute
}
\description{
Allocate number for alternative single and separate sex length based catch curve models,
and growth curve types (1=von Beralanffy, 2=Schnute)
}
\keyword{internal}
