% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetGrowthParams_AgeAndLengthBasedCatchCurvesCalcs}
\alias{GetGrowthParams_AgeAndLengthBasedCatchCurvesCalcs}
\title{Get growth parameter values for age and length based catch curve}
\usage{
GetGrowthParams_AgeAndLengthBasedCatchCurvesCalcs(
  params,
  NatMortType,
  GrowthCurveType,
  SelectivityType
)
}
\arguments{
\item{params}{estimated model parameters (varies, depending on growth curve type, selectivity type and catch curve type)}

\item{NatMortType}{1 = fixed, 2 = estimated with prior to allow for uncertainty}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params,}
}
\value{
list including GrowthParams (von Bertalanffy or Schnute growth parameters), GrowthModelType
1=von Bertalanffy 1 sex, 2= von Bertalanffy 2 sexes, 3=Schnute 1 sex, 4=Schnute 2 sexes, and
CVSizeAtAge for 1 or 2 sex model
}
\description{
Get growth parameter values for age and length based catch curve given specified catch curve, growth and selectivity model options
}
\keyword{internal}
