% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetInputsForLengthTransitionMatrices}
\alias{GetInputsForLengthTransitionMatrices}
\title{Get inputs for length transition matrices}
\usage{
GetInputsForLengthTransitionMatrices(
  MaxAge,
  nLenCl,
  midpt,
  GrowthModelType,
  Linf,
  vbK,
  t1,
  t2,
  y1,
  y2,
  a,
  b,
  GrowthParams,
  RefnceAges
)
}
\arguments{
\item{MaxAge}{maximum age considered in model}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{GrowthModelType}{1=von Bertalanffy single sex, 2=von Bertalanffy both sexes, 3=Schnute single sex,
4=Schnute both sexes}

\item{Linf}{asymptotic length (von Bertalanffy growth curve)}

\item{vbK}{growth coefficient (von Bertalanffy growth curve)}

\item{t1}{first reference age (Schnute growth curve)}

\item{t2}{second reference age (Schnute growth curve)}

\item{y1}{length at first reference age (Schnute growth curve)}

\item{y2}{length at second reference age (Schnute growth curve)}

\item{a}{Schnute growth curve parameter}

\item{b}{Schnute growth curve parameter}

\item{GrowthParams}{growth parameters (used for GrowthModelType 2 or 4)}

\item{RefnceAges}{Schnute reference ages (used for GrowthModelType 4)}
}
\value{
mean size at age from growth curve (MeanSizeAtAge), estimated length after one year for each length class
mid point, given growth curve (MeanEndingLength) and associated change in length after one year of growth (TimestepGrowthSizeInc)
}
\description{
Get key inputs required to calculate length transition matrices, for models with either one or two sexes,
and using von Bertalanffy or Schnute growth model
}
\keyword{internal}
