% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetLengthTransitionMatrices}
\alias{GetLengthTransitionMatrices}
\title{Get length transition matrices}
\usage{
GetLengthTransitionMatrices(
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  MaxAge,
  TimeStep,
  lbnd,
  midpt,
  ubnd,
  nLenCl,
  CVSizeAtAge
)
}
\arguments{
\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{growth parameters}

\item{RefnceAges}{reference ages for growth curves (Schnute)}

\item{MaxAge}{maximum age considered by model}

\item{TimeStep}{model timestep}

\item{lbnd}{lower bounds of length classes}

\item{midpt}{mid points of length classes}

\item{ubnd}{upper bounds of length classes}

\item{nLenCl}{number of length classes}

\item{CVSizeAtAge}{growth variation}
}
\value{
LTM_Fem, LTM_Mal
}
\description{
Get sex-specific length transition matrices
}
\keyword{internal}
