% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetLinearCatchCurveResults}
\alias{GetLinearCatchCurveResults}
\title{Fit linear catch curve and get results}
\usage{
GetLinearCatchCurveResults(RecAssump, SpecRecAge, MinFreq, Ages, ObsAgeFreq)
}
\arguments{
\item{RecAssump}{0=age at peak frequency, 1=age at peak frequency + 1}

\item{SpecRecAge}{specified age at full recruitment. Set to NA if RecAssump = 0 or 1.}

\item{MinFreq}{minimum frequency of fish for including data for old fish}

\item{Ages}{ages for analysis}

\item{ObsAgeFreq}{observed age frequency}
}
\value{
linear catch curve y intercept and associated standard error (yint, yint_se),
total mortality and associated standard error (ZMort, ZMort_se), age at full recruitment (RecAge),
last age included in analysis (LastAgeForLinearCC), age classes included in analysis (AgeClassesForLinearCC),
natural logarithms of age frequencies in analysis (lnObsAgeFreqForLinearCC), natural logarithms of estimated
frequencies and associated 95 percent confidence limits (EstlnFreq, EstlnFreq_Zlow, EstlnFreq_Zup), ages
for observed frequency data (Ages), estimated frequencies at age, in normal space, with associated 95 percent
confidence limits (EstFreq, EstFreq_Zlow, EstFreq_Zup)
}
\description{
This function fits a linear catch curve to fish age composition data.
The age at full recruitment is determined using on the age at
peak frequency of fish. If RecAssump=0, RecAge is set to the age at
peak frequency, but if RecAssump=1, RecAge is set to one year above the age at
peak frequency. The oldest age that should be considered for analysis with
the linear catch curve analysis, is calculated given a specified minimum frequency
(must be >0). The catch curve is fitted to the natural logarithms of the frequency
data using the lm function
}
\examples{
# simulate data
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = as.vector(unlist(Res$CatchSample))
res=GetLinearCatchCurveResults(RecAssump=0, SpecRecAge=NA, MinFreq=1, Ages, ObsAgeFreq)
}
