% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetLTMInputsForPerRecruitAnalysis}
\alias{GetLTMInputsForPerRecruitAnalysis}
\title{Calculates inputs required to calculate length transition matrices, for length-based per recruit analysis}
\usage{
GetLTMInputsForPerRecruitAnalysis(
  GrowthCurveType,
  TimeStep,
  MaxModelAge,
  GrowthParams,
  RefnceAges,
  midpt
)
}
\arguments{
\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{growth parameters of either von Bertalanffy or Schnute model}

\item{RefnceAges}{reference ages for Schnute model, set to NA if using von Bertalanffy model}

\item{midpt}{length class mid-points}

\item{Ages}{ages considered by model}
}
\value{
MeanEndingLength

TimestepGrowthSizeInc
}
\description{
This function calculates inputs required to calculate length transition matrices, for length-based per
recruit analysis, including the mean size to which a fish grows, for a  model timestep, given its initial length and specified
growth curve (MeanEndingLength), and the amount by which it grows (TimestepGrowthSizeInc)
}
\keyword{internal}
