% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetMaxLenAtAgeCV_AgeAndLengthBasedCatchCurve}
\alias{GetMaxLenAtAgeCV_AgeAndLengthBasedCatchCurve}
\title{Calculate CV for length corresponding to max age (model diagnostic)}
\usage{
GetMaxLenAtAgeCV_AgeAndLengthBasedCatchCurve(
  TimeStep,
  res,
  MaxAge,
  SelectivityType
)
}
\arguments{
\item{res}{output from AgeAndLengthBasedCatchCurvesCalcs function}

\item{MaxAge}{maximum age}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params,  3=logistic selectivity + retention params}

\item{Timestep}{model timestep}
}
\value{
ParamEst
}
\description{
Calculate CV for length corresponding to max age (model diagnostic), for
single sex, or both females and males if model is for separate sexes
}
\keyword{internal}
