% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetMeanLengthStats_DynSimMod}
\alias{GetMeanLengthStats_DynSimMod}
\title{Calculate mean length statistics for dynamic simulation model}
\usage{
GetMeanLengthStats_DynSimMod(
  nYears,
  RandObsCatchLenFreq_Fem,
  RandObsCatchLenFreq_Mal
)
}
\arguments{
\item{nYears}{number of years of length data}

\item{RandObsCatchLenFreq_Fem}{randomly generated annual length frequencies for females}

\item{RandObsCatchLenFreq_Mal}{randomly generated annual length frequencies for males}
}
\value{
MeanLengthStats
}
\description{
Calculates mean length and associated 95 percent confidence limits for
each year of generated length data from dynamic simulation model
}
\keyword{internal}
