% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetParamEst_LogisticCatchCurve}
\alias{GetParamEst_LogisticCatchCurve}
\title{Get parameter estimates and associated 95% CLs, in normal space, for age-based catch curve with
logistic selectivity}
\usage{
GetParamEst_LogisticCatchCurve(
  params,
  SelDeltaVal,
  ses,
  nSexes,
  DistnType,
  NatMortType
)
}
\arguments{
\item{params}{model parameters log(c(FMort, SelA50, SelDelta) for multinomial NLL or
log(c(FMort, SelA50, SelDelta, theta) for Dirichlet multinomial NLL}

\item{SelDeltaVal}{NA, or fixed value if not estimating this parameter}

\item{ses}{standard errors calculated for each of the estimated model parameter}

\item{nSexes}{1 or 2}

\item{DistnType}{1 = Multinomial, 2 = Dirichlet multinomial}

\item{NatMortType}{1 = fixed, 2 = estimated with prior}
}
\value{
All estimated parameters and 95 percent CLs (ParamEst), and also for estimates of individual selectivity parameters
}
\description{
This function returns estimated parameters, in normal space, for the age-based catch curve with
logistic selectivity (which are then outputted from the function GetLogisticCatchCurveResults).
}
\keyword{internal}
