% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetParamRes_AgeAndLengthBasedCatchCurve}
\alias{GetParamRes_AgeAndLengthBasedCatchCurve}
\title{Get estimates of model parameters and 95 percent confidence intervals from age and length based catch curve}
\usage{
GetParamRes_AgeAndLengthBasedCatchCurve(
  NatMortType,
  GrowthModelType,
  SelectivityType,
  nlmb,
  params,
  ses,
  sims
)
}
\arguments{
\item{NatMortType}{1 = fixed, 2 = estimated with prior to allow for uncertainty}

\item{GrowthModelType}{# 1 = von Bert single sex, 2 = von Bert, 2 sexes, 3 = Schnute - single sex, 4 = Schunte 2 sexes}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params,  3=logistic selectivity + retention params}

\item{nlmb}{output from nlminb optimisation}

\item{params}{estimated parameters, varying depending on growth model type, selectivity type and number of sexes}

\item{ses}{estimates of standard errors for estimated parameters}

\item{sims}{random values of estimated parameters drawn from multivariate normal distribution}
}
\value{
ParamEst
}
\description{
Get estimates of model parameters and associated 95 percent confidence intervals from age and length based catch curve. Outputs vary
depending on growth model type, selectivity type and number of sexes
}
\keyword{internal}
