% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetPerRecruitGrowthPredIntervals_LB}
\alias{GetPerRecruitGrowthPredIntervals_LB}
\title{Get 95 percent prediction intervals for growth curves used in length-based per recruit analysis}
\usage{
GetPerRecruitGrowthPredIntervals_LB(
  MaxModelAge,
  nLenCl,
  midpt,
  lbnd,
  ubnd,
  Res
)
}
\arguments{
\item{MaxModelAge}{maximum age considered by model}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points bounds of length classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{Res}{object containing results outputted by CalcYPRAndSPRForFMort_LB function}
}
\description{
This function outputs 95 percent prediction intervals for growth curves used in length-based per recruit analysis
}
