% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRandFishLengths_DynMod_ALB}
\alias{GetRandFishLengths_DynMod_ALB}
\title{Get random retained and discarded fish lengths}
\usage{
GetRandFishLengths_DynMod_ALB(
  nYears,
  midpt,
  ObsLenClRetCatchMidPt_Fem,
  ObsLenClRetCatchMidPt_Mal,
  ObsLenClDiscCatchMidPt_Fem,
  ObsLenClDiscCatchMidPt_Mal,
  SampleSize_Fem,
  SampleSize_Mal,
  DiscSampleSize_Fem,
  DiscSampleSize_Mal,
  SimAnnSampSize
)
}
\arguments{
\item{nYears}{number of years}

\item{midpt}{mid points of length classes}

\item{ObsLenClRetCatchMidPt_Fem}{randomly-generated observed retained catch female length class data}

\item{ObsLenClRetCatchMidPt_Mal}{randomly-generated observed retained catch male length class data}

\item{ObsLenClDiscCatchMidPt_Fem}{randomly-generated observed discarded catch female length class data}

\item{ObsLenClDiscCatchMidPt_Mal}{randomly-generated observed discarded catch male length class data}

\item{SampleSize_Fem}{specified sample size for female retained catch sample}

\item{SampleSize_Mal}{specified sample size for male retained catch sample}

\item{DiscSampleSize_Fem}{specified sample size for female discarded catch sample}

\item{DiscSampleSize_Mal}{specified sample size for male discarded catch sample}

\item{SimAnnSampSize}{specified sample size, all fish categories}
}
\value{
ObsRandLenRetCatch_Fem, ObsRandLenRetCatch_Mal, ObsRandLenRetCatch, ObsRandLenDiscCatch_Fem,
ObsRandLenDiscCatch_Mal, ObsRandLenDiscCatch
}
\description{
Get random retained and discarded fish lengths, given random length classes,
for age and length-based dynamic simulation model
}
\keyword{internal}
