% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRandomAgesAndLengths_DynSimMod_ALB}
\alias{GetRandomAgesAndLengths_DynSimMod_ALB}
\title{Get individual ages and lengths from random, conditional age at length distributions}
\usage{
GetRandomAgesAndLengths_DynSimMod_ALB(
  nYears,
  TimeStep,
  nTimeSteps,
  midpt,
  nLenCl,
  SimAnnSampSize,
  ObsRetCatchFreqLenAndAge
)
}
\arguments{
\item{nYears}{number of length classes}

\item{TimeStep}{model timestep}

\item{nTimeSteps}{number of mode time steps per year}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}

\item{ObsRetCatchFreqLenAndAge}{random age at length distributions, from simulation model}

\item{FishSampSize}{annual sample size for data to be simulated (depends on sex and retn vs disc)}
}
\value{
ObsAgeRetCatch, ObsLenRetCatch
}
\description{
Gets individual ages and lengths from conditional age at length distributions, generated
from dynamic simulation model
}
\keyword{internal}
