% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRandomParamVals_PreRecruit_LB_with_err}
\alias{GetRandomParamVals_PreRecruit_LB_with_err}
\title{Get random parameter values for use in length-based per recruit analysis}
\usage{
GetRandomParamVals_PreRecruit_LB_with_err(
  Steepness,
  Steepness_sd,
  Current_F,
  Current_F_sd,
  NatMort,
  NatMort_sd,
  ParamSims,
  nReps
)
}
\arguments{
\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{Steepness_sd}{specified error for steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{Current_F}{estimate of current fishing mortality}

\item{Current_F_sd}{error for estimate of current fishing mortality}

\item{NatMort}{natural mortality}

\item{NatMort_sd}{specified error for natural mortality}

\item{ParamSims}{random values of parameters estimated by the age-based catch curve with logistic selectivity}

\item{nReps}{number of random parameter sets from parametric resampling to generate per recruit outputs with error}
}
\value{
random parameter values
}
\description{
This function get random parameter values (M, h, F, and gear selectivity) of specified number,
and according to how these are generated as specified by the user, for use in length-based per recruit analysis
}
\keyword{internal}
