% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRecruitmentDeviations_DynSimMod}
\alias{GetRecruitmentDeviations_DynSimMod}
\title{Generate random annual recruitment deviations for dynamic data simulation model}
\usage{
GetRecruitmentDeviations_DynSimMod(nYears, lnSigmaR, autocorr)
}
\arguments{
\item{nYears}{number of years of deviations required}

\item{lnSigmaR}{natural logarithm of standard deviation of annual recruitment deviations}

\item{autocorr}{level of autocorrelation of annual recruitment deviations}
}
\value{
random_dev
}
\description{
Generates random annual recruitment deviations, with specified level of process
error and autocorrelation, for using in simulating composition data from a
dynamic model
}
\keyword{internal}
