% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetReqdSampleSizesForRandData}
\alias{GetReqdSampleSizesForRandData}
\title{Get required sample sizes for randomly-generated data}
\usage{
GetReqdSampleSizesForRandData(
  MLL,
  RetenParams,
  SampleSize,
  RetCatchAtLen_Fem,
  RetCatchAtLen_Mal,
  DiscCatchAtLen_Fem,
  DiscCatchAtLen_Mal
)
}
\arguments{
\item{MLL}{minimum legal length}

\item{RetenParams}{retention parameters}

\item{SampleSize}{size of length classes}

\item{RetCatchAtLen_Fem}{randomly-generated observed retained catch female length class data}

\item{RetCatchAtLen_Mal}{randomly-generated observed retained catch male length class data}

\item{DiscCatchAtLen_Fem}{randomly-generated observed discarded catch female length class data}

\item{DiscCatchAtLen_Mal}{randomly-generated observed discarded catch male length class data}
}
\value{
SampleSize_Fem, SampleSize_Mal, DiscSampleSize_Fem, DiscSampleSize_Mal, DiscSampleSize
}
\description{
Get required sample sizes for randomly-generated size and age data for retained and discarded catches
}
\keyword{internal}
