% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetResampLengthsAtAge_AgeLengthCatchCurve_Growth}
\alias{GetResampLengthsAtAge_AgeLengthCatchCurve_Growth}
\title{Get lengths at age and associated quantiles from resampling}
\usage{
GetResampLengthsAtAge_AgeLengthCatchCurve_Growth(
  nReps,
  nTimeSteps,
  DecAges,
  GrowthCurveType,
  RefnceAges,
  params,
  vcov.params,
  ObsRetCatchFreqAtLen
)
}
\arguments{
\item{nReps}{number of samples}

\item{DecAges}{decimal ages}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{RefnceAges}{references ages for Schnute growth curve}

\item{params}{parameter estimates}

\item{vcov.params}{variance covariance matrix for parameter estimates}

\item{ObsRetCatchFreqAtLen}{observed catch length frequency}

\item{TimeStep}{observed frequencies in length classes}
}
\value{
EstProp.sim, EstProp.sim_low, EstProp.sim_up, EstPropF.sim, EstPropF.sim_low, EstPropF.sim_up,
EstPropM.sim, EstPropM.sim_low, EstPropM.sim_up, EstLenAtAge.sim, EstLenAtAgeF.sim, EstLenAtAgeM.sim
}
\description{
Get lengths at age and associated quantiles from resampling, from multivariate normal distribution
for parameter estimates
}
\keyword{internal}
