% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRetCatchFreqAtLenAndIntAge}
\alias{GetRetCatchFreqAtLenAndIntAge}
\title{Get catch frequencies at length and integer ages}
\usage{
GetRetCatchFreqAtLenAndIntAge(
  TimeStep,
  MaxAge,
  midpt,
  nLenCl,
  ObsAgeClRetCatch_Fem,
  ObsAgeClRetCatch_Mal,
  ObsLenClRetCatchMidPt_Fem,
  ObsLenClRetCatchMidPt_Mal
)
}
\arguments{
\item{TimeStep}{model time step}

\item{MaxAge}{maximum age}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}

\item{ObsAgeClRetCatch_Fem}{generated female catch in each integer age class}

\item{ObsAgeClRetCatch_Mal}{generated male catch in each integer age class}

\item{ObsLenClRetCatchMidPt_Fem}{generated female catch in each length class}

\item{ObsLenClRetCatchMidPt_Mal}{generated male catch in each length class}
}
\value{
ObsCatchFreqAtLengthAndIntAge_Fem, ObsCatchFreqAtLengthAndIntAge_Mal, ObsCatchFreqAtLengthAndIntAge
}
\description{
Get catch frequencies in each of the length classes at each integer age
}
\keyword{internal}
