% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetSelAndMortAtAge_LogisticCatchCurve}
\alias{GetSelAndMortAtAge_LogisticCatchCurve}
\title{Get selectivity at age, and fishing and total mortality at age}
\usage{
GetSelAndMortAtAge_LogisticCatchCurve(
  params,
  EstFMort,
  Ages,
  EstSelA50,
  EstSelA95,
  EstSelA50_1,
  EstSelA95_1,
  EstSelA50_2,
  EstSelA95_2
)
}
\arguments{
\item{params}{model parameters log(c(FMort, SelA50, SelDelta) for multinomial NLL or
log(c(FMort, SelA50, SelDelta, theta) for Dirichlet multinomial NLL}

\item{EstFMort}{estimated fishing mortality}

\item{Ages}{Ages considered in analysis}

\item{EstSelA50}{selectivity parameter (single sex)}

\item{EstSelA95}{selectivity parameter (single sex)}

\item{EstSelA50_1}{selectivity parameter (two sexes)}

\item{EstSelA95_1}{selectivity parameter (two sexes)}

\item{EstSelA50_2}{selectivity parameter (two sexes)}

\item{EstSelA95_2}{selectivity parameter (two sexes)}
}
\value{
Selectivity at age (SelAtAge), fishing mortality at age (FAtAge), total mortality at age (ZAtAge),
number of sexes (nSexes)
}
\description{
This function returns values for selectivity at age, and fishing and total mortality at age
(which are then subsequently outputted in the function GetLogisticCatchCurveResults)
}
\keyword{internal}
