% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetSelectParams_AgeAndLengthBasedCatchCurvesCalcs}
\alias{GetSelectParams_AgeAndLengthBasedCatchCurvesCalcs}
\title{Get selectivity parameter values for age and length based catch curve}
\usage{
GetSelectParams_AgeAndLengthBasedCatchCurvesCalcs(
  Res,
  params,
  CatchCurveType,
  SelectivityType
)
}
\arguments{
\item{params}{estimated model parameters (varies, depending on growth curve type, selectivity type and catch curve type)}

\item{CatchCurveType}{1=length based, 2=age and length based}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}
}
\value{
values of growth parameters given specified catch curve, growth and selectivity model option
}
\description{
Get selectivity parameter values for age and length based catch curve given specified catch curve, growth and selectivity model options
}
\keyword{internal}
