% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetTheta_AgeAndLengthBasedCatchCurve}
\alias{GetTheta_AgeAndLengthBasedCatchCurve}
\title{Get Dirichlet multinomial theta value, given model specification}
\usage{
GetTheta_AgeAndLengthBasedCatchCurve(
  params,
  NatMortType,
  SelectivityType,
  GrowthModelType
)
}
\arguments{
\item{params}{parameters, including fishing mortality, growth and logistic selectivity parameters}

\item{NatMortType}{1 = fixed, 2 = estimated with prior to allow for uncertainty}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params}

\item{GrowthModelType}{1=length-based von Bertalanffy, 2=length-based Schnute,
3=age and length based von Bertalanffy, 4=age and length based Schnute}
}
\value{
DM_theta
}
\description{
Get Dirichlet multinomial theta value, given model specification
}
\keyword{internal}
