% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GillnetSelectivity_Calcs}
\alias{GillnetSelectivity_Calcs}
\title{Gillnet selectivity calculations (Kirkwood and Walker, 1986 method), used in objective function}
\usage{
GillnetSelectivity_Calcs(params)
}
\arguments{
\item{params}{theta and theta2 parameters of Kirkwood and Walker (1986) model}
}
\value{
RelNum_popn=RelNum_popn, Sel_ij, Comb_Sel, EstNum_ij, alpha_beta, alpha, beta
}
\description{
Calculate gillnet selectivity at length, by mesh and overall (assuming equal fishing intensity
among meshes) using the method of Kirkwood and Walker (1986), given model parameter values
(theta1 and theta2)
}
\keyword{internal}
