% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GillnetSelectivityObj_func}
\alias{GillnetSelectivityObj_func}
\title{Objective function for Kirkwood and Walker (1986) method}
\usage{
GillnetSelectivityObj_func(params)
}
\arguments{
\item{params}{theta and theta2 parameters of Kirkwood and Walker (1986) model}
}
\value{
NLL
}
\description{
Objective function for Kirkwood and Walker (1986) method, returns negative log-likelihood
}
