% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{InverseSchnuteGrowthfunction}
\alias{InverseSchnuteGrowthfunction}
\title{Inverse Schnute growth function}
\usage{
InverseSchnuteGrowthfunction(MaxAge, FishLen, t1, t2, y1, y2, a, b)
}
\arguments{
\item{MaxAge}{maximumm age of species to be considered by model}

\item{FishLen}{specified length}

\item{t1}{first reference age}

\item{t2}{second reference age}

\item{y1}{length at t1}

\item{y2}{length at t2}

\item{a}{growth curve parameter}
}
\value{
Age at specified length
}
\description{
Calculate age given length, from Schnute growth function
}
\keyword{internal}
