% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotAgeBasedCatchCurveResults_NormalSpace}
\alias{PlotAgeBasedCatchCurveResults_NormalSpace}
\title{Plot age based catch curve results in normal space}
\usage{
PlotAgeBasedCatchCurveResults_NormalSpace(
  RecAssump,
  SpecRecAge,
  MinFreq,
  MinAge,
  MaxAge,
  nSexes,
  DistnType,
  SelDeltaVal,
  NatMortType,
  NatMort,
  NatMort_sd,
  ObsAgeFreq,
  CatchCurveModel,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  xmax,
  xint,
  ymax,
  yint,
  PlotCLs,
  CCResults
)
}
\arguments{
\item{RecAssump}{0=age at peak frequency, 1=age at peak frequency + 1}

\item{SpecRecAge}{specified at at full recruitment, set to NA when RecAssump is set to 0 or 1}

\item{MinFreq}{minimum frequency of fish for including data for old fish}

\item{MinAge}{minimum age}

\item{MaxAge}{maximum age}

\item{nSexes}{1 or 2}

\item{DistnType}{1 = Multinomial, 2 = Dirichlet multinomial}

\item{SelDeltaVal}{NA or fixed value, if not estimating this parameter}

\item{NatMortType}{1 = fixed, 2 = estimated with prior}

\item{NatMort}{natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space}

\item{NatMort_sd}{standard deviation for natural mortality in log space. Set to NA if NatMortType = 1}

\item{ObsAgeFreq}{observed age frequency data}

\item{CatchCurveModel}{1=C&R, 2=Linear, 3=logist. sel.}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{PlotCLs}{logical (TRUE=plot 95 percent confidence limits for lines)}

\item{CCResults}{results of applied catch curve model 1, 2 or 3. Can set to NA.}
}
\value{
plot of expected vs observed proportions at age
}
\description{
This function produces plots of outputs of age-based catch curve analyses in normal space
}
\examples{
# Simulate data from Multinomial distribution (single sex)
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = unlist(as.vector(Res$CatchSample)) # input as vector for single sex model
# Specify catch curve model and required inputs for that model
# CatchCurveModel = 1 # Chapman Robson
# nSexes = NA
# DistnType = NA
# NatMortType = NA
# NatMort = NA
# NatMort_sd = NA
# RecAssump = 1
# MinFreq = NA
# CCResults = NA
# CatchCurveModel = 2 # Linear
# nSexes = NA
# DistnType = NA
# NatMortType = NA
# NatMort = NA
# NatMort_sd = NA
# RecAssump = 1
# MinFreq = 1
# CCResults = NA
CatchCurveModel = 3 # Logistic selectivity
RecAssump = NA
nSexes = 1
DistnType = 1 # 1 = multinomial, 2 = Dirichlet multinomial
SelDeltaVal = NA # NA, or fixed value if not estimating this parameter
NatMortType = 1 # 1 = fixed, 2 = estimated with prior
NatMort = exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
NatMort_sd = NA # in log space if NatMort is estimated, otherwise set to NA
MinFreq = NA # # Logistic selectivity
Init_FMort = 0.2
Init_SelA50 = 5
Init_SelDelta = 2
params = log(c(Init_FMort, Init_SelA50, Init_SelDelta))
CCResults=GetLogisticCatchCurveResults(params, nSexes, DistnType, SelDeltaVal, NatMortType, NatMort, NatMort_sd, Ages, ObsAgeFreq)
CCResults$ParamEst
PlotAgeBasedCatchCurveResults_NormalSpace(RecAssump, SpecRecAge=NA, MinFreq, MinAge, MaxAge, nSexes, DistnType, SelDeltaVal,
                                          NatMortType, NatMort, NatMort_sd, ObsAgeFreq, CatchCurveModel, MainLabel=NA,
                                          xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA, ymax=NA, yint=NA, PlotCLs=T, CCResults)

# Simulate data from Dirichlet multinomial distribution (single sex)
library(dirmult)
# J = number of fish sampling events
# K = number of age classes
# n = number of fish sampled from each sampling event
# pi = expected proportion at age
# theta = amount of autocorrelation between ages of fish within sampling events
# Fit catch curve with Dirichlet multinomial distribution objective function (single sex)
CatchCurveModel = 3 # Logistic selectivity
set.seed(123)
theta_val = 0.3
simDat = simPop(J=50, K=nAges, n=10, pi=as.vector(unlist(Res$PropAtAge[1,])), theta=theta_val)
simAges = data.frame(simDat$data)
colnames(simAges)=Ages
ObsAgeFreq = as.vector(colSums(simAges))
nSexes = 1
DistnType = 2 # 1 = multinomial, 2 = Dirichlet multinomial
SelDeltaVal = NA # NA, or fixed value if not estimating this parameter
NatMortType = 1 # 1 = fixed, 2 = estimated with prior
NatMort = exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
NatMort_sd = NA # in log space if NatMort is estimated, otherwise set to NA
Init_FMort = 0.1
Init_SelA50 = 5
Init_SelDelta = 2
Init_theta = 0.35
Init_theta_logit = log(Init_theta/(1-Init_theta)) # logit transform (so theta is always between 0 and 1)
params = c(log(Init_FMort), log(Init_SelA50), log(Init_SelDelta), Init_theta_logit)
CCResults=GetLogisticCatchCurveResults(params, nSexes, DistnType, SelDeltaVal, NatMortType, NatMort, NatMort_sd, Ages, ObsAgeFreq)
CCResults$ParamEst
CCResults$EffSampSize
PlotAgeBasedCatchCurveResults_NormalSpace(RecAssump, SpecRecAge=NA, MinFreq, MinAge, MaxAge, nSexes, DistnType, SelDeltaVal,
                                          NatMortType, NatMort, NatMort_sd, ObsAgeFreq, CatchCurveModel, MainLabel=NA,
                                          xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA, ymax=NA, yint=NA, PlotCLs=T, CCResults)

# Simulate data from Multinomial distribution (two sexes)
set.seed(123)
CatchCurveModel = 3 # Logistic selectivity
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = c(6,6.5)
SelA95 = c(8,8.5)
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = Res$CatchSample
# fit model (multinomial distribution - two sexes)
nSexes = 2
DistnType = 1 # 1 = multinomial, 2 = Dirichlet multinomial
SelDeltaVal = NA # NA, or fixed value if not estimating this parameter
NatMortType = 1 # 1 = fixed, 2 = estimated with prior
NatMort = exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
NatMort_sd = NA # in log space if NatMort is estimated, otherwise set to NA
Init_FMort = 0.2
Init_SelA50 = c(5,5)
Init_SelDelta = c(2,2.5)
params = log(c(Init_FMort, Init_SelA50, Init_SelDelta))
CCResults=GetLogisticCatchCurveResults(params, nSexes, DistnType, SelDeltaVal, NatMortType, NatMort, NatMort_sd, Ages, ObsAgeFreq)
CCResults$ParamEst
PlotAgeBasedCatchCurveResults_NormalSpace(RecAssump, SpecRecAge=NA, MinFreq, MinAge, MaxAge, nSexes, DistnType, SelDeltaVal,
                                          NatMortType, NatMort, NatMort_sd, ObsAgeFreq, CatchCurveModel, MainLabel=NA,
                                          xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA, ymax=NA, yint=NA, PlotCLs=T, CCResults)

# Simulate data from Dirichlet multinomial distribution (two sexes)
set.seed(123)
theta_val = 0.3
# Females
simDat_Fem = simPop(J=50, K=nAges, n=10, pi=as.vector(unlist(Res$PropAtAge[1,])), theta=theta_val)
simAges_Fem = data.frame(simDat_Fem$data)
colnames(simAges_Fem)=Ages
ObsAgeFreq_Fem = as.vector(colSums(simAges_Fem))
# Males
simDat_Mal = simPop(J=50, K=nAges, n=10, pi=as.vector(unlist(Res$PropAtAge[2,])), theta=theta_val)
simAges_Mal = data.frame(simDat_Mal$data)
colnames(simAges_Mal)=Ages
ObsAgeFreq_Mal = as.vector(colSums(simAges_Mal))
ObsAgeFreq = t(data.frame(ObsAgeFreq_Fem,ObsAgeFreq_Mal))
colnames(ObsAgeFreq)=Ages
# fit model (Dirichlet multinomial distribution - two sexes)
nSexes = 2
DistnType = 2 # 1 = Multinomial, 2 = Dirichlet multinomial
SelDeltaVal = NA # NA, or fixed value if not estimating this parameter
NatMortType = 1 # 1 = fixed, 2 = estimated with prior
NatMort = exp(1.46 - (1.01 * (log(MaxAge))))
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1
Init_FMort = 0.1
Init_SelA50 = c(5,5)
Init_SelDelta = c(2.5,2.5)
Init_theta = 0.3
Init_theta_logit = log(Init_theta/(1-Init_theta)) # logit transform (so theta is always between 0 and 1)
params = c(log(Init_FMort), log(Init_SelA50), log(Init_SelDelta), Init_theta_logit)
CCResults=GetLogisticCatchCurveResults(params, nSexes, DistnType, SelDeltaVal, NatMortType, NatMort, NatMort_sd, Ages, ObsAgeFreq)
CCResults$ParamEst
CCResults$EffSampSize
PlotAgeBasedCatchCurveResults_NormalSpace(RecAssump, SpecRecAge=NA, MinFreq, MinAge, MaxAge, nSexes, DistnType, SelDeltaVal,
                                          NatMortType, NatMort, NatMort_sd, ObsAgeFreq, CatchCurveModel, MainLabel=NA,
                                          xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA, ymax=NA, yint=NA, PlotCLs=T, CCResults)
}
