% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotAgeLengthCatchCurve_Pears_Resid}
\alias{PlotAgeLengthCatchCurve_Pears_Resid}
\title{Plot Pearson residuals for proportions at age in each length class, from age and length catch curve model}
\usage{
PlotAgeLengthCatchCurve_Pears_Resid(
  params,
  RefnceAges,
  MLL,
  DistnType,
  nSexes,
  GrowthCurveType,
  SelectivityType,
  ObsRetCatchFreqAtLen,
  ObsRetCatchFreqAtLengthAndAge,
  lbnd,
  ubnd,
  midpt,
  SelectivityAtLen,
  DiscMort,
  MaxAge,
  NatMortType,
  NatMort,
  NatMort_sd,
  TimeStep,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  xmin,
  xmax,
  xint,
  ymin,
  ymax,
  yint,
  CircleScale,
  MinLenClFreq,
  ShowLegend,
  FittedRes
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated (if FittedRes=NA)}

\item{RefnceAges}{Reference ages for Schnute growth curve (set to NA for von Bertalanffy growth curve)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{DistnType}{1 = Multinomial, 2 = Dirichlet multinomial  (for marginal length comp)}

\item{nSexes}{1 or 2 (for estimating growth parameters)}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params,  3=logistic selectivity + retention params}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{ObsRetCatchFreqAtLengthAndAge}{observed frequencies in length and age classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityAtLen}{selectivity at length}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{MaxAge}{maximum age considered in model}

\item{NatMortType}{1 = fixed, 2 = estimated with prior to allow for uncertainty}

\item{NatMort}{natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space}

\item{NatMort_sd}{standard deviation for natural mortality in log space. Set to NA if NatMortType = 1}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{CircleScale}{size of number increases or decreases sizes of circles for Pearson residuals}

\item{MinLenClFreq}{minimum number of fish in length class to show Pearson residuals}

\item{FittedRes}{saved results from GetAgeAndLengthBasedCatchCurveResults function (model will be refitted if set to NA)}
}
\value{
Plot of Pearson residuals for fit of age and length based catch curve
}
\description{
This function provides a plot of Pearson residuals for observed and expected proportions at age in each length class,
from the age and length catch curve model. The model is fitted to a sample of fish length and age data, by minimising
the overall negative log-likelihood, including the NLL associated with the marginal length composition and a conditional age at length NLL,
given the parameters (selectivity, growth and mortality) and data, using nlminb.
}
\examples{
# Simulate data
set.seed(123)
SampleSize=5000
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMortType = 1 # 1 = fixed, 2 = estimated with prior to allow for uncertainty
NatMort = 4.22/MaxAge # natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1.
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
RefnceAges = NA
GrowthParams = c(Linf, vbK, CVSizeAtAge)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# 2 sexes, von Bertalanffy
# GrowthCurveType = 1 # 1 = von Bertalanffy
# Linf = c(700,850)
# vbK = c(0.25,0.2)
# CVSizeAtAge = c(0.05,0.05)
# RefnceAges = NA
# GrowthParams = data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge)
# Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
#                          SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = as.vector(Res$ObsRetCatchFreqAtLen) # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# # get data - 2 sexes
# ObsRetCatchFreqAtLen <- data.frame(matrix(nrow = 2, ncol = length(midpt))) # 2 sex
# colnames(ObsRetCatchFreqAtLen) <- midpt
# ObsRetCatchFreqAtLen[1,] = Res$ObsRetCatchFreqAtLen_Fem
# ObsRetCatchFreqAtLen[2,] = Res$ObsRetCatchFreqAtLen_Mal
# ObsRetCatchFreqAtLengthAndAge = array(c(unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem), unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Mal)),
#                                       c(nTimeSteps, length(midpt), 2), dimnames=list(rownames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem),
#                                                                                      colnames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem)))
# get params - 1 sex
nSexes = 1
InitFishMort = 0.3 # specify starting parameters
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
# # get params - 2 sexes
# nSexes = 2
# InitFishMort = 0.3 # specify starting parameters
# InitL50 = 320
# InitDelta = 50 # L95-L50
# InitLinf = c(800,800)
# InitvbK = c(0.25,0.25)
# InitCVSizeAtAge = 0.05
# InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
# params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)
# Example with specified selectivity vector
# Simulate data
SelectivityType=1 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = 1 / (1 + exp(-log(19)*(midpt-400)/(500-400)))
SelParams = c(NA, NA) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial (for marginal length comp)
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
GrowthParams = c(Linf, vbK, CVSizeAtAge)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = as.vector(Res$ObsRetCatchFreqAtLen) # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# get params - 1 sex
nSexes = 1
InitFishMort = 0.3 # specify starting parameters
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort / (2 - InitFishMort)) # logit transform (so F is always between 0 and 2)
params = c(InitFishMort_logit, log(c(InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep)

PlotAgeLengthCatchCurve_Pears_Resid(params, RefnceAges, MLL, DistnType, nSexes, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityAtLen, DiscMort, MaxAge, NatMortType, NatMort, NatMort_sd, TimeStep, MainLabel=NA,
                                                xaxis_lab=NA, yaxis_lab=NA, xmin=NA, xmax=NA, xint=NA, ymin=NA, ymax=NA, yint=NA, CircleScale=5, MinLenClFreq=10, ShowLegend=T, FittedRes)

}
