% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotLenConvCatchCurveResults}
\alias{PlotLenConvCatchCurveResults}
\title{Plot results for specified length-converted catch model}
\usage{
PlotLenConvCatchCurveResults(
  MaxAge,
  ModelType,
  GrowthParams,
  RefnceAges,
  ObsRetCatchFreqAtLen,
  MinFreq,
  lbnd,
  midpt,
  ubnd
)
}
\arguments{
\item{MaxAge}{maximum age}

\item{ModelType}{1=von Bertalanffy growth, 2=Schnute growth}

\item{GrowthParams}{c(Linf, vbK) von Bertalanffy growth or c(y1, y2, a, b) Schnute growth}

\item{RefnceAges}{Schnute growth curve reference ages}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{MinFreq}{minimum frequency to include}

\item{lbnd}{lower bounds of length classes}

\item{midpt}{mid points of length classes}

\item{ubnd}{upper bounds of length classes}
}
\value{
plots associated with fitted length-converted catch curve models
}
\description{
Plot results for specified length-converted catch curve model, including either Pauly's method,
or a model allowing for Schnute growth with  logistic selectivity or knife edge selectivity
}
\examples{
# Simulate data
SampleSize=1000
set.seed(123)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1500
LenInc = 50
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityVec = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
# # von Bertalanffy
# GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
# Linf = 800
# vbK = 0.2
# CVSizeAtAge = 0.08
# GrowthParams = c(Linf, vbK)
# RefnceAges = NA
# Schnute
GrowthCurveType = 2 # 1 = von Bertalanffy, 2 = Schnute
t1 = 1 # growth - Schnute
t2 = 10 # growth - Schnute
y1 = 400 # growth - Schnute
y2 = 1000 # growth - Schnute
a = 0.1 # growth - Schnute
b = 2.0 # growth - Schnute
CVSizeAtAge = 0.08
GrowthParams = c(y1, y2, a, b)
RefnceAges = c(t1,t2)
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
ObsRetCatchFreqAtLen = as.vector(Res$ObsRetCatchFreqAtLen)
MinFreq = 20 # set minimum frequency for larger lengths for analysis
# note, this needs to be high enough so that data for ln(n/dt) vs relative age for essentially straight
# line - if not, Z will be biased!!!
midpt=Res$midpt
lbnd=Res$lbnd
ubnd=Res$ubnd
ModelType = 2 # 1 = von Bertalanffy growth curve (Pauly), 2 = length-converted catch curve - Schnute growth curve)
res=GetLenConvCatchCurveResults(ModelType, GrowthParams, RefnceAges, ObsRetCatchFreqAtLen,
                               MinFreq, lbnd, midpt, ubnd, MaxAge)
PlotLenConvCatchCurveResults(MaxAge, ModelType, GrowthParams, RefnceAges, ObsRetCatchFreqAtLen, MinFreq,
                            lbnd, midpt, ubnd)
}
