% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotPerRecruit_Param_Err_Distns}
\alias{PlotPerRecruit_Param_Err_Distns}
\title{Plot assumed error distributions for M, F and h in per recruit analysis allowing for error}
\usage{
PlotPerRecruit_Param_Err_Distns(
  NatMort,
  NatMort_sd,
  Current_F,
  Current_F_sd,
  Steepness,
  Steepness_sd
)
}
\arguments{
\item{NatMort}{mean for natural mortality}

\item{NatMort_sd}{sd for natural mortality}

\item{Current_F}{mean for fishing mortality}

\item{Current_F_sd}{sd for fishing mortality}

\item{Steepness}{mean for steepness of stock recruitment relationship}

\item{Steepness_sd}{sd for steepness of stock recruitment relationship}
}
\description{
This function plots assumed error distributions for M, F and h in per recruit analysis allowing for error, based on user inputs
for mean and sd of these parameters. Parameters are assumed to be normally distributed.
}
