% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotPerRecruit_RiskMatrix_RelBiom}
\alias{PlotPerRecruit_RiskMatrix_RelBiom}
\title{Produces WA risk matrix for relative biomass estimates from per recruit analysis, and probabilities}
\usage{
PlotPerRecruit_RiskMatrix_RelBiom(RefPointOpt, FittedRes, nReps)
}
\arguments{
\item{RefPointOpt}{1=proxy, 2=BMSY ratio reference points}

\item{FittedRes}{save results from age or length based per recruit analysis with err (get functions)}

\item{nReps}{number of estimates of relative biomass from resampling}
}
\value{
plot of risk matrix for relative biomass, likelihoods for consequence levels, and risks for consequence
levels and overall
}
\description{
This function outputs a risk matrix plot Produces WA risk matrix for relative biomass estimates from
age or length based per recruit analysis (allowing for error in h, M and F), and associated
likelihoods and risk levels (for each consequence level and overall)
}
\examples{
InitRecruit <- 1 # Initial recruitment
MaxModelAge <- 20 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be appropriate to use a smaller time step)
Linf <- c(550, 500) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.5, 0.5) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
EstLenAtAge <- data.frame(EstFemLenAtAge=NA, EstMalLenAtAge=NA) # length at age (from age 0), inputted as values in data frame
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtAge <- data.frame(EstFemWtAtAge=NA, EstMalWtAtAge=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at age zero
FinalSex_A50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_A95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
EstSexRatioAtAge <- NA  # sex ratio at age (from age 0) inputted as vector
EggFertParam <- NA # (NA or from ~0.2-1) NA = no effect, 0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects
mat_A50 <- c(2.5, 2.5) # females, males - Logistic length (mm) at maturity relationship parameters
mat_A95 <- c(3.5, 3.5) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtAge <- data.frame(EstFemMatAtAge=NA, EstMalMatAtAge=NA) # maturity at age (from age 0), inputted as values in data frame
Gear_sel_A50 <- c(2.5, 2.5) # females, males - Logistic age selectivity relationship parameters
Gear_sel_A95 <- c(3.5, 3.5) # females, males - Logistic age selectivity relationship parameters
EstGearSelAtAge <- data.frame(FemGearSelAtAge=NA, MalGearSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
Land_sel_A50 <- c(2.5, 2.5) # females, males - Logistic age selectivity relationship parameters
Land_sel_A95 <- c(3.5, 3.5) # females, males - Logistic age selectivity relationship parameters
EstLandSelAtAge <- data.frame(FemSelLandAtAge=NA, MalSelLandAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
ret_Pmax <- NA  # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_A50 <- NA  # females, males - Logistic age fish retention at age parameters
ret_A95 <- NA  # females, males - Logistic age fish retention at age parameters
EstRetenAtAge <- data.frame(EstFemRetenAtAge=rep(1,MaxModelAge+1), EstMalRetenAtAge=rep(1,MaxModelAge+1)) # fish retention at age (from age 0), inputted as values in data frame
DiscMort <- 0.0 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
Steepness_sd <- 0.025
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort <- 0.2 # natural mortality  (year-1)
NatMort_sd <- 0.025
Current_F <- 0.3 # estimate of fishing mortality, e.g. from catch curve analysis
Current_F_sd <- 0.005
PlotOpt <- 1 # 1=females, 2=males, 3=combined sex
RefPointPlotOpt <- 1 # 0=don't plot, 1=plot defaults, 2=plot BMSY ref points
nReps=25
FittedRes=GetPerRecruitResults_AB_with_err(MaxModelAge, TimeStep, Linf, vbK, tzero, EstLenAtAge, lenwt_a, ln_lenwt_a,
                                 lenwt_b, WLrel_Type, EstWtAtAge, ReprodScale, ReprodPattern, InitRatioFem,
                                 FinalSex_A50, FinalSex_A95, EstSexRatioAtAge, EggFertParam, mat_A50, mat_A95,
                                 EstMatAtAge, Gear_sel_A50, Gear_sel_A95, EstGearSelAtAge, Land_sel_A50, Land_sel_A95,
                                 EstLandSelAtAge, ret_Pmax, ret_A50, ret_A95, EstRetenAtAge, DiscMort, Steepness,
                                 Steepness_sd, SRrel_Type, NatMort, NatMort_sd, Current_F, Current_F_sd, nReps)
PlotPerRecruit_RiskMatrix_RelBiom(RefPointOpt, FittedRes, nReps)
}
