% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotRiskMatrix_CatchCurve_FishMort}
\alias{PlotRiskMatrix_CatchCurve_FishMort}
\title{Produces WA risk matrix for fishing mortality from catch curve analysis, and probabilities}
\usage{
PlotRiskMatrix_CatchCurve_FishMort(
  EstFMort,
  EstFMort_se,
  Sim_FVals,
  F_targ,
  F_thresh,
  F_lim
)
}
\arguments{
\item{EstFMort}{Fishing mortality point estimate}

\item{EstFMort_se}{standard error for fishing mortality estimate}

\item{Sim_FVals}{Simulated fishing mortality values (set to NA, or input as vector, which will override EstFMort, EstFMort_se)}

\item{F_targ}{target reference value for fishing mortality}

\item{F_thresh}{threshold reference value for fishing mortality}

\item{F_lim}{limit reference value for fishing mortality}
}
\value{
plot of risk matrix for fishing mortality, likelihoods for consequence levels, and risks for consequence
levels and overall
}
\description{
This function outputs a risk matrix plot Produces WA risk matrix for fishing mortality estimates from age-based,
length-based or age and length-based catch curve analysis analysis, and associated
likelihoods and risk levels (for each consequence level and overall)
}
\examples{
NatMort = 0.15
EstFMort = 0.15
EstFMort_se = 0.025
Sim_FVals = NA
F_targ = 2/3*NatMort
F_thresh = NatMort
F_lim = 3/2*NatMort
PlotRiskMatrix_CatchCurve_FishMort(EstFMort, EstFMort_se, Sim_FVals,
                                   F_targ, F_thresh, F_lim)
}
