% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotRiskMatrix_PerRecruit_RelBiom}
\alias{PlotRiskMatrix_PerRecruit_RelBiom}
\title{Produces WA risk matrix for relative biomass estimates from per recruit analysis, and probabilities}
\usage{
PlotRiskMatrix_PerRecruit_RelBiom(
  EstBrel,
  EstBrel_se,
  Sim_BrelVals,
  B_targ,
  B_thresh,
  B_lim
)
}
\arguments{
\item{EstBrel}{relative biomass point estimate}

\item{EstBrel_se}{standard error for relative biomass estimate}

\item{Sim_BrelVals}{Simulated relative biomass values (set to NA, or input as vector, which will override EstBrel, EstBrel_se)}

\item{B_targ}{target reference value for relative biomass}

\item{B_thresh}{threshold reference value for relative biomass}

\item{B_lim}{limit reference value for relative biomass}
}
\value{
plot of risk matrix for relative biomass, likelihoods for consequence levels, and risks for consequence
levels and overall
}
\description{
This function outputs a risk matrix plot Produces WA risk matrix for relative biomass estimates from
age or length based per recruit analysis and associated likelihoods and risk levels (for each consequence
level and overall)
}
\examples{
EstBrel = 0.32
EstBrel_se = 0.05
Sim_BrelVals = NA
B_targ = 0.4
B_thresh = 0.3
B_lim = 0.2
PlotRiskMatrix_PerRecruit_RelBiom(EstBrel, EstBrel_se, Sim_BrelVals,
                                  B_targ, B_thresh, B_lim)
}
