% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SchnuteGrowthfunction}
\alias{SchnuteGrowthfunction}
\title{Schnute growth function}
\usage{
SchnuteGrowthfunction(Age, t1, t2, y1, y2, a, b)
}
\arguments{
\item{Age}{specified age}

\item{t1}{first reference age}

\item{t2}{second reference age}

\item{y1}{length at t1}

\item{y2}{length at t2}

\item{a}{growth curve parameter}
}
\value{
Estimated length at a specified age
}
\description{
Calculates estimated length at age from Schnute growth curve, given the age, growth parameters and
reference ages
}
\keyword{internal}
