% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SimAgeFreqData_DynMod_AB}
\alias{SimAgeFreqData_DynMod_AB}
\title{Simulate length composition data from a length-based dynamic model with recruitment variation}
\usage{
SimAgeFreqData_DynMod_AB(
  SimAnnSampSize,
  nYears,
  lnSigmaR,
  autocorr,
  InitRec,
  MaxModelAge,
  TimeStep,
  Linf,
  vbK,
  tzero,
  EstLenAtAge,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtAge,
  ReprodScale,
  ReprodPattern,
  InitRatioFem,
  FinalSex_A50,
  FinalSex_A95,
  EstSexRatioAtAge,
  EggFertParam,
  mat_A50,
  mat_A95,
  EstMatAtAge,
  Gear_sel_A50,
  Gear_sel_A95,
  EstGearSelAtAge,
  Land_sel_A50,
  Land_sel_A95,
  EstLandSelAtAge,
  ret_Pmax,
  ret_A50,
  ret_A95,
  EstRetenAtAge,
  DiscMort,
  Steepness,
  SRrel_Type,
  NatMort,
  FMortByYear
)
}
\arguments{
\item{SimAnnSampSize}{specified sample size (for both sexes)}

\item{nYears}{number of years of deviations required}

\item{lnSigmaR}{natural logarithm of standard deviation of annual recruitment deviations}

\item{autocorr}{level of autocorrelation of annual recruitment deviations}

\item{InitRec}{initial recruitment (R0)}

\item{MaxModelAge}{maximum age considered by model}

\item{TimeStep}{model time step (in y)}

\item{Linf}{von Bertalanffy growth parameter}

\item{vbK}{von Bertalanffy growth parameter}

\item{tzero}{von Bertalanffy growth parameter}

\item{EstLenAtAge}{input vector length at age}

\item{lenwt_a}{weight-length parameter}

\item{ln_lenwt_a}{weight-length parameter}

\item{lenwt_b}{weight-length parameter}

\item{WLrel_Type}{1=power, 2=log-log}

\item{EstWtAtAge}{weight at ages}

\item{ReprodScale}{1=default (standard calculations for spawning biomass), use alternative number to specify reproductive scaling with female mass factor (i.e. BOFFF effects)}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{proportion of fish that are females at hatching}

\item{FinalSex_A50}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_A95}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{EstSexRatioAtAge}{NA  # sex ratio at age, inputted as vector}

\item{EggFertParam}{<- NA # (NA or from ~0.2-1) NA = no effect, 0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects}

\item{mat_A50}{logistic age at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{mat_A95}{logistic age at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{EstMatAtAge}{vector of proportion mature at age (set to NA if using age at maturity parameters)}

\item{EstGearSelAtAge}{gear selectivity curve inputted as vector}

\item{ret_Pmax}{logistic parameter for fish retention curve}

\item{ret_A50}{logistic parameter for fish retention curve}

\item{ret_A95}{logistic parameter for fish retention curve}

\item{EstRetenAtAge}{retention curve inputted as vector}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{SRrel_Type}{1 = Beverton-Holt, 2=Ricker stock-recruitment relationship}

\item{NatMort}{natural mortality}

\item{FMortByYear}{annual fully-selected fishing mortality}

\item{sel_A50}{logistic parameter for gear selectivity curve}

\item{sel_A95}{logistic parameter for gear selectivity curve}
}
\value{
catches in numbers for females, males and combined sexes (Catch_Fem, Catch_Mal, Catch_CombSex),
catch biomass for combined sexes (Catch_Biom), female, male and combined sex spawning biomass (FemSpBiom,
MalSpBiom, CombSexSpBiom), annual recruitment (AnnRecruit), recruitment deviations (random_dev),
simulated sample sizes for females and males, given combined sex imput and model assumptions (AnnSimSampSize_Fem,
AnnSimSampSize_Mal), randomly-generated age frequencies, by year, for females, males and combined sexes
(RandObsCatchAgeFreq_Fem, RandObsCatchAgeFreq_Mal, RandObsCatchAgeFreq_CombSex), age clases (Ages), mean and 95 percent
confidence limits for each year, for simulated age data (MeanAgeStats)
}
\description{
This function provides simulated length composition data from a length-based dynamic model
with specified annual fishing and (natural mortality), annual recruitment with specified process
error and autocorrelation.
}
\examples{
InitRecruit <- 1 # Initial recruitment
MaxModelAge <- 20 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be approprriate to use a smaller time step)
Linf <- c(550, 500) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.5, 0.5) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
EstLenAtAge <- data.frame(EstFemLenAtAge=NA, EstMalLenAtAge=NA) # length at age (from age 0), inputted as values in data frame
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtAge <- data.frame(EstFemWtAtAge=NA, EstMalWtAtAge=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at age zero
FinalSex_A50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_A95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
EstSexRatioAtAge <- NA  # sex ratio at age (from age 0) inputted as vector
EggFertParam <- NA # (NA or from ~0.2-1) NA = no effect, ~0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects
mat_A50 <- c(2.5, 2.5) # females, males - Logistic length (mm) at maturity relationship parameters
mat_A95 <- c(3.5, 3.5) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtAge <- data.frame(EstFemMatAtAge=NA, EstMalMatAtAge=NA) # maturity at age (from age 0), inputted as values in data frame
Gear_sel_A50 <- NA # females, males - Logistic age selectivity relationship parameters
Gear_sel_A95 <- NA # females, males - Logistic age selectivity relationship parameters
EstGearSelAtAge <- data.frame(FemGearSelAtAge=NA, MalGearSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
Land_sel_A50 <- c(2.5, 2.5) # females, males - Logistic age selectivity relationship parameters
Land_sel_A95 <- c(3.5, 3.5) # females, males - Logistic age selectivity relationship parameters
EstLandSelAtAge <- data.frame(FemSelLandAtAge=NA, MalSelLandAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
ret_Pmax <- NA  # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_A50 <- NA  # females, males - Logistic age fish retention at age parameters
ret_A95 <- NA  # females, males - Logistic age fish retention at age parameters
EstRetenAtAge <- data.frame(FemRetProbAtAge=NA, MalRetProbAtAge=NA) # fish retention at age (from age 0), inputted as values in data frame
DiscMort <- 0.0 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort = 0.2 # natural mortality  (year-1)
InitRec <- 1000
lnSigmaR <- 0.6
autocorr <- 0.3
# get exploitation history - initial in increase F, reduction in F then stable F period
nPeriods <- 3 # number of periods of mortality trends
InitYr <- 1970 # first year of mortality
PeriodEndYr <- c(1995,2000,2020) # periods when mortality changes
InitYr_FMort <- 0.33*NatMort # initial mortality, at beginning of first period
PeriodEndYr_FMort <- c(3*NatMort,0.67*NatMort,0.67*NatMort)
F_autocorr <- 0.2
F_randerr_sd <- 0.1
set.seed(123)
res=SimExploitationHistory_DynMod(nPeriods, InitYr, PeriodEndYr, InitYr_FMort,
                                  PeriodEndYr_FMort, F_autocorr, F_randerr_sd)
# Get random age data
nYears <- res$nYears
FMortByYear <- res$rand_FMort
SimAnnSampSize <- 1000
res=SimAgeFreqData_DynMod_AB(SimAnnSampSize, nYears, lnSigmaR, autocorr, InitRec, MaxModelAge, TimeStep, Linf, vbK, tzero, EstLenAtAge,
                             lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtAge, ReprodScale,
                             ReprodPattern, InitRatioFem, FinalSex_A50, FinalSex_A95, EstSexRatioAtAge,
                             EggFertParam, mat_A50, mat_A95, EstMatAtAge, Gear_sel_A50, Gear_sel_A95,
                             EstGearSelAtAge, Land_sel_A50, Land_sel_A95, EstLandSelAtAge, ret_Pmax,
                             ret_A50, ret_A95, EstRetenAtAge, DiscMort, Steepness, SRrel_Type, NatMort, FMortByYear)
}
