% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SimAgeFreqData_EqMod}
\alias{SimAgeFreqData_EqMod}
\title{Simulate age frequency data}
\usage{
SimAgeFreqData_EqMod(
  SampleSize,
  MinAge,
  MaxAge,
  SelA50,
  SelA95,
  NatMort,
  FMort
)
}
\arguments{
\item{SampleSize}{required sample size}

\item{MinAge}{minimum age}

\item{MaxAge}{maximum age}

\item{SelA50}{age at 50 percent selectivity}

\item{SelA95}{age at 95 percent selectivity}

\item{NatMort}{natural mortality}

\item{FMort}{fully-selected fishing mortality}
}
\value{
Age classes (Ages), selectivity at age (SelAtAge), expected proportions at age (PropAtAge), observed
catch frequency at age (CatchSample)
}
\description{
Simulate age frequency data with specified selectivity and mortality parameters
}
\examples{
# simulate data
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
# single sex
SelA50 = 6
SelA95 = 8
# two sexes
# assuming same natural and fishing mortality,
# but potentially different age-based selectivity, e.g. due to different growth
# SelA50 = c(6,5.5)
# SelA95 = c(8,7.5)
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = Res$CatchSample
}
