% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SimExploitationHistory_DynMod}
\alias{SimExploitationHistory_DynMod}
\title{Generate an exploitation history, for use in generating length data from a dynamic
length-structured model}
\usage{
SimExploitationHistory_DynMod(
  nPeriods,
  InitYr,
  PeriodEndYr,
  InitYr_FMort,
  PeriodEndYr_FMort,
  autocorr,
  randerr_sd
)
}
\arguments{
\item{nPeriods}{number of specified exploitation periods with different trends}

\item{InitYr}{first year of first exploitation period}

\item{PeriodEndYr}{final year of each exploitation period}

\item{InitYr_FMort}{fishing mortality at beginning of first exploitation period}

\item{PeriodEndYr_FMort}{fishing mortality in final year of each exploitation period}

\item{autocorr}{specified level of autocorrelation}

\item{randerr_sd}{random error used to generate autocorrelation}

\item{FMort_SD}{random error for fishing mortality deviations}
}
\value{
total number of years of exploitation (nYears), fishing mortality trend
without error (FMort), fishing mortality tren with error (rand_FMort)
}
\description{
Generate exploitation history, for dynamic length simulation model a dynamic length-structured
model (using SimLenFreqData_DynMod_LB). Values of fishing mortality are generated
given specified number of exploitation periods, fishing mortality at the beginning
and end of each period, with values in between for each period linearly interpolated.
Random autocorrelated errord for annual fishing mortality values can also be specified.
}
\examples{
# example exploitation history - initial in increase F, reduction in F then stable F period
nPeriods = 3 # number of periods of mortality trends
InitYr = 1970 # first year of mortality
NatMort = 4.22 / 41 # natural mortality  (year-1)
PeriodEndYr = c(1995,2000,2020) # periods when mortality changes
InitYr_FMort = 0.33*NatMort # initial mortality, at beginning of first period
PeriodEndYr_FMort = c(3*NatMort,0.67*NatMort,0.67*NatMort)
autocorr = 0.2
randerr_sd = 0.01
set.seed(123)
res=SimExploitationHistory_DynMod(nPeriods, InitYr, PeriodEndYr, InitYr_FMort,
                                  PeriodEndYr_FMort, autocorr, randerr_sd)
# example exploitation history -  stable F with no error
nPeriods = 1 # number of periods of mortality trends
InitYr = 1970 # first year of mortality
NatMort = 4.22 / 41 # natural mortality  (year-1)
InitYr_FMort = NatMort # initial mortality, at beginning of first period
PeriodEndYr_FMort = NatMort
autocorr = 0
randerr_sd = 0.01
set.seed(123)
res=SimExploitationHistory_DynMod(nPeriods, InitYr, PeriodEndYr, InitYr_FMort,
                                  PeriodEndYr_FMort, autocorr, randerr_sd)
}
