% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SimLenAndAgeFreqData_EqMod}
\alias{SimLenAndAgeFreqData_EqMod}
\title{Simulate length and age data}
\usage{
SimLenAndAgeFreqData_EqMod(
  SampleSize,
  MaxAge,
  TimeStep,
  NatMort,
  FishMort,
  MaxLen,
  LenInc,
  MLL,
  SelectivityType,
  SelParams,
  RetenParams,
  SelectivityAtLen,
  DiscMort,
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  CVSizeAtAge
)
}
\arguments{
\item{SampleSize}{required same size, single value generates this number of retained fish, and routines
then determines an expected sample size for discarded fish based on input assumptions for selectivity and
retention. Can also specify vector, e.g. SampleSize=c(10000, 2000) produces 10000 retained fish, 2000 discarded fish}

\item{MaxAge}{maximum age}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{NatMort}{natural mortality}

\item{FishMort}{fully-selected fishing mortality}

\item{MaxLen}{maximum length}

\item{LenInc}{length class interval}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{SelectivityAtLen}{selectivity at length, set to NA if selectivity parameters inputted}

\item{DiscMort}{Proportion of fish that die following to capture and release}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{c(Linf, vbK) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK) separate sex von Bertalanffy,
c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b) separate sex Schnute}

\item{RefnceAges}{Schnute reference ages, either c(t1,t2) single sex, c(t1=t1, t2=t2) separate sex, or set to NA if using von Bertalanffy model}

\item{CVSizeAtAge}{coefficient of variation for mean lengths at age}

\item{L50}{length at 50 percent selectivity}

\item{L95}{length at 95 percent selectivity}
}
\value{
Observed catch frequency at age, ObsCatchFreqAtAge (Vector), Observed catch proportions at
age, ObsRelCatchAtAge (Vectors), Observed catch frequency at length, ObsRetCatchFreqAtLen (Vector),
observed catch proportions at length, ObsRelCatchAtLen (Vectors), Observed catch frequency at length
and age, ObsRetCatchFreqAtLengthAndAge (Matrix), age classes, ObsAgeCl (Vector), length class
mid points, ObslenClMidPt (Vector)
}
\description{
Simulate length and age data, given specified growth, mortality and selectivity
}
\examples{
# Simulate data
set.seed(123)
SampleSize=5000 # sample size for retained catches (and same number for released fish, if an MLL is specified)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.08
GrowthParams = c(Linf, vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# 2 sexes, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = c(700,850)
vbK = c(0.3,0.2)
CVSizeAtAge = c(0.08,0.08)
GrowthParams = data.frame(Linf=Linf, vbK=vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# 1 sex, Schnute
GrowthCurveType = 2 # 1 = von Bertalanffy, 2 = Schnute
t1 = 0.5 # growth - Schnute
t2 = 25 # growth - Schnute
y1 = 100 # growth - Schnute
y2 = 1000 # growth - Schnute
a = 0.02 # growth - Schnute
b = 3.0 # growth - Schnute
GrowthParams = c(y1, y2, a, b)
RefnceAges = c(t1,t2)
CVSizeAtAge = 0.05
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# 2 sexes, Schnute
GrowthCurveType = 2 # 1 = von Bertalanffy, 2 = Schnute
t1 = c(0.5,0.5) # growth - Schnute
t2 = c(25,25) # growth - Schnute
y1 = c(100,100) # growth - Schnute
y2 = c(1000,1000) # growth - Schnute
a = c(0.02,0.02) # growth - Schnute
b = c(3,3) # growth - Schnute
CVSizeAtAge = c(0.05, 0.05)
GrowthParams = data.frame(y1=y1, y2=y2, a=a, b=b)
RefnceAges = data.frame(t1=t1,t2=t2)
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# Example with selectivity specified as a vector
# Simulate data
SampleSize=1000
set.seed(123)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=1 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
lbnd = seq(0,MaxLen - LenInc, LenInc)
midpt = lbnd + (LenInc/2)
# # single sex selectivity input
# SelectivityAtLen = 1 / (1 + exp(-log(19)*(midpt-400)/(500-400)))
# two sex selectivity input
FemSelAtLen = 1 / (1 + exp(-log(19)*(midpt-400)/(500-400)))
MalSelAtLen = 1 / (1 + exp(-log(19)*(midpt-450)/(550-450)))
SelectivityAtLen = t(data.frame(FemSelAtLen=FemSelAtLen,MalSelAtLen=MalSelAtLen))
colnames(SelectivityAtLen) = midpt
SelParams = c(NA, NA) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
# 2 sexes, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = c(700,850)
vbK = c(0.3,0.2)
CVSizeAtAge = c(0.06,0.06)
GrowthParams = data.frame(Linf=Linf, vbK=vbK)
RefnceAges = NA
SimRes=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                            SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
}
