% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SimLenFreqDat_DirMultDistn_EqMod}
\alias{SimLenFreqDat_DirMultDistn_EqMod}
\title{Get length frequency sample data simulated from the Dirichlet multinomial distribution.}
\usage{
SimLenFreqDat_DirMultDistn_EqMod(
  nSampEvents,
  nFishPerSampEvent,
  theta_val,
  midpt,
  ExpPropAtLen
)
}
\arguments{
\item{nSampEvents}{minimum legal length}

\item{nFishPerSampEvent}{retention parameters}

\item{theta_val}{size of length classes}

\item{midpt}{randomly-generated observed retained catch female length class data}

\item{ExpPropAtLen}{randomly-generated observed retained catch male length class data}
}
\value{
midpt, simSampEventLenDat, simLenFreq, TotSampSize
}
\description{
Get length frequency data simulated from the Dirichlet multinomial distribution. Using this
distribution allows for auto-correlation in length frequency data associated with, for example,
fish schooling (according to size) behaviours. This routine uses the dirmult package. Expected
proportions in each length class may be calculated using the SimLenAndAgeFreqData_EqMod function.
}
\examples{
# Simulate length-frequency data from Dirchlet multinomial distribution.
# First, use SimLenAndAgeFreqData_EqMod function to calculate expected proportions at length
set.seed(123)
SampleSize=5000 # sample size for retained catches (and same number for released fish, if an MLL is specified)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityAtLen = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(500, 50) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.08
GrowthParams = c(Linf, vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# Now simulate length-frequency data from Dirchlet multinomial distribution.
set.seed(123)
nSampEvents = 50
nFishPerSampEvent = 20
theta_val = 0.3
midpt = Res$midpt
ExpPropAtLen = Res$ModelDiag$ExpRetCatchPropAtLen
res=SimLenFreqDat_DirMultDistn_EqMod(nSampEvents, nFishPerSampEvent, theta_val, midpt, ExpPropAtLen)
plot(res$midpt, res$simLenFreq, "o")
}
