% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{UpdatePopnDynamics_DynSimMod_AB}
\alias{UpdatePopnDynamics_DynSimMod_AB}
\title{Update population dynamics for age-based dynamic simulation model}
\usage{
UpdatePopnDynamics_DynSimMod_AB(
  nYears,
  MaxModelAge,
  NatMort,
  ReprodPattern,
  InitRatioFem,
  PropFemAtAge,
  FemPropMatAtAge,
  MalPropMatAtAge,
  FemWtAtAge,
  MalWtAtAge,
  FishFemSurvAtAge,
  FishMalSurvAtAge,
  Fish_FemSpBiomPerRec,
  Fish_CombSexSpBiomPerRec,
  BH_SRRa,
  BH_SRRb,
  random_dev,
  FMortByYear,
  DiscMort,
  FemSelLandAtAge,
  MalSelLandAtAge,
  FemSelDiscAtAge,
  MalSelDiscAtAge
)
}
\arguments{
\item{nYears}{number of years of deviations required}

\item{MaxModelAge}{maximum age considered by model}

\item{NatMort}{natural mortality}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{ratio of females to males at birth}

\item{PropFemAtAge}{proportion females at age}

\item{FemPropMatAtAge}{female proportion mature at age}

\item{MalPropMatAtAge}{male proportion mature at age}

\item{FemWtAtAge}{female weight at age}

\item{MalWtAtAge}{male weight at age}

\item{FishFemSurvAtAge}{numbers of females per recruit in each age class}

\item{FishMalSurvAtAge}{numbers of males per recruit in each age class}

\item{Fish_FemSpBiomPerRec}{female spawning biomass per recruit}

\item{Fish_CombSexSpBiomPerRec}{combined sex spawning biomass per recruit}

\item{BH_SRRa}{stock recruitment parameter}

\item{BH_SRRb}{stock recruitment parameter}

\item{random_dev}{annual recruitment deviations}

\item{FMortByYear}{annual values for fully selected fishing mortality}

\item{DiscMort}{level of discard mortality}

\item{FemSelLandAtAge}{selectivity of landings for females}

\item{MalSelLandAtAge}{selectivity of landings for males}

\item{FemSelDiscAtAge}{selectivity of discards for females}

\item{MalSelDiscAtAge}{selectivity of discards for males}
}
\value{
FemSpBiom, MalSpBiom, CombSexSpBiom, AnnRecruit, Catch_Fem, Catch_Mal, Catch_CombSex, Catch_Biom
}
\description{
Generates outputs from a dynamic age-based simulation model (with age-based selectivity) for generation of
annual age composition data
}
\keyword{internal}
