% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{UpdatePopnDynamics_DynSimMod_ALB}
\alias{UpdatePopnDynamics_DynSimMod_ALB}
\title{Update population dynamics for age-based dynamic simulation model allowing for variation in growth}
\usage{
UpdatePopnDynamics_DynSimMod_ALB(
  nYears,
  TimeStep,
  MaxModelAge,
  nLenCl,
  midpt,
  NatMort,
  ReprodPattern,
  InitRatioFem,
  PropFemAtLen,
  RecLenDist,
  FemPropMatAtLen,
  MalPropMatAtLen,
  FemWtAtLen,
  MalWtAtLen,
  LTM_Fem,
  LTM_Mal,
  Fish_FemNPerRecAtAge,
  Fish_MalNPerRecAtAge,
  Fish_FemSpBiomPerRec,
  Fish_CombSexSpBiomPerRec,
  BH_SRRa,
  BH_SRRb,
  random_dev,
  FMortByYear,
  DiscMort,
  FemSelLandAtLen,
  MalSelLandAtLen,
  FemSelDiscAtLen,
  MalSelDiscAtLen
)
}
\arguments{
\item{nYears}{number of years of deviations required}

\item{TimeStep}{model timestep}

\item{MaxModelAge}{maximum age assumed by model}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{NatMort}{natural mortality}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{ratio of females to males at birth}

\item{PropFemAtLen}{proportion females at length}

\item{RecLenDist}{length distribution of recruits}

\item{FemPropMatAtLen}{female proportion mature at length}

\item{MalPropMatAtLen}{male proportion mature at length}

\item{FemWtAtLen}{female weight at length}

\item{MalWtAtLen}{male weight at length}

\item{LTM_Fem}{female length-transition matrix}

\item{LTM_Mal}{male length-transition matrix}

\item{Fish_MalNPerRecAtAge}{numbers of males per recruit in each length and age class}

\item{Fish_FemSpBiomPerRec}{female spawning biomass per recruit}

\item{Fish_CombSexSpBiomPerRec}{combined sex spawning biomass per recruit}

\item{BH_SRRa}{stock recruitment parameter}

\item{BH_SRRb}{stock recruitment parameter}

\item{random_dev}{annual recruitment deviations}

\item{FMortByYear}{annual values for fully selected fishing mortality}

\item{DiscMort}{level of discard mortality}

\item{FemSelLandAtLen}{selectivity of landings for females}

\item{MalSelLandAtLen}{selectivity of landings for males}

\item{FemSelDiscAtLen}{selectivity of discards for females}

\item{MalSelDiscAtLen}{selectivity of discards for males}
}
\value{
FemSpBiom, MalSpBiom, CombSexSpBiom, AnnRecruit, RetCatch_Num, RetCatch_Fem_yl, RetCatch_Mal_yl, RetCatch_CombSex_yl,
RetCatch_Fem_yal, RetCatch_Mal_yal, RetCatch_CombSex_yal, DiscCatch_Num, RetCatchBiom_Fem, RetCatchBiom_Mal, RetCatchBiom_CombSex,
DiscCatch_Fem_yl, DiscCatch_Mal_yl, DiscCatch_CombSex_yl, DiscCatch_Fem_yal, DiscCatch_Mal_yal, DiscCatch_CombSex_yal,
DiscCatchBiom_Fem, DiscCatchBiom_Mal, DiscCatchBiom_CombSex
}
\description{
Generates outputs from a dynamic age-based simulation model, using sex-specific length transition matrices
to allow for growth variation (and length-based selectivity), for generation of length, age, and conditional
age-at-length data
}
\keyword{internal}
