% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{UpdatePopnDynamics_DynSimMod_LB}
\alias{UpdatePopnDynamics_DynSimMod_LB}
\title{Update population dynamics for dynamic simulation model}
\usage{
UpdatePopnDynamics_DynSimMod_LB(
  nYears,
  nLenCl,
  midpt,
  NatMort,
  ReprodPattern,
  InitRatioFem,
  RecLenDist,
  FemPropMatAtLen,
  MalPropMatAtLen,
  FemWtAtLen,
  MalWtAtLen,
  LTM_Fem,
  LTM_Mal,
  Fish_FemNPerRecAtLen,
  Fish_MalNPerRecAtLen,
  Fish_FemSpBiomPerRec,
  Fish_CombSexSpBiomPerRec,
  BH_SRRa,
  BH_SRRb,
  random_dev,
  FMortByYear,
  DiscMort,
  FemSelLandAtLen,
  MalSelLandAtLen,
  FemSelDiscAtLen,
  MalSelDiscAtLen
)
}
\arguments{
\item{nYears}{number of years of deviations required}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{NatMort}{natural mortality}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{ratio of females to males at birth}

\item{RecLenDist}{length distribution of recruits}

\item{FemPropMatAtLen}{female proportion mature at length}

\item{MalPropMatAtLen}{male proportion mature at length}

\item{FemWtAtLen}{female weight at length}

\item{MalWtAtLen}{male weight at length}

\item{Fish_FemNPerRecAtLen}{numbers of females per recruit in each length class}

\item{Fish_MalNPerRecAtLen}{numbers of males per recruit in each length class}

\item{Fish_FemSpBiomPerRec}{female spawning biomass per recruit}

\item{Fish_CombSexSpBiomPerRec}{combined sex spawning biomass per recruit}

\item{BH_SRRa}{stock recruitment parameter}

\item{BH_SRRb}{stock recruitment parameter}

\item{random_dev}{annual recruitment deviations}

\item{FMortByYear}{annual values for fully selected fishing mortality}

\item{DiscMort}{level of discard mortality}

\item{FemSelLandAtLen}{selectivity of landings for females}

\item{MalSelLandAtLen}{selectivity of landings for males}

\item{FemSelDiscAtLen}{selectivity of discards for females}

\item{MalSelDiscAtLen}{selectivity of discards for males}
}
\value{
FemSpBiom, MalSpBiom, CombSexSpBiom, AnnRecruit, Catch_Fem, Catch_Mal, Catch_CombSex, Catch_Biom
}
\description{
Generates outputs from a dynamic simulation model for generation of
annual composition data
}
\keyword{internal}
