% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{VarRecCC_ObjFunc}
\alias{VarRecCC_ObjFunc}
\title{Calculate objective function for multi-year catch curve allowing for variable recruitment}
\usage{
VarRecCC_ObjFunc(params)
}
\arguments{
\item{params}{Estimated parameters (fishing mortality and recruitment deviations)}
}
\value{
NLL
}
\description{
This function calculates the negative multinomial log-likelihood for a  multi-year catch curve
allowing for variable recruitment. As currently implemented, this method assumes knife-edge recruitment.
}
\keyword{internal}
