% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{VisualiseGrowthApplyingLTM}
\alias{VisualiseGrowthApplyingLTM}
\title{Visualise length at age data when applying a length transition matrix}
\usage{
VisualiseGrowthApplyingLTM(
  nFish,
  MaxAge,
  Growth_params,
  CVSizeAtAge,
  midpt,
  nLenCl,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  xmax,
  xint,
  ymax,
  yint
)
}
\arguments{
\item{nFish}{number of growth trajectories for individual fish to simulate}

\item{CVSizeAtAge}{coefficient of variation (CV), common across all mean lengths at age}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{MinAge}{maximum age}
}
\value{
plot of mean growth curve and simulated, individual fish growth trajectories
}
\description{
This function simulates growth trajectories for individual fish for visualising length
at age data produced when applying a length transition matrix, specified
mean sizes at ages and associated variation (common CV for mean lengths at age). Note
that this diagnostic analysis does not account for any fishing mortality (or selectivity)
effects on variation in data.
}
\examples{
set.seed(123)
MaxAge = 20
Ages = 1:MaxAge
MaxLen = 1000
LenInc = 20
lbnd = seq(0,MaxLen - LenInc, LenInc)
ubnd = lbnd + LenInc
midpt = lbnd + (LenInc/2)
nLenCl = length(midpt)
Linf = 800
vbK = 0.2
tzero = 0
Growth_params = c(Linf, vbK, tzero) # currently only set up for von Bertalanffy growth equation
CVSizeAtAge = 0.05
nFish = 100
VisualiseGrowthApplyingLTM(nFish, MaxAge, Growth_params, CVSizeAtAge, midpt, nLenCl, MainLabel=NA,
                           xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA, ymax=NA, yint=NA)
}
