% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{AddAxesAndTickLabelsToPlot}
\alias{AddAxesAndTickLabelsToPlot}
\title{Generic function to add axes and axes labels to plots}
\usage{
AddAxesAndTickLabelsToPlot(
  xmin,
  xmax,
  xint,
  ymin,
  ymax,
  yint,
  cexval,
  cexaxisval,
  lwdval,
  lineval,
  lasval
)
}
\arguments{
\item{xmin}{x axis minimum}

\item{xmax}{x axis maximum}

\item{xint}{x axis tick interval}

\item{ymin}{y axis minimum}

\item{ymax}{y axis maximum}

\item{yint}{y axis tick interval}

\item{cexval}{label size}

\item{cexaxisval}{axis size}

\item{lwdval}{line widths}

\item{lineval}{axis offset}

\item{lasval}{axis orientation}
}
\value{
adds axes to plots
}
\description{
Generic function to add axes and axes labels to plots
}
\keyword{internal}
