% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcPerRecruitFishLenAndFishWtStats}
\alias{CalcPerRecruitFishLenAndFishWtStats}
\title{Calculates expected lengths and associated statistics from length-based per recruit analysis}
\usage{
CalcPerRecruitFishLenAndFishWtStats(
  midpt,
  FemCatchNum,
  MalCatchNum,
  CombSexCatchNum,
  FemWtAtLen,
  MalWtAtLen
)
}
\arguments{
\item{midpt}{length class mid-points}

\item{FemCatchNum}{female per recruit catch numbers at length}

\item{MalCatchNum}{male per recruit catch numbers at length}

\item{CombSexCatchNum}{}
}
\value{
PerRecRandFishLen

PerRecMeanCatchLen

PerRecLenFreq

MeanCatchLen.95ci

MeanCatchLen.60pi

MeanCatchLen.95pi

PerRecRandFishWt

PerRecMeanCatchWt

MeanCatchWt.95ci

MeanCatchWt.60pi

MeanCatchWt.95pi
}
\description{
This function outputs random fish lengths, calculated from length-based per recruit analysis
with specified fishing mortality (PerRecFishLen), expected mean length (PerRecLenFreq),
95 percent confidence limits for mean length (MeanCatchLen.95ci), and 60 and 95 percent
prediction intervals for mean length (MeanCatchLen.60pi, MeanCatchLen.95pi). Values are calculated
for females, males and combined sexes.
}
\keyword{internal}
