% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcSelectivityAndRetentionAtAge}
\alias{CalcSelectivityAndRetentionAtAge}
\title{Calculate gear selectivity and retention at age}
\usage{
CalcSelectivityAndRetentionAtAge(
  EstGearSelAtAge,
  EstRetenAtAge,
  Ages,
  Gear_sel_A50,
  Gear_sel_A95,
  Land_sel_A50,
  Land_sel_A95,
  EstLandSelAtAge,
  ret_A50,
  ret_A95,
  ret_Pmax,
  DiscMort
)
}
\arguments{
\item{EstGearSelAtAge}{input vector for gear selectivity at age (may be set to NA)}

\item{EstRetenAtAge}{input vector for fish retention at age (may be set to NA)}

\item{Ages}{ages for analysis calculated from max age and model timestep}

\item{Gear_sel_A50}{logistic gear selectivity curve parameter}

\item{Gear_sel_A95}{logistic gear selectivity curve parameter}

\item{Land_sel_A50}{logistic parameter for selectivity of landings curve}

\item{Land_sel_A95}{logistic parameter for selectivity of landings curve}

\item{EstLandSelAtAge}{vector for for selectivity of landings at age (set to NA if using age at selectivity of landings parameters)}

\item{ret_A50}{logistic gear retention curve parameter}

\item{ret_A95}{logistic gear retention curve parameter}

\item{ret_Pmax}{logistic gear retention curve parameter}

\item{DiscMort}{proportion of fish that die following capture and release}
}
\value{
FemGearSelAtAge, MalGearSelAtAge, FemRetProbAtAge, MalRetProbAtAge, FemSelLandAtAge, MalSelLandAtAge, FemSelDiscAtAge, MalSelDiscAtAge
}
\description{
Calculate gear selectivity and retention at age, given input parameters or vectors
}
\keyword{internal}
