% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcSelectivityAndRetentionAtLen}
\alias{CalcSelectivityAndRetentionAtLen}
\title{Calculate gear selectivity and retention at length}
\usage{
CalcSelectivityAndRetentionAtLen(
  midpt,
  sel_L50,
  sel_L95,
  ret_L50,
  ret_L95,
  ret_Pmax
)
}
\arguments{
\item{sel_L50}{logistic gear selectivity curve parameter}

\item{sel_L95}{logistic gear selectivity curve parameter}

\item{ret_L50}{logistic gear retention curve parameter}

\item{ret_L95}{logistic gear retention curve parameter}

\item{ret_Pmax}{logistic gear retention curve parameter}

\item{Ages}{ages for analysis calculated from max age and model timestep}
}
\value{
FemGearSelAtAge, MalGearSelAtAge, FemRetProbAtAge, MalRetProbAtAge
}
\description{
Calculate gear selectivity and retention at the mid point for eac length class,
given input parameters
}
\keyword{internal}
