% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{CalcStockRecruitParams_DynSimMod}
\alias{CalcStockRecruitParams_DynSimMod}
\title{Calculate B-H stock recruitment params and unfished spawning biomass}
\usage{
CalcStockRecruitParams_DynSimMod(
  ReprodPattern,
  Unfish_FemSpBiomPerRec,
  Unfish_CombSexSpBiomPerRec,
  InitRec,
  Steepness
)
}
\arguments{
\item{ReprodPattern}{1=gonochoristic, 2=protogynous, 3=protandrous}

\item{Unfish_FemSpBiomPerRec}{unfished spawning biomass per recruit for females}

\item{Unfish_CombSexSpBiomPerRec}{unfished spawning biomass per recruit for combined sexes}

\item{InitRec}{initial recruitment (R0)}

\item{Steepness}{steepness parameters}
}
\value{
Unfish_SpBiom, BH_SRRa, BH_SRRb
}
\description{
Calculate B-H stock recruitment parameters and unfished spawning biomass
for dynamic simulation model
}
\keyword{internal}
