% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Calculate_NLL_LogisticCatchCurve}
\alias{Calculate_NLL_LogisticCatchCurve}
\title{Return negative log-likelihood for a catch curve with age-based, logistic selectivity}
\usage{
Calculate_NLL_LogisticCatchCurve(ln_params)
}
\arguments{
\item{ln_params}{model parameters log(c(FMort, SelA50, SelA95)}
}
\value{
negative log-likelihood (NLL)
}
\description{
This function returns the negative log-likelihood, given age frequency data, a value total mortality,
and age-based selectivity parameters, for a catch curve with age-based, logistic selectivity.
Function requires an estimate of natural mortality (NatMort), a value for maximum age (MaxAge)
and age frequency data (stored in memory in R)
}
\examples{
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 500 # required number of fish for age sample
Res=SimAgeFreqData(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = unlist(as.vector(Res$CatchSample))
Calculate_NLL_LogisticCatchCurve(ln_params)
}
\keyword{internal}
