% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetAgeAndLengthBasedCatchCurveResults}
\alias{GetAgeAndLengthBasedCatchCurveResults}
\title{Get results for a fitted age and length-based catch curve}
\usage{
GetAgeAndLengthBasedCatchCurveResults(
  params,
  MLL,
  SelectivityType,
  ObsCatchFreqAtLen,
  ObsCatchFreqAtLengthAndAge,
  lbnd,
  ubnd,
  midpt,
  SelectivityVec,
  DiscMort,
  MaxAge,
  NatMort
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{ObsCatchFreqAtLen}{observed frequencies in length classes}

\item{ObsCatchFreqAtLengthAndAge}{observed frequencies in length and age classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityVec}{selectivity at length}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{MaxAge}{maximum age considered in model}

\item{NatMort}{natural mortality}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence)
sample size (SampleSize), growth parameter estimates with lower and upper 95 percent
confidence limits (ParamEst), point estimates for estimated parameters (params)
and variance-covariance matrix (vcov.Params), selectivity at length (SelAtLength),
growth curve (MeanSizeAtAge), midpoint of each length class (midpt), mean length after 1 year from
growth curve, given initial length (MeanEndingLength), mean change in length after 1 year,
from initial length - note, assuming normal a distribution allows for possibility of negative growth
if above asyptotic length (TimestepGrowthSizeInc), length distribution of 1+ year old recruits (RecLenDist),
expected relative catches at length and age (Catch), expected catches at length (ExpCatchAtLen),
catch proportions at length (ExpCatchPropInLenClass), expected catches at age (ExpCatchAtAge),
catch proportions at age of females, males and sexes combined (ExpCatchPropAtAge, ExpCatchPropAtAge_Fem,
ExpCatchPropAtAge_Mal), catch proportions at length and age for females, males and sexes combined
(ExpCatchPropLengthGivenAge, ExpCatchPropLengthGivenAge_Fem, ExpCatchPropLengthGivenAge_Mal),
observed catch frequency at length data (ObsCatchFreqAtLen), observed catch frequency at length and age
(ObsCatchFreqAtLengthAndAge)
}
\description{
This function fits a length-based catch curve with length-based selectivity, based on
either 1) specified length-based selectivity inputted as a vector or 2) estimated asymptotic logistic
selectivity curve, and estimated von Bertalalnffy growth curve. The model is fitted to a
sample of fish length and age data, by minimising the overall negative log-likelihood,
including the NLL associated with the marginal length composition and a conditional age at length NLL,
given the parameters (selectivity, growth and mortality) and data, using nlminb.
It provides various statistical outputs in include convergence statistics, parameter estimates
and associated 95 percent confidence limits and associated variance-covariance matrix, calculated using
the MASS package.
}
\examples{
set.seed(123)
# Simulate data
SampleSize=5000
set.seed(123)
MaxAge = 26
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1100
LenInc = 50
MLL = NA
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
L50 = 400 # selectivity
L95 = 450 # selectivity
SelectivityVec = NA
DiscMort = 0
# # single sex, von Bertalanffy
# GrowthCurveType = 1 # 1 = von Bertalanffy
# Linf = 800
# vbK = 0.2
# CVSizeAtAge = 0.08
# GrowthParams = c(Linf, vbK, CVSizeAtAge)
# RefnceAges = NA
# 2 sexes, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = c(700,850)
vbK = c(0.25,0.2)
CVSizeAtAge = c(0.05,0.05)
RefnceAges = NA
GrowthParams = data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge)
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         L50, L95, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# # get data - 1 sex (or combined sexes)
# ObsCatchFreqAtLen = Res$ObsCatchFreqAtLen # 1 sex
# ObsCatchFreqAtLengthAndAge = as.matrix(Res$ObsCatchFreqAtLengthAndAge) # 1 sex
# get data - 2 sexes
ObsCatchFreqAtLen <- data.frame(matrix(nrow = 2, ncol = length(midpt))) # 2 sex
colnames(ObsCatchFreqAtLen) <- midpt
ObsCatchFreqAtLen[1,] = Res$ObsCatchFreqAtLen_Fem
ObsCatchFreqAtLen[2,] = Res$ObsCatchFreqAtLen_Mal
ObsCatchFreqAtLengthAndAge = array(c(unlist(Res$ObsCatchFreqAtLengthAndAge_Fem), unlist(Res$ObsCatchFreqAtLengthAndAge_Mal)),
                                   c(MaxAge, length(midpt), 2), dimnames=list(rownames(Res$ObsCatchFreqAtLengthAndAge_Fem),
                                                                              colnames(Res$ObsCatchFreqAtLengthAndAge_Fem)))
# # get params - 1 sex
# InitFishMort = 0.3 # specify starting parameters
# InitL50 = 300
# InitL95 = 500
# InitLinf = 800
# InitvbK = 0.2
# InitCVSizeAtAge = 0.05
# get params - 2 sexes
InitFishMort = 0.3 # specify starting parameters
InitL50 = 300
InitL95 = 500
InitLinf = c(800,800)
InitvbK = c(0.25,0.25)
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform (so F is always between 0 and 1)
params = c(InitFishMort_logit, log(c(InitL50, InitL95, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, MLL, SelectivityType, ObsCatchFreqAtLen, ObsCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityVec, DiscMort, MaxAge, NatMort)
}
