% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetAgeAndLengthBasedCatchCurveResults}
\alias{GetAgeAndLengthBasedCatchCurveResults}
\title{Get results for a fitted age and length-based catch curve}
\usage{
GetAgeAndLengthBasedCatchCurveResults(
  params,
  RefnceAges,
  MLL,
  GrowthCurveType,
  SelectivityType,
  ObsRetCatchFreqAtLen,
  ObsRetCatchFreqAtLengthAndAge,
  lbnd,
  ubnd,
  midpt,
  SelectivityVec,
  DiscMort,
  MaxAge,
  NatMort,
  TimeStep
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated}

\item{RefnceAges}{Reference ages for Schnute growth curve (set to NA for von Bertalanffy growth curve)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{ObsRetCatchFreqAtLengthAndAge}{observed frequencies in length and age classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityVec}{selectivity at length}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{MaxAge}{maximum age considered in model}

\item{NatMort}{natural mortality}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence)
sample size (SampleSize), growth parameter estimates with lower and upper 95 percent
confidence limits (ParamEst), point estimates for estimated parameters (params)
and variance-covariance matrix (vcov.Params), gear selectivity at length (SelAtLength), retention
at length (RetAtLength), selectivity of landings at length (SelLandAtLength),
growth curve (MeanSizeAtAge), midpoint of each length class (midpt), mean length after 1 year from
growth curve, given initial length (MeanEndingLength), mean change in length after 1 year,
from initial length - note, assuming normal a distribution allows for possibility of negative growth
if above asyptotic length (TimestepGrowthSizeInc), length distribution of 1+ year old recruits (RecLenDist),
expected retained catches at length and integer age class (RetCatchAtIntAgeLen), expected catches at length (ExpCatchAtLen),
catch proportions at length (ExpCatchPropInLenClass), expected retained catches at integer age classes (RetCatchAtIntAge),
catch proportions at integer age class of females, males and sexes combined (ExpRetCatchPropAtIntAge, ExpRetCatchPropAtIntAge_Fem,
ExpRetCatchPropAtIntAge_Mal), catch proportions at length and integer age class for females, males and sexes combined
(ExpRetCatchPropLengthGivenIntAge, ExpRetCatchPropLengthGivenIntAge_Fem, ExpRetCatchPropLengthGivenIntAge_Mal),
observed catch frequency at length data (ObsRetCatchFreqAtLen), observed catch frequency at length and age
(ObsRetCatchFreqAtLengthAndAge), approximate CV at length corresponding to maximum age (CV_LenAtMaxAge, FemCV_LenAtMaxAge,
MalCV_LenAtMaxAge)
}
\description{
This function fits a length-based catch curve with length-based selectivity, based on
either 1) specified length-based selectivity inputted as a vector or 2) estimated asymptotic logistic
selectivity curve, and estimated von Bertalalnffy growth curve. The model is fitted to a
sample of fish length and age data, by minimising the overall negative log-likelihood,
including the NLL associated with the marginal length composition and a conditional age at length NLL,
given the parameters (selectivity, growth and mortality) and data, using nlminb.
It provides several statistical outputs including convergence statistics, parameter estimates
and associated 95 percent confidence limits and associated variance-covariance matrix, calculated using
the MASS package.
}
\examples{
# Simulate data
set.seed(123)
SampleSize=5000
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityVec = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
RefnceAges = NA
GrowthParams = c(Linf, vbK, CVSizeAtAge)
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# 2 sexes, von Bertalanffy
# GrowthCurveType = 1 # 1 = von Bertalanffy
# Linf = c(700,850)
# vbK = c(0.25,0.2)
# CVSizeAtAge = c(0.05,0.05)
# RefnceAges = NA
# GrowthParams = data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge)
# Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
#                          SelParams, RetenParams, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# # get data - 2 sexes
# ObsRetCatchFreqAtLen <- data.frame(matrix(nrow = 2, ncol = length(midpt))) # 2 sex
# colnames(ObsRetCatchFreqAtLen) <- midpt
# ObsRetCatchFreqAtLen[1,] = Res$ObsRetCatchFreqAtLen_Fem
# ObsRetCatchFreqAtLen[2,] = Res$ObsRetCatchFreqAtLen_Mal
# ObsRetCatchFreqAtLengthAndAge = array(c(unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem), unlist(Res$ObsRetCatchFreqAtLengthAndDecAge_Mal)),
#                                       c(nTimeSteps, length(midpt), 2), dimnames=list(rownames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem),
#                                                                                      colnames(Res$ObsRetCatchFreqAtLengthAndDecAge_Fem)))
# get params - 1 sex
InitFishMort = 0.3 # specify starting parameters
InitL50 = 320
InitDelta = 50 # L95-L50
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform (so F is always between 0 and 1)
params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
# # get params - 2 sexes
# InitFishMort = 0.3 # specify starting parameters
# InitL50 = 320
# InitDelta = 50 # L95-L50
# InitLinf = c(800,800)
# InitvbK = c(0.25,0.25)
# InitCVSizeAtAge = 0.05
# InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform (so F is always between 0 and 1)
# params = c(InitFishMort_logit, log(c(InitL50, InitDelta, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityVec, DiscMort, MaxAge, NatMort, TimeStep)
# Example with specified selectivity vector
# Simulate data
SelectivityType=1 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityVec = 1 / (1 + exp(-log(19)*(midpt-400)/(500-400)))
SelParams = c(NA, NA) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.05
GrowthParams = c(Linf, vbK, CVSizeAtAge)
RefnceAges = NA
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# get data - 1 sex (or combined sexes)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen # 1 sex
ObsRetCatchFreqAtLengthAndAge = as.matrix(Res$ObsRetCatchFreqAtLengthAndDecAge) # 1 sex
# get params - 1 sex
InitFishMort = 0.3 # specify starting parameters
InitLinf = 800
InitvbK = 0.2
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform (so F is always between 0 and 1)
params = c(InitFishMort_logit, log(c(InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, RefnceAges, MLL, GrowthCurveType, SelectivityType, ObsRetCatchFreqAtLen, ObsRetCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityVec, DiscMort, MaxAge, NatMort, TimeStep)
}
