% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetGrowthInputsForLengthTransitionMatrices}
\alias{GetGrowthInputsForLengthTransitionMatrices}
\title{Get inputs for length transition matrices}
\usage{
GetGrowthInputsForLengthTransitionMatrices(
  MaxAge,
  TimeStep,
  nLenCl,
  midpt,
  GrowthCurveType,
  GrowthModelType,
  GrowthParams,
  RefnceAges
)
}
\arguments{
\item{MaxAge}{maximum age considered in model}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{nLenCl}{number of length classes}

\item{midpt}{mid points of length classes}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute, 3=any model with annual size increment input}

\item{GrowthModelType}{1=von Bertalanffy - 1 sex, 2=von Bertalanffy - 2 sexes,
3=Schnute - 1 sex, 4=Schnute - 2 sexes, 5=size increment input - 1 sex, 6=size increment input - 1 sexes}

\item{GrowthParams}{growth parameters or list with tag-model growth outputs}

\item{RefnceAges}{Schnute reference ages}
}
\value{
mean size at age from growth curve (MeanSizeAtAge), estimated length after one year for each length class
mid point, given growth curve (MeanEndingLength) and associated change in length after one year of growth (TimestepGrowthSizeInc)
}
\description{
Get key inputs required to calculate length transition matrices, for models with either one or two sexes,
and using von Bertalanffy or Schnute growth model
}
\keyword{internal}
