% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetGrowthModelType}
\alias{GetGrowthModelType}
\title{Allocate number for alternative length based catch curve models}
\usage{
GetGrowthModelType(CatchCurveType, GrowthCurveType, GrowthParams)
}
\arguments{
\item{CatchCurveType}{1=length based, 2=age and length based}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{(von Bertalanffy or Schnute growth parameters), GrowthModelType
1=von Bertalanffy 1 sex, 2= von Bertalanffy 2 sexes, 3=Schnute 1 sex, 4=Schnute 2 sexes, and
CVSizeAtAge for 1 or 2 sex model}
}
\value{
GrowthModelType 1=length-based von Bertalanffy, 2=length-based Schnute,
3=age and length based von Bertalanffy, 4=age and length based Schnute
}
\description{
Allocate number for alternative single and separate sex length based catch curve models,
and growth curve types (1=von Beralanffy, 2=Schnute)
}
\keyword{internal}
