% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetInputsForPlotting_Cond_AL}
\alias{GetInputsForPlotting_Cond_AL}
\title{Output observed and expected quantities associated with conditional age-length plots}
\usage{
GetInputsForPlotting_Cond_AL(
  params,
  RefnceAges,
  MLL,
  GrowthCurveType,
  SelectivityType,
  ObsRetCatchFreqAtLen,
  ObsRetCatchFreqAtLengthAndAge,
  lbnd,
  ubnd,
  midpt,
  SelectivityVec,
  DiscMort,
  MaxAge,
  NatMort,
  TimeStep,
  FittedRes
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated (if FittedRes=NA)}

\item{RefnceAges}{Reference ages for Schnute growth curve (set to NA for von Bertalanffy growth curve)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{ObsRetCatchFreqAtLengthAndAge}{observed frequencies in length and age classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityVec}{selectivity at length}

\item{DiscMort}{Proportion of fish that die following to capture and release}

\item{MaxAge}{maximum age considered in model}

\item{NatMort}{natural mortality}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{FittedRes}{saved results from GetLengthBasedCatchCurveResults function (model will be refitted if set to NA)}
}
\value{
ObsRetCatchFreqAtLengthAndIntAge, ObsCatchPropAgeAtLength, ExpRetCatchPropIntAgeGivenLength (single sex),
ObsRetCatchFreqAtLengthAndIntAge_Fem, ObsRetCatchFreqAtLengthAndIntAge_Mal, ObsCatchPropAgeAtLength_Fem,
ObsCatchPropAgeAtLength_Mal, ExpRetCatchPropIntAgeGivenLength_Fem, ExpRetCatchPropIntAgeGivenLength_Mal (separaate sex)
}
\description{
This function outputs several observed and expected quantities associated with conditional age-length plots,
including observed vs expected age distribution at length plots, and pearson residual plots
}
