% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetLengthBasedCatchCurveResults}
\alias{GetLengthBasedCatchCurveResults}
\title{Get results for a fitted length-based catch curve}
\usage{
GetLengthBasedCatchCurveResults(
  params,
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  MLL,
  SelectivityType,
  ObsCatchFreqAtLen,
  lbnd,
  ubnd,
  midpt,
  SelectivityVec,
  PropReleased,
  DiscMort,
  CVSizeAtAge,
  MaxAge,
  NatMort
)
}
\arguments{
\item{params}{vector of model parameters in log space (params) to be estimated}

\item{GrowthCurveType}{1 = von Bertalanffy, 2 = Schnute}

\item{GrowthParams}{c(Linf, vbK, CVSizeAtAge) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge),
both sexes von Bertalanffy, or c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b), both sexes Schnute}

\item{RefnceAges}{reference ages for Schnute function (set to NA if growth based on another function)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{ObsCatchFreqAtLen}{observed frequency of fish at length in sample}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityVec}{selectivity at length}

\item{PropReleased}{proportion of fish that are released}

\item{DiscMort}{Proportion of fish that die following to capture and release}

\item{CVSizeAtAge}{coefficient of variation (CV), common across all mean lengths at age}

\item{MaxAge}{maximum age considered in model}

\item{NatMort}{natural mortality}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence)
sample size (SampleSize), growth parameter estimates with lower and upper 95 percent
confidence limits (ParamEst), point estimates for estimated parameters (params)
and variance-covariance matrix (vcov.Params), selectivity at length (SelAtLength),
growth curve (MeanSizeAtAge), midpoint of each length class (midpt), mean length after 1 year from
growth curve, given initial length (MeanEndingLength), mean change in length after 1 year,
from initial length - note, assuming normal a distribution allows for possibility of negative growth
if above asymptotic length (AnnGrowthSizeInc), length distribution of 1+ year old recruits (RecLenDist),
expected catches, at length (ExpCatchAtLen), proportion of catch in each length classes for sexes combined
and separate (ExpCatchPropInLenClass, ExpCatchPropInLenClass_Fem, ExpCatchPropInLenClass_Mal),
observed catch in each length class for sexes combined (ObsCatchFreqAtLen), CV for modelled lengthsat age, around mean
length at MaxAge, for sexes combined or separate (CV_LenAtMaxAge, FemCV_LenAtMaxAge, MalCV_LenAtMaxAge)
}
\description{
This function fits a length-based catch curve with length-based selectivity, based on
either 1) specified length-based selectivity inputted as a vector, or 2) an estimated asymptotic logistic
selectivity curve, and specified growth curve (either von Bertalanffy or Schnute growth model). The model is
to a sample of fish length frequency data, by minimising the negative log-likelihood associated
with the parameters and data, using nlminb. It provides various statistical outputs in include
convergence statistics, parameter estimates and associated 95 percent confidence limits and associated
variance-covariance matrix, calculated using the MASS package.
}
\examples{
# Simulate data
SampleSize=5000
set.seed(123)
MaxAge = 30
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1500
LenInc = 20
MLL = NA
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
L50 = 600 # selectivity
L95 = 700 # selectivity
SelectivityVec = NA
PropReleased = NA # vector including mean and sd
DiscMort = 0 # proportion of fish that die due to natural mortality
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.08
GrowthParams = c(Linf, vbK, CVSizeAtAge)
RefnceAges = NA
# # 2 sexes, von Bertalanffy
# GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
# Linf = c(700,850)
# vbK = c(0.3,0.2)
# CVSizeAtAge = c(0.08,0.08)
# GrowthParams = data.frame(Linf=Linf, vbK=vbK, CVSizeAtAge=CVSizeAtAge)
# RefnceAges = NA
# 1 sex, Schnute
# GrowthCurveType = 2 # 1 = von Bertalanffy, 2 = Schnute
# t1 = 1 # growth - Schnute
# t2 = 10 # growth - Schnute
# y1 = 400 # growth - Schnute
# y2 = 1000 # growth - Schnute
# a = 0.1 # growth - Schnute
# b = 2.0 # growth - Schnute
# GrowthParams = c(y1, y2, a, b)
# RefnceAges = c(t1,t2)
# CVSizeAtAge = 0.08
# # 2 sexes, Schnute
# GrowthCurveType = 2 # 1 = von Bertalanffy, 2 = Schnute
# t1 = c(1,1) # growth - Schnute
# t2 = c(10,10) # growth - Schnute
# y1 = c(435.3,400.3) # growth - Schnute
# y2 = c(1089.1,1034.5) # growth - Schnute
# a = c(0.044,0.136) # growth - Schnute
# b = c(2.748,1.971) # growth - Schnute
# CVSizeAtAge = c(0.08, 0.08)
# GrowthParams = data.frame(y1=y1, y2=y2, a=a, b=b)
# RefnceAges = data.frame(t1=t1,t2=t2)
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         L50, L95, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
ObsCatchFreqAtLen = Res$ObsCatchFreqAtLen
midpt=Res$midpt
lbnd=Res$lbnd
ubnd=Res$ubnd
InitFishMort = 0.25 # specify starting parameters
InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform
InitL50 = 600
InitL95 = 700
params = c(InitFishMort_logit, log(InitL50), log(InitL95))
Res=GetLengthBasedCatchCurveResults(params, GrowthCurveType, GrowthParams, RefnceAges, MLL, SelectivityType, ObsCatchFreqAtLen,
                                          lbnd, ubnd, midpt, SelectivityVec, PropReleased, DiscMort, CVSizeAtAge, MaxAge, NatMort)
}
