% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetLogisticCatchCurveResults}
\alias{GetLogisticCatchCurveResults}
\title{Get statistical outputs from a fitted catch curve with age-based, logistic selectivity}
\usage{
GetLogisticCatchCurveResults(ln_params, NatMort, Ages, ObsAgeFreq)
}
\arguments{
\item{ln_params}{initial values of model parameters log(c(FMort, SelA50, SelA95)}

\item{NatMort}{natural mortality}

\item{Ages}{ages in observed data}

\item{ObsAgeFreq}{observed age frequency data}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence)
sample size (SampleSize), parameter estimates with lower and upper 95 percent
confidence limits (ParamEst), point estimates for parameters estimated in log space (Estlnparams),
variance-covariance matrix for estimated parameters (vcov.Params), standard errors for estimated
parameters (lnEstFMort_se, lnEstSelA50_se, lnEstFSelA95_se), selectivity at age (SelAtAge), fishing
mortality at age (FAtAge), estimated frequencies at age with associated 95 percent confidence limits
(EstFreq, EstFreq_Zlow, EstFreq_Zup), random values of parameters in log space from parametric resampling,
using rmultinom function (lnparams.sims), and associated median and lower 2.5 and upper 97.5 percentiles
of estimates for frequency at age (EstFreq.sim), selectivity parameters in normal space (SelA50.sim, SelA95.sim),
fishing mortality (FMort.sim) and total mortality (EstZMort.sim)
}
\description{
This function fits a catch curve with an asymptotic, age-based logistic selectivity curve,
to a sample of fish age frequency data, by minimising the negative log-likelihood associated
with the parameters and data, using nlminb. It provides various statistical outputs including
convergence statistics, parameter estimates and associated 95 percent confidence limits and associated
variance-covariance matrix, calculated using the MASS package
}
\examples{
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 1000 # required number of fish for age sample
Res=SimAgeFreqData(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = unlist(as.vector(Res$CatchSample))
Init_FMort = 0.2
Init_SelA50 = 5
Init_SelA95 = 7
ln_params = log(c(Init_FMort, Init_SelA50, Init_SelA95))
res=GetLogisticCatchCurveResults(ln_params, NatMort, Ages, ObsAgeFreq)
}
