% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetMeanAgeStats_DynSimMod}
\alias{GetMeanAgeStats_DynSimMod}
\title{Calculate mean age statistics for dynamic simulation model}
\usage{
GetMeanAgeStats_DynSimMod(
  nYears,
  MaxModelAge,
  RandObsCatchAgeFreq_Fem,
  RandObsCatchAgeFreq_Mal
)
}
\arguments{
\item{nYears}{number of years of length data}

\item{MaxModelAge}{maximum age considered by model}

\item{RandObsCatchLenFreq_Fem}{randomly generated annual length frequencies for females}

\item{RandObsCatchLenFreq_Mal}{randomly generated annual length frequencies for males}
}
\value{
MeanAgeStats
}
\description{
Calculates mean ages and associated 95 percent confidence limits for
each year of generated age data from dynamic simulation model
}
\keyword{internal}
