% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetParamEst_LogisticCatchCurve}
\alias{GetParamEst_LogisticCatchCurve}
\title{Get parameter estimates and associated 95% CLs, in normal space, for age-based catch curve with
logistic selectivity}
\usage{
GetParamEst_LogisticCatchCurve(params, ses)
}
\arguments{
\item{params}{model parameters log(c(FMort, SelA50, SelDelta) for multinomial NLL or
log(c(FMort, SelA50, SelDelta, theta) for Dirichlet multinomial NLL}

\item{ses}{standard errors calculated for each of the estimated model parameters}
}
\value{
All estimated parameters and 95 percent CLs (ParamEst), estimates of the following
individual parameters with 96 percent CLs (EstFMort, EstSelA50, EstSelA95, EstSelA50_1,
EstSelA95_1, EstSelA50_2, EstSelA95_2, EstTheta)
}
\description{
This function returns estimated parameters, in normal space, for the age-based catch curve with
logistic selectivity (which are then outputted from the function GetLogisticCatchCurveResults).
}
\keyword{internal}
