% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetPerRecruitResults_AB}
\alias{GetPerRecruitResults_AB}
\title{Get outputs from age-based per recruit analysis across a range of fishing mortality values}
\usage{
GetPerRecruitResults_AB(
  MaxModelAge,
  TimeStep,
  Linf,
  vbK,
  tzero,
  EstLenAtAge,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtAge,
  ReprodScale,
  ReprodPattern,
  InitRatioFem,
  FinalSex_Pmax,
  FinalSex_A50,
  FinalSex_A95,
  mat_A50,
  mat_A95,
  EstMatAtAge,
  Gear_sel_A50,
  Gear_sel_A95,
  EstGearSelAtAge,
  Land_sel_A50,
  Land_sel_A95,
  EstLandSelAtAge,
  ret_Pmax,
  ret_A50,
  ret_A95,
  EstRetenAtAge,
  DiscMort,
  Steepness,
  SRrel_Type,
  NatMort,
  Current_F
)
}
\arguments{
\item{MaxModelAge}{maximum age considered in model}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{Linf}{asymptotic length (von Bertalanffy growth parameter, set to NA if inputting lengths at ages directly from any growth model)}

\item{vbK}{growth coefficient (von Bertalanffy growth parameter, set to NA if inputting lengths at ages directly from any growth model)}

\item{tzero}{hypothetical age at zero length (von Bertalanffy growth parameter, set to NA if inputting lengths at ages directly from any growth model)}

\item{EstLenAtAge}{vector or estimated lengths at ages from any growth model, set to NA if von Bertalanffy growth parameters specified}

\item{lenwt_a}{weight-length parameter (power relationship)}

\item{ln_lenwt_a}{weight-length parameter (log-log relationship)}

\item{lenwt_b}{weight-length parameter (power or log-log relationship)}

\item{WLrel_Type}{1=power, 2=log-log relationship (set to NA if inputting weights at ages directly)}

\item{EstWtAtAge}{vector of weights at ages (set to NA if weight-length growth parameters specified)}

\item{ReprodScale}{<- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{proportion of fish that are females at hatching}

\item{FinalSex_Pmax}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_A50}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_A95}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{mat_A50}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{mat_A95}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{EstMatAtAge}{vector of proportion mature at age (set to NA if using age at maturity parameters)}

\item{Gear_sel_A50}{logistic parameter for gear selectivity curve}

\item{Gear_sel_A95}{logistic parameter for gear selectivity curve}

\item{EstGearSelAtAge}{vector for gear selectivity at age (set to NA if using age at gear selectivity parameters)}

\item{Land_sel_A50}{logistic parameter for selectivity of landings curve}

\item{Land_sel_A95}{logistic parameter for selectivity of landings curve}

\item{EstLandSelAtAge}{vector for for selectivity of landings at age (set to NA if using age at selectivity of landings parameters)}

\item{ret_Pmax}{logistic parameter for fish retention curve}

\item{ret_A50}{logistic parameter for fish retention curve}

\item{ret_A95}{logistic parameter for fish retention curve}

\item{EstRetenAtAge}{vector for fish retention at age (set to NA if using age at retention parameters)}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{SRrel_Type}{1 = Beverton-Holt, 2=Ricker stock-recruitment relationship}

\item{NatMort}{natural mortality}

\item{Current_F}{estimate of current fishing mortality}
}
\value{
Age at each timestep (Ages), female and male length at age (FemLenAtAge, MalLenAtAge),
female and male weight at age (FemWtAtAge, FemWtAtAge), proportion female at age (PropFemAtAge), female and male
proportion mature at age (FemPropMatAtAge, MalPropMatAtAge), female and male selectivity at age (FemSelAtAge, MalSelAtAge),
yield-per-recruit (YPR), unfished female, male and combined sex spawning potential ratio (Fem_SPR, Mal_SPR,
CombSex_SPR), unfished female and male per recruit survival at age (UnfishFemSurvAtAge, UnfishMalSurvAtAge),
unfished female and male mature biomass at age (UnfishFemSpBiomAtAge, UnfishMalSpBiomAtAge), fished female and
male spawning potential ratio (Fem_SPR, Mal_SPR), fished female and male per recruit survival at age
(FishFemSurvAtAge, FishMalSurvAtAge), fished female and male mature biomass at age (FishFemSpBiomAtAge,
FishMalSpBiomAtAge), female and male per recruit retained catches at age in numbers (FemCatchAtAgeNum, MalCatchAtAgeNum),
female and male per recruit released catches at age in numbers, (FemRelCatchAtAgeNum, MalRelCatchAtAgeNum),
female and male per recruit retained catches at age in biomass (FemCatchAtAge, MalCatchAtAge), and for the extended model, equilibrium recruitment (Eq_Rec),
equilibrium catch (Eq_Catch), equilibrium female and male spawning biomass (Eq_FemSpBiom, Eq_MalSpBiom) and
relative female, male and combined sex spawning biomass (Eq_FemRelSpBiom, Eq_MalRelSpBiom, Eq_CombSexRelSpBiom),
female and male gear selectivity at age (FemGearSelAtAge, MalGearSelAtAge), female and male retention at age probabilities,
(FemRetProbAtAge, MalRetProbAtAge), selectivity of female and male fish landings (FemSelLandAtAge, MalSelLandAtAge),
female and male selectivity of discards (FemSelDiscAtAge, MalSelDiscAtAge), female and male fishing mortality
associated with discarding (FemDiscFAtAge, MalDiscFAtAge), female and male fishing mortality associated with
fish landings (FemLandFAtAge, MalLandFAtAge), female and male total fishing mortality at age (FemFAtAge, MalFAtAge),
female and male total mortality at age (FemZAtAge, MalZAtAge)range of fishing mortality values applied for which per quantities are calculated
(FishMort = seq(0,2,0.01)), equilibrium recruitment vs FMort (Eq_Rec), equilibrium catch vs FMort (Eq_Catch),
equilibrium female spawning biomass vs FMort (Eq_FemSpBiom), equilibrium spawning potential ratio vs FMort
(Eq_SPR), maximum yield per recruit (YPRmax), maximum equilibrium catch (maxeqCatch), fishing mortality
associated with YPRmax (Fmax), fishing mortality associated with maxeqCatch (F_MSY or FmaxeqCatch), biomass target at 1.2maxeqCatch,
(BMSY_Targ), biomass threshold at maxeqCatch (BMSY_Thresh), biomass threshold at 0.5maxeqCatch (BMSY_Lim),
YPR vs FishMort (YPRResults), Eq_Catch vs FishMort (Eq_CatchResults), Eq_FemRelSpBiom vs FishMort (Fem_SPRResults),
Eq_MalRelSpBiom vs FishMort (Mal_SPRResults), Eq_FemRelSpBiom vs FishMort (Eq_FemRelSpBiomResults),
Eq_MalRelSpBiom vs FishMort (Eq_MalRelSpBiomResults), Eq_CombSexRelSpBiom vs FishMort (CombSex_SPRResults),
Eq_Rec vs FishMort (Eq_RecResults)
}
\description{
This function provides outputs associated with per recruit analysis, and an
extended form of this analysis with a Beverton-Holt stock recruitment relationship to account
for potential impacts of fishing on recruitment. Outputs are provided for a range of
fishing mortality values, including the current, estimated value.
}
\examples{
# Example 1. Non-hermaphroditic species
InitRecruit <- 1 # Initial recruitment
MaxModelAge <- 20 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be appropriate to use a smaller time step)
Linf <- c(550, 500) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.5, 0.5) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
EstLenAtAge <- data.frame(EstFemLenAtAge=NA, EstMalLenAtAge=NA) # length at age (from age 0), inputted as values in data frame
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtAge <- data.frame(EstFemWtAtAge=NA, EstMalWtAtAge=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at age zero
FinalSex_Pmax <- NA # Logistic sex change relationship parameters (max probability of final sex)
FinalSex_A50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_A95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
mat_A50 <- c(2.5, 2.5) # females, males - Logistic length (mm) at maturity relationship parameters
mat_A95 <- c(3.5, 3.5) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtAge <- data.frame(EstFemMatAtAge=NA, EstMalMatAtAge=NA) # maturity at age (from age 0), inputted as values in data frame
Gear_sel_A50 <- NA # females, males - Logistic age selectivity relationship parameters
Gear_sel_A95 <- NA # females, males - Logistic age selectivity relationship parameters
EstGearSelAtAge <- data.frame(EstFemSelAtAge=NA, EstMalSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
Land_sel_A50 <- c(2.5, 2.5) # females, males - Logistic age selectivity relationship parameters
Land_sel_A95 <- c(3.5, 3.5) # females, males - Logistic age selectivity relationship parameters
EstLandSelAtAge <- data.frame(EstFemSelAtAge=NA, EstMalSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
ret_Pmax <- NA  # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_A50 <- NA  # females, males - Logistic age fish retention at age parameters
ret_A95 <- NA  # females, males - Logistic age fish retention at age parameters
EstRetenAtAge <- data.frame(EstFemRetenAtAge=rep(1,MaxModelAge+1), EstMalRetenAtAge=rep(1,MaxModelAge+1)) # fish retention at age (from age 0), inputted as values in data frame
DiscMort <- 0.0 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort = 0.2 # natural mortality  (year-1)
Current_F <- 0.07 # estimate of fishing mortality, e.g. from catch curve analysis
Res = GetPerRecruitResults_AB(MaxModelAge, TimeStep, Linf, vbK, tzero, EstLenAtAge,
                           lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtAge, ReprodScale,
                           ReprodPattern, InitRatioFem, FinalSex_Pmax, FinalSex_A50, FinalSex_A95,
                           mat_A50, mat_A95, EstMatAtAge, Gear_sel_A50, Gear_sel_A95, EstGearSelAtAge,
                           Land_sel_A50, Land_sel_A95, EstLandSelAtAge, ret_Pmax, ret_A50, ret_A95,
                           EstRetenAtAge, DiscMort, Steepness, SRrel_Type, NatMort, Current_F)
# Example 2: hermaphroditic species
InitRecruit <- 1 # Initial recruitment
MaxModelAge <- 100 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be appropriate to use a smaller time step)
Linf <- c(1000, 1000) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.1, 0.1) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
EstLenAtAge <- data.frame(EstFemLenAtAge=NA, EstMalLenAtAge=NA) # length at age (from age 0), inputted as values in data frame
lenwt_a <- NA # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- -11.0 # for log-log relationship
lenwt_b <- 3.0 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 2 # 1=power, 2=log-log relationship
EstWtAtAge <- data.frame(EstFemWtAtAge=NA, EstMalWtAtAge=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 2 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 1 # Ratio of females to males at age zero
FinalSex_Pmax <- 1 # Logistic sex change relationship parameters (max probability of final sex)
FinalSex_A50 <- 35 # Logistic sex change relationship parameters (inflection point)
FinalSex_A95 <- 60 # Logistic sex change relationship parameters (95\% of max probability)
mat_A50 <- c(20, 20) # females, males - Logistic length (mm) at maturity relationship parameters
mat_A95 <- c(30, 30) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtAge <- data.frame(EstFemMatAtAge=NA, EstMalMatAtAge=NA) # maturity at age (from age 0), inputted as values in data frame
Gear_sel_A50 <- NA # females, males - Logistic age selectivity relationship parameters
Gear_sel_A95 <- NA # females, males - Logistic age selectivity relationship parameters
EstGearSelAtAge <- data.frame(EstFemSelAtAge=NA, EstMalSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
Land_sel_A50 <- c(2.5, 2.5) # females, males - Logistic age selectivity relationship parameters
Land_sel_A95 <- c(3.5, 3.5) # females, males - Logistic age selectivity relationship parameters
EstLandSelAtAge <- data.frame(EstFemSelAtAge=NA, EstMalSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
ret_Pmax <- NA  # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_A50 <- NA  # females, males - Logistic age fish retention at age parameters
ret_A95 <- NA  # females, males - Logistic age fish retention at age parameters
EstRetenAtAge <- data.frame(EstFemRetenAtAge=rep(1,MaxModelAge+1), EstMalRetenAtAge=rep(1,MaxModelAge+1)) # fish retention at age (from age 0), inputted as values in data frame
DiscMort <- 0.0 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort = 0.07 # natural mortality  (year-1)
Current_F <- 0.07 # estimate of fishing mortality, e.g. from catch curve analysis
Res = GetPerRecruitResults_AB(MaxModelAge, TimeStep, Linf, vbK, tzero, EstLenAtAge,
                           lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtAge, ReprodScale,
                           ReprodPattern, InitRatioFem, FinalSex_Pmax, FinalSex_A50, FinalSex_A95,
                           mat_A50, mat_A95, EstMatAtAge, Gear_sel_A50, Gear_sel_A95, EstGearSelAtAge,
                           Land_sel_A50, Land_sel_A95, EstLandSelAtAge, ret_Pmax, ret_A50, ret_A95,
                           EstRetenAtAge, DiscMort, Steepness, SRrel_Type, NatMort, Current_F)
}
