% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetPerRecruitResults_AB_with_err}
\alias{GetPerRecruitResults_AB_with_err}
\title{Get outputs from per age-based recruit analysis across a range of fishing mortality values, allowing for error in
M, h and F}
\usage{
GetPerRecruitResults_AB_with_err(
  MaxModelAge,
  TimeStep,
  Linf,
  vbK,
  tzero,
  EstLenAtAge,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtAge,
  ReprodScale,
  ReprodPattern,
  InitRatioFem,
  FinalSex_Pmax,
  FinalSex_A50,
  FinalSex_A95,
  mat_A50,
  mat_A95,
  EstMatAtAge,
  sel_A50,
  sel_A95,
  EstSelAtAge,
  ret_Pmax,
  ret_A50,
  ret_A95,
  EstRetenAtAge,
  DiscMort,
  Steepness,
  Steepness_sd,
  SRrel_Type,
  NatMort,
  NatMort_sd,
  Current_F,
  Current_F_sd,
  nReps
)
}
\arguments{
\item{MaxModelAge}{maximum age considered in model}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{Linf}{asymptotic length (von Bertalanffy growth parameter, set to NA if inputting lengths at ages directly from any growth model)}

\item{vbK}{growth coefficient (von Bertalanffy growth parameter, set to NA if inputting lengths at ages directly from any growth model)}

\item{tzero}{hypothetical age at zero length (von Bertalanffy growth parameter, set to NA if inputting lengths at ages directly from any growth model)}

\item{EstLenAtAge}{vector or estimated lengths at ages from any growth model, set to NA if von Bertalanffy growth parameters specified}

\item{lenwt_a}{weight-length parameter (power relationship)}

\item{ln_lenwt_a}{weight-length parameter (log-log relationship)}

\item{lenwt_b}{weight-length parameter (power or log-log relationship)}

\item{WLrel_Type}{1=power, 2=log-log relationship (set to NA if inputting weights at ages directly)}

\item{EstWtAtAge}{vector of weights at ages (set to NA if weight-length growth parameters specified)}

\item{ReprodScale}{<- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{proportion of fish that are females at hatching}

\item{FinalSex_Pmax}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_A50}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_A95}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{mat_A50}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{mat_A95}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{EstMatAtAge}{vector of proportion mature at age (set to NA if using age at maturity parameters)}

\item{sel_A50}{logistic parameter for gear selectivity curve}

\item{sel_A95}{logistic parameter for gear selectivity curve}

\item{EstSelAtAge}{vector of gear selectivity at age (set to NA if using age at gear selectivity parameters)}

\item{ret_Pmax}{logistic parameter for fish retention curve}

\item{ret_A50}{logistic parameter for fish retention curve}

\item{ret_A95}{logistic parameter for fish retention curve}

\item{EstRetenAtAge}{vector of fish retention at age (set to NA if using age at fish retention parameters)}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{Steepness_sd}{specified error for steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{SRrel_Type}{1 = Beverton-Holt, 2=Ricker stock-recruitment relationship}

\item{NatMort}{natural mortality}

\item{NatMort_sd}{specified error for natural mortality}

\item{Current_F}{estimate of current fishing mortality}

\item{Current_F_sd}{error for estimate of current fishing mortality}

\item{nReps}{number of random parameter sets from parametric resampling to generate per recruit outputs with error}

\item{RefPointPlotOpt}{plotting option for reference points, 0=don't plot, 1=plot defaults, 2=plot BMSY ref points}
}
\value{
fishing mortality values for analysis (PerRec_FValues), female SPR estimates vs F (Fem_SPR_Vals)
estimates of equilbrium female spawning biomass vs F (Eq_RelFemSpBiom_Vals), estimated BMSY values (BMSY_Vals)
resampled female SPR values vs F (Sim_FemSPR=Sim_FemSPR), resampled equilbrium female, male and combined sex
spawning biomass values vs F (Sim_Eq_RelFemSpBiom, Sim_Eq_RelMalSpBiom, Sim_Eq_RelCombSexSpBiom),
median female, male and combined sex SPR values vs F (EstFemSPR, EstMalSPR, EstCombSexSPR), median BMSY ratio estimate
(EstMedBMSYratio), summary of estimates with associated 95 percent confidence limits (ResSummary_with_err).
}
\description{
This function provides outputs associated with age-based per recruit analysis, and an
extended form of this analysis with a Beverton-Holt stock recruitment relationship to account
for potential impacts of fishing on recruitment. Outputs are provided for a range of
fishing mortality values, including the current, estimated value, with specified error in M, h and F.
}
\examples{
# Example. Non-hermaphroditic species
InitRecruit <- 1 # Initial recruitment
MaxModelAge <- 20 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be approprriate to use a smaller time step)
Linf <- c(550, 500) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.5, 0.5) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
EstLenAtAge <- data.frame(EstFemLenAtAge=NA, EstMalLenAtAge=NA) # length at age (from age 0), inputted as values in data frame
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtAge <- data.frame(EstFemWtAtAge=NA, EstMalWtAtAge=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at age zero
FinalSex_Pmax <- NA # Logistic sex change relationship parameters (max probability of final sex)
FinalSex_A50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_A95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
mat_A50 <- c(2.5, 2.5) # females, males - Logistic length (mm) at maturity relationship parameters
mat_A95 <- c(3.5, 3.5) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtAge <- data.frame(EstFemMatAtAge=NA, EstMalMatAtAge=NA) # maturity at age (from age 0), inputted as values in data frame
sel_A50 <- c(2.5, 2.5) # females, males - Logistic age selectivity relationship parameters
sel_A95 <- c(3.5, 3.5) # females, males - Logistic age selectivity relationship parameters
EstSelAtAge <- data.frame(EstFemSelAtAge=NA, EstMalSelAtAge=NA) # gear selectivity at age (from age 0), inputted as values in data frame
ret_Pmax <- NA  # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_A50 <- NA  # females, males - Logistic age fish retention at age parameters
ret_A95 <- NA  # females, males - Logistic age fish retention at age parameters
EstRetenAtAge <- data.frame(EstFemRetenAtAge=rep(1,MaxModelAge+1), EstMalRetenAtAge=rep(1,MaxModelAge+1)) # fish retention at age (from age 0), inputted as values in data frame
DiscMort <- 0.0 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
Steepness_sd <- 0.025
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort <- 0.2 # natural mortality  (year-1)
NatMort_sd <- 0.025
Current_F <- 0.1 # estimate of fishing mortality, e.g. from catch curve analysis
Current_F_sd <- 0.005
RefPointPlotOpt <- 1 # 0=don't plot, 1=plot defaults, 2=plot BMSY ref points
nReps = 50
GetPerRecruitResults_AB_with_err(MaxModelAge, TimeStep, Linf, vbK, tzero, EstLenAtAge,
                              lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtAge, ReprodScale, ReprodPattern,
                              InitRatioFem, FinalSex_Pmax, FinalSex_A50, FinalSex_A95, mat_A50, mat_A95,
                              EstMatAtAge, sel_A50, sel_A95, EstSelAtAge, ret_Pmax, ret_A50, ret_A95, EstRetenAtAge,
                              DiscMort, Steepness, Steepness_sd, SRrel_Type, NatMort, NatMort_sd, Current_F,
                              Current_F_sd, nReps)
}
