% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRetainedCatchStats}
\alias{GetRetainedCatchStats}
\title{Get retained catch summary statistics}
\usage{
GetRetainedCatchStats(midpt, RetCatchAtLen, TotCatchAtLen, SampleSize)
}
\arguments{
\item{midpt}{mid points of length classes}

\item{RetCatchAtLen}{discarded catch at length}

\item{TotCatchAtLen}{total catch at length}

\item{SampleSize}{sample sizes}
}
\value{
PropReleased, PropReleased_sd, MeanLenReleased, MeanLenReleased_sd
}
\description{
Calculate prop retained and mean length of retained fish
}
\keyword{internal}
