% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetRetCatchFreqAtLenAndIntAge}
\alias{GetRetCatchFreqAtLenAndIntAge}
\title{Get retained catch frequencies at length in each integer age, and in each age class}
\usage{
GetRetCatchFreqAtLenAndIntAge(
  TimeStep,
  MaxAge,
  midpt,
  nLenCl,
  ObsAgeClRetCatch_Fem,
  ObsAgeClRetCatch_Mal,
  ObsLenClRetCatchMidPt_Fem,
  ObsLenClRetCatchMidPt_Mal
)
}
\arguments{
\item{TimeStep}{model time step}

\item{MaxAge}{maximum age}

\item{midpt}{mid points of length classes}

\item{nLenCl}{number of length classes}

\item{ObsAgeClRetCatch_Fem}{generated female catch in each integer age class}

\item{ObsAgeClRetCatch_Mal}{generated male catch in each integer age class}

\item{ObsLenClRetCatchMidPt_Fem}{generated female catch in each length class}

\item{ObsLenClRetCatchMidPt_Mal}{generated male catch in each length class}
}
\value{
ObsRetCatchFreqAtLengthAndIntAge_Fem, ObsRetCatchFreqAtLengthAndIntAge_Mal, ObsRetCatchFreqAtLengthAndIntAge,
ObsRetCatchFreqAtIntAge_Fem, ObsRetCatchFreqAtIntAge_Mal, ObsRetCatchFreqAtIntAge
}
\description{
Get retained catch frequencies in each of the length classes at each integer age, and in each age class
}
\keyword{internal}
