% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{GetSelectivityAndRetention}
\alias{GetSelectivityAndRetention}
\title{Get fish selectivity and retention vectors}
\usage{
GetSelectivityAndRetention(
  midpt,
  SelectivityType,
  SelParams,
  SelectivityAtLen,
  RetenAtLen,
  RetenParams,
  MLL
)
}
\arguments{
\item{midpt}{mid points of length classes}

\item{SelectivityType}{1=selectivity inputted as vector, 2=logistic selectivity params,  3=logistic selectivity + retention params}

\item{SelParams}{params for gear selectivity}

\item{SelectivityAtLen}{inputted selectivity as length vector}

\item{RetenAtLen}{inputted retention as length vector}

\item{RetenParams}{params for fish retention}

\item{MLL}{minimum legal length}
}
\value{
SelAtLength, RetAtLength
}
\description{
Get fish selectivity and retention vectors, depending on user inputs
}
\keyword{internal}
