% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{Plot_Param_Correlations}
\alias{Plot_Param_Correlations}
\title{Plot parameter correlations estimated by age, length and age-length catch curves}
\usage{
Plot_Param_Correlations(ParamSims, plot_labels = NA)
}
\arguments{
\item{ParamSims}{randomly-generated parameter values}

\item{plot_labels}{names of parameters which can be specified, otherwise set to NA}
}
\value{
plot of parameter correlations
}
\description{
This function produced a plot to visualise parameter correlations for multiple catch
curve models, using the variable ParamSims, which contains randomly-generated values
of the parameters, assuming (prior to any back-transformations), they conform to a
multivariate normal distribution.
}
\examples{
# Simulate data
library(dirmult)
SampleSize=1000 # sample size for retained catches (and same number for released fish, if an MLL is specified)
set.seed(123)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge # natural mortality, in normal space
FishMort = NatMort
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=logistic selectivity params
SelectivityAtLen = NA # selectivity vector
SelParams = c(500, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
# single sex, von Bertalanffy
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.08
GrowthParams = c(Linf, vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)

Ages = Res$ModelDiag$DecAges
ObsAgeFreq  = as.vector(table(factor(Res$ObsDecAgeRetCatch, levels=Ages)))
# Simulate marginal age data from Dirichlet multinomial distribution (single sex)
# J = number of fish sampling events
# K = number of age classes
# n = number of fish sampled from each sampling event
# pi = expected proportion at age
# theta = amount of autocorrelation between ages of fish within sampling events
# Fit catch curve with Dirichlet multinomial distribution objective function (single sex)
set.seed(123)
nSampEvents = 100
nFishPerSampEvent = 10
theta_val = 0.6
simDat = simPop(J=nSampEvents, K=nAges, n=nFishPerSampEvent, pi=as.vector(unlist(Res$ModelDiag$ExpRetCatchPropAtDecAge)), theta=theta_val)
simAges = data.frame(simDat$data)
colnames(simAges)=Ages
ObsAgeFreq = as.vector(colSums(simAges))
# Simulate marginal length data from Dirichlet multinomial distribution (single sex)
midpt = Res$midpt
ExpPropAtLen = Res$ModelDiag$ExpRetCatchPropAtLen_Fem
# plot(midpt, Res$ModelDiag$ExpRetCatchPropAtLen_Fem, "l")
# lines(midpt, Res$ModelDiag$ExpRetCatchPropAtLen_Mal, col="blue") # check that same for both sexes (as single sex model)
res=SimLenFreqDat_DirMultDistn_EqMod(nSampEvents, nFishPerSampEvent, theta_val, midpt, ExpPropAtLen)
ObsRetCatchFreqAtLen = res$simLenFreq

# Age-based catch curve
# Specify catch curve model and required inputs for that model
nSexes = 1
DistnType = 1 # 1 = Multinomial, 2 = Dirichlet multinomial
NatMortType = 1 # 1 = fixed, 2 = estimated with prior
NatMort = exp(1.46 - (1.01 * (log(MaxAge))))
NatMort_sd = NA # standard deviation for natural mortality in log space. Set to NA if NatMortType = 1
Init_FMort = 0.2
Init_SelA50 = 5
Init_SelDelta = 2
params = log(c(Init_FMort, Init_SelA50, Init_SelDelta))
FittedRes=GetLogisticCatchCurveResults(params, nSexes, DistnType, NatMortType, NatMort, NatMort_sd, Ages, ObsAgeFreq)
FittedRes$ParamEst
ParamSims <- FittedRes$ParamSims[,c(1,3,4)]
plot_labels <- c("F","A[50]","A[95]")
Plot_Param_Correlations(ParamSims, plot_labels)
}
