% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotAgeBasedCatchCurveResults_LogSpace}
\alias{PlotAgeBasedCatchCurveResults_LogSpace}
\title{Plot age based catch curve results in log space}
\usage{
PlotAgeBasedCatchCurveResults_LogSpace(
  RecAssump,
  SpecRecAge,
  MinFreq,
  MinAge,
  MaxAge,
  nSexes,
  DistnType,
  NatMortType,
  NatMort,
  NatMort_sd,
  ObsAgeFreq,
  CatchCurveModel,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  ymin,
  xmax,
  xint,
  ymax,
  yint,
  PlotCLs,
  CCResults
)
}
\arguments{
\item{RecAssump}{0=age at peak frequency, 1=age at peak frequency + 1}

\item{SpecRecAge}{specified at at full recruitment, set to NA when RecAssump is set to 0 or 1}

\item{MinFreq}{minimum frequency of fish for including data for old fish}

\item{MinAge}{minimum age}

\item{MaxAge}{maximum age}

\item{nSexes}{1 or 2}

\item{DistnType}{1 = Multinomial, 2 = Dirichlet multinomial}

\item{NatMortType}{1 = fixed, 2 = estimated with prior}

\item{NatMort}{natural mortality, in normal space. If NatMortType = 2, this is expected value of this parameter in normal space}

\item{NatMort_sd}{standard deviation for natural mortality in log space. Set to NA if NatMortType = 1}

\item{ObsAgeFreq}{observed age frequency data}

\item{CatchCurveModel}{1=C&R, 2=Linear, 3=logist. sel.}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{PlotCLs}{logical (TRUE=plot 95 percent confidence limits for lines)}

\item{CCResults}{results of applied catch curve model 1, 2 or 3. Can set to NA.}
}
\value{
plot of expected vs observed proportions at age
}
\description{
This function produces plots of outputs of age-based catch curve analyses in log space
}
\examples{
# simulate data
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = unlist(as.vector(Res$CatchSample)) # input as vector for single sex model
# Specify catch curve model and required inputs for that model
# CatchCurveModel = 1 # Chapman Robson
# nSexes = NA
# DistnType = NA
# NatMortType = NA
# NatMort = NA
# NatMort_sd = NA
# RecAssump = 1
# MinFreq = NA
# CCResults = NA
# CatchCurveModel = 2 # Linear
# nSexes = NA
# DistnType = NA
# NatMortType = NA
# NatMort = NA
# NatMort_sd = NA
# RecAssump = 1
# MinFreq = 1
# CCResults = NA
CatchCurveModel = 3 # Logistic selectivity
nSexes = 1
DistnType = 1 # 1 = multinomial, 2 = Dirichlet multinomial
NatMortType = 1 # 1 = fixed, 2 = estimated with prior
NatMort = exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
NatMort_sd = NA # in log space if NatMort is estimated, otherwise set to NA
RecAssump = NA
NatMort = 0.104
MinFreq = NA #
CCResults = NA
Init_FMort = 0.2
Init_SelA50 = 5
Init_SelA95 = 7
params = log(c(Init_FMort, Init_SelA50, SelA95))
PlotAgeBasedCatchCurveResults_LogSpace(RecAssump, SpecRecAge=NA, MinFreq, MinAge, MaxAge, nSexes, DistnType, NatMortType,
                                       NatMort, NatMort_sd, ObsAgeFreq, CatchCurveModel, MainLabel=NA, xaxis_lab=NA,
                                       yaxis_lab=NA, ymin=NA, xmax=NA, xint=NA, ymax=NA, yint=NA, PlotCLs=T, CCResults=NA)
}
