% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotAgeBasedCatchCurveResults_LogSpace}
\alias{PlotAgeBasedCatchCurveResults_LogSpace}
\title{Plot age based catch curve results in log space}
\usage{
PlotAgeBasedCatchCurveResults_LogSpace(
  RecAssump,
  SpecRecAge,
  MinFreq,
  MinAge,
  MaxAge,
  NatMort,
  ObsAgeFreq,
  CatchCurveModel,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  ymin,
  xmax,
  xint,
  ymax,
  yint,
  PlotCLs
)
}
\arguments{
\item{RecAssump}{0=age at peak frequency, 1=age at peak frequency + 1}

\item{SpecRecAge}{specified at at full recruitment, set to NA when RecAssump is set to 0 or 1}

\item{MinFreq}{minimum frequency of fish for including data for old fish}

\item{MinAge}{minimum age}

\item{MaxAge}{maximum age}

\item{NatMort}{natural mortality}

\item{ObsAgeFreq}{observed age frequency data}

\item{CatchCurveModel}{1=C&R, 2=Linear, 3=logist. sel.}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{PlotCLs}{logical (TRUE=plot 95 percent confidence limits for lines)}
}
\value{
plot of expected vs observed proportions at age
}
\description{
This function produces plots of outputs of age-based catch curve analyses in log space
}
\examples{
# simulate data
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort <- exp(1.46 - (1.01 * (log(MaxAge)))) # i.e. Hoenig's (1983) eqn for fish
FMort = 0.1
ZMort = FMort + NatMort
SelA50 = 6
SelA95 = 8
SampleSize = 1000 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = unlist(as.vector(Res$CatchSample)) # input as vector for single sex model
# Specify catch curve model and required inputs for that model
# CatchCurveModel = 1 # Chapman Robson
# NatMort = NA # Chapman Robson
# RecAssump = 1 # Chapman Robson
# MinFreq = NA # Chapman Robson
# CatchCurveModel = 2 # Linear
# NatMort = NA # Linear
# RecAssump = 1 # Linear
# MinFreq = 1 # Linear
CatchCurveModel = 3 # Logistic selectivity
RecAssump = NA # Logistic selectivity
NatMort = 0.104 # Logistic selectivity
MinFreq = NA # # Logistic selectivity
Init_FMort = 0.2
Init_SelA50 = 5
Init_SelA95 = 7
params = log(c(Init_FMort, Init_SelA50, SelA95))
PlotAgeBasedCatchCurveResults_LogSpace(RecAssump, SpecRecAge=NA, MinFreq, MinAge, MaxAge, NatMort,
                                       ObsAgeFreq, CatchCurveModel, MainLabel=NA,
                                       xaxis_lab=NA, yaxis_lab=NA, ymin=NA, xmax=NA, xint=NA,
                                       ymax=NA, yint=NA, PlotCLs=T)
}
