% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotAgeLengthCatchCurve_MargLength}
\alias{PlotAgeLengthCatchCurve_MargLength}
\title{Show fit of age and length catch curve model to marginal length composition}
\usage{
PlotAgeLengthCatchCurve_MargLength(
  params,
  MLL,
  SelectivityType,
  ObsCatchFreqAtLen,
  ObsCatchFreqAtLengthAndAge,
  lbnd,
  ubnd,
  midpt,
  SelectivityVec,
  DiscMort,
  MaxAge,
  NatMort,
  TimeStep,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  xmax,
  xint,
  ymax,
  yint,
  PlotCLs,
  FittedRes,
  nReps
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated (if FittedRes=NA)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{ObsCatchFreqAtLen}{observed frequencies in length classes}

\item{ObsCatchFreqAtLengthAndAge}{observed frequencies in length and age classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityVec}{selectivity at length}

\item{DiscMort}{Proportion of fish that die following to capture and release}

\item{MaxAge}{maximum age considered in model}

\item{NatMort}{natural mortality}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{PlotCLs}{logical (TRUE=plot 95 percent confidence limits for line)}

\item{FittedRes}{saved results from GetLengthBasedCatchCurveResults function (model will be refitted if set to NA)}

\item{nReps}{number of random parameter sets from parametric resampling to generate outputs with error}
}
\value{
Plot of fitted age and length-based catch curve to marginal length composition
}
\description{
This function provides a plot of the fit of an age and length-based catch curve model, with length-
based selectivity to marginal length composition data. The model is fitted to a sample of fish
length and age data, by minimising the overall negative log-likelihood, including the NLL associated
with the marginal length composition and a conditional age at length NLL,
given the parameters (selectivity, growth and mortality) and data, using nlminb.
It provides various statistical outputs in include convergence statistics, parameter estimates
and associated 95 percent confidence limits and associated variance-covariance matrix, calculated using
the MASS package.
}
\examples{
set.seed(123)
# Simulate data
SampleSize=5000
set.seed(123)
MaxAge = 26
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
MinAge = floor(TimeStep)
nAgeCl = length(MinAge:MaxAge)
nTimeSteps = length(seq(TimeStep,MaxAge,TimeStep))
FishMort = 0.2
MaxLen = 1100
LenInc = 10
MLL = NA
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
L50 = 400 # selectivity
L95 = 450 # selectivity
SelectivityVec = NA
DiscMort = 0
# # single sex, von Bertalanffy
# GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute (no yet implemented for this method)
# Linf = 800
# vbK = 0.2
# CVSizeAtAge = 0.08
# GrowthParams = c(Linf, vbK)
# RefnceAges = NA
# 2 sexes, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute (no yet implemented for this method)
Linf = c(700,850)
vbK = c(0.25,0.2)
CVSizeAtAge = c(0.05,0.05)
RefnceAges = NA
GrowthParams = data.frame(Linf=Linf, vbK=vbK)
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         L50, L95, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
lbnd=Res$lbnd
midpt=Res$midpt
ubnd=Res$ubnd
# # get data - 1 sex (or combined sexes)
# ObsCatchFreqAtLen = Res$ObsCatchFreqAtLen # 1 sex
# ObsCatchFreqAtLengthAndAge = as.matrix(Res$ObsCatchFreqAtLengthAndDecAge) # 1 sex
# get data - 2 sexes
ObsCatchFreqAtLen <- data.frame(matrix(nrow = 2, ncol = length(midpt))) # 2 sex
colnames(ObsCatchFreqAtLen) <- midpt
ObsCatchFreqAtLen[1,] = Res$ObsCatchFreqAtLen_Fem
ObsCatchFreqAtLen[2,] = Res$ObsCatchFreqAtLen_Mal
ObsCatchFreqAtLengthAndAge = array(c(unlist(Res$ObsCatchFreqAtLengthAndDecAge_Fem), unlist(Res$ObsCatchFreqAtLengthAndDecAge_Mal)),
                                   c(nTimeSteps, length(midpt), 2), dimnames=list(rownames(Res$ObsCatchFreqAtLengthAndDecAge_Fem),
                                                                                  colnames(Res$ObsCatchFreqAtLengthAndDecAge_Fem)))
# # get params - 1 sex
# InitFishMort = 0.3 # specify starting parameters
# InitL50 = 300
# InitL95 = 500
# InitLinf = 800
# InitvbK = 0.2
# InitCVSizeAtAge = 0.05
# get params - 2 sexes
InitFishMort = 0.3 # specify starting parameters
InitL50 = 300
InitL95 = 500
InitLinf = c(800,800)
InitvbK = c(0.25,0.25)
InitCVSizeAtAge = 0.05
InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform (so F is always between 0 and 1)
params = c(InitFishMort_logit, log(c(InitL50, InitL95, InitLinf, InitvbK, InitCVSizeAtAge)))
FittedRes=GetAgeAndLengthBasedCatchCurveResults(params, MLL, SelectivityType, ObsCatchFreqAtLen, ObsCatchFreqAtLengthAndAge,
                                                lbnd, ubnd, midpt, SelectivityVec, DiscMort, MaxAge, NatMort, TimeStep)
# Plot. Note, can skip above step and set FittedRes=NA (plot function will be slower)
PlotAgeLengthCatchCurve_MargLength(params, MLL, SelectivityType, ObsCatchFreqAtLen, ObsCatchFreqAtLengthAndAge,
                                  lbnd, ubnd, midpt, SelectivityVec, DiscMort, MaxAge, NatMort, TimeStep, MainLabel=NA,
                                  xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA,
                                  ymax=0.10, yint=0.02, PlotCLs=TRUE, FittedRes, nReps=200)
}
