% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotLengthBasedCatchCurve_RetCatch}
\alias{PlotLengthBasedCatchCurve_RetCatch}
\title{Produce plot of fitted length-based catch curve to length composition data for retained catches}
\usage{
PlotLengthBasedCatchCurve_RetCatch(
  params,
  MLL,
  SelectivityType,
  ObsRetCatchFreqAtLen,
  lbnd,
  ubnd,
  midpt,
  SelectivityVec,
  PropReleased,
  ObsDiscCatchFreqAtLen,
  DiscMort,
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  MaxAge,
  NatMort,
  TimeStep,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  xmax,
  xint,
  ymax,
  yint,
  PlotCLs,
  FittedRes,
  nReps
)
}
\arguments{
\item{params}{vector of model parameters (in log space) to be estimated (if FittedRes=NA)}

\item{MLL}{minimum legal length (for setting knife edge retention, set to NA if not assumed)}

\item{SelectivityType}{1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve}

\item{ObsRetCatchFreqAtLen}{observed frequencies in length classes}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points of length classes}

\item{SelectivityVec}{selectivity at length}

\item{PropReleased}{proportion of fish that are released}

\item{ObsDiscCatchFreqAtLen}{Mean length of fish that are released}

\item{DiscMort}{selectivity at length}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{c(Linf, vbK) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK) separate sex von Bertalanffy,
c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b) separate sex Schnute}

\item{RefnceAges}{Schnute reference ages, either c(t1,t2) single sex, c(t1=t1, t2=t2) separate sex, or set to NA if using von Bertalanffy model}

\item{MaxAge}{maximum age considered in model}

\item{NatMort}{natural mortality}

\item{TimeStep}{model timestep (e.g. 1 = annual, 1/12 = monthly)}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{PlotCLs}{logical (TRUE=plot 95 percent confidence limits for line)}

\item{FittedRes}{saved results from GetLengthBasedCatchCurveResults function (model will be refitted if set to NA)}

\item{nReps}{number of random parameter sets from parametric resampling to generate outputs with error}
}
\value{
plot of fitted length-based catch curve to length composition data for retained fish
}
\description{
Produce plot of fitted length-based catch curve to length composition data for retained catches
}
\examples{
# Simulate data
SampleSize=5000
set.seed(123)
MaxAge = 40
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
FishMort = 0.15
MaxLen = 1200
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified is knife-edged at MLL
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityVec = NA # selectivity vector
SelParams = c(300, 500) # L50, L95 for gear selectivity
RetenParams = c(NA, NA) # L50, L95 for retention
DiscMort = 0.5
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 900
vbK = 0.11
CVSizeAtAge = 0.08
GrowthParams = c(Linf, vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
ObsRetCatchFreqAtLen = Res$ObsRetCatchFreqAtLen
ObsDiscCatchFreqAtLen = Res$ObsDiscCatchFreqAtLen
PropReleased=NA
midpt=Res$midpt
lbnd=Res$lbnd
ubnd=Res$ubnd
InitFishMort = 0.25 # specify starting parameters
InitFishMort_logit = log(InitFishMort/(1-InitFishMort)) # logit transform
InitL50 = 300
InitDelta = 100
params = c(InitFishMort_logit, log(InitL50), log(InitDelta))
FittedRes=GetLengthBasedCatchCurveResults(params, GrowthCurveType, GrowthParams, RefnceAges, MLL, SelectivityType, ObsRetCatchFreqAtLen,
                                          lbnd, ubnd, midpt, SelectivityVec, PropReleased, ObsDiscCatchFreqAtLen, DiscMort, CVSizeAtAge,
                                          MaxAge, NatMort, TimeStep)
PlotLengthBasedCatchCurve_RetCatch(params, MLL, SelectivityType, ObsRetCatchFreqAtLen, lbnd, ubnd, midpt,
                                 SelectivityVec, PropReleased, ObsDiscCatchFreqAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, MaxAge, NatMort, TimeStep, MainLabel=NA,
                                 xaxis_lab=NA, yaxis_lab=NA, xmax=1500, xint=500,
                                 ymax=0.15, yint=0.05, PlotCLs=TRUE, FittedRes, nReps=200)
}
