% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{PlotPerRecruit_Biom_with_err_LB}
\alias{PlotPerRecruit_Biom_with_err_LB}
\title{Plot of SPR and relative equilibrium biomass vs F from length-based per recruit analysis and extended analysis
with a stock-recruitment relationship, with error for M, h and F}
\usage{
PlotPerRecruit_Biom_with_err_LB(
  MaxModelAge,
  TimeStep,
  Linf,
  vbK,
  tzero,
  EstLenAtAge,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtAge,
  ReprodScale,
  ReprodPattern,
  InitRatioFem,
  FinalSex_Pmax,
  FinalSex_A50,
  FinalSex_A95,
  mat_A50,
  mat_A95,
  EstMatAtAge,
  sel_A50,
  sel_A95,
  EstSelAtAge,
  ret_Pmax,
  ret_A50,
  ret_A95,
  EstRetenAtAge,
  DiscMort,
  Steepness,
  Steepness_sd,
  SRrel_Type,
  NatMort,
  NatMort_sd,
  Current_F,
  Current_F_sd,
  PlotOpt,
  RefPointPlotOpt,
  FittedRes,
  nReps,
  MainLabel,
  xaxis_lab,
  yaxis_lab,
  xmax,
  xint,
  ymax,
  yint
)
}
\arguments{
\item{MaxModelAge}{maximum age considered by model}

\item{TimeStep}{model time step (in y)}

\item{lenwt_a}{weight-length parameter}

\item{ln_lenwt_a}{weight-length parameter}

\item{lenwt_b}{weight-length parameter}

\item{WLrel_Type}{1=power, 2=log-log}

\item{ReprodScale}{<- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{proportion of fish that are females at hatching}

\item{FinalSex_Pmax}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{ret_Pmax}{logistic parameter for fish retention curve}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{Steepness_sd}{standard deviation for steepness parameter}

\item{SRrel_Type}{1 = Beverton-Holt, 2=Ricker stock-recruitment relationship}

\item{NatMort}{natural mortality}

\item{NatMort_sd}{standard deviation for natural mortality}

\item{Current_F}{estimated current fishing mortality}

\item{Current_F_sd}{standard deviation for estimated current fishing mortality}

\item{RefPointPlotOpt}{# 0=don't plot, 1=plot defaults, 2=plot BMSY ref points}

\item{FittedRes}{option to input per recruit results into plot function to increase speed, can set to NA}

\item{nReps}{number of resamping trials}

\item{MainLabel}{plot label}

\item{xaxis_lab}{y axis label}

\item{yaxis_lab}{x axis label}

\item{xmax}{maximum x axis value}

\item{xint}{x axis interval}

\item{ymax}{maximum y axis value}

\item{yint}{y axis interval}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points bounds of length classes}

\item{nLenCl}{number of length classes}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{growth parameters of either von Bertalanffy or Schnute model}

\item{RefnceAges}{reference ages for Schnute model, set to NA if using von Bertalanffy model}

\item{CVSizeAtAge}{coefficient of variation for size at age}

\item{EstWtAtLen}{user-specified weights at lengths}

\item{FinalSex_L50}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_L95}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{mat_L50}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{mat_L95}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{EstMatAtLen}{vector of proportion mature at length (set to NA if using age at maturity parameters)}

\item{sel_L50}{logistic parameter for gear selectivity curve}

\item{sel_L95}{logistic parameter for gear selectivity curve}

\item{ret_L50}{logistic parameter for fish retention curve}

\item{ret_L95}{logistic parameter for fish retention curve}
}
\description{
This function provides a plot of SPR and relative equilibrium biomass vs F from length-based per recruit analysis
and extended analysis with a stock-recruitment relationship, with specified error for M, h and F
}
\examples{
MaxModelAge <- 20 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be appropriate to use a smaller time step)
MaxLen = 800
LenInc = 20
lbnd = seq(0,MaxLen - LenInc, LenInc)
ubnd = lbnd + LenInc
midpt = lbnd + (LenInc/2)
nLenCl = length(midpt)
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf <- c(550, 500) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.2, 0.2) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
GrowthParams = data.frame(Linf=Linf, vbK=vbK, tzero=tzero)
RefnceAges = NA
# GrowthParams c(Linf, vbK, tzero) single sex von Bertalanffy, or data.frame(Linf=Linf, vbK=vbK, tzero=tzero),
#' both sexes von Bertalanffy, or c(y1, y2, a, b) single sex Schnute, or data.frame(y1=y1, y2=y2, a=a, b=b), both sexes Schnute
CVSizeAtAge = c(0.05, 0.05)
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtLen <- data.frame(EstFemWtAtLen=NA,
                         EstMalWtAtLen=NA) # weight at age (from age 0), inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at age zero
FinalSex_Pmax <- NA # Logistic sex change relationship parameters (max probability of final sex)
FinalSex_L50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_L95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
mat_L50 <- c(250, 250) # females, males - Logistic length (mm) at maturity relationship parameters
mat_L95 <- c(300, 300) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtLen <- data.frame(EstFemMatAtLen=NA,
                          EstMalMatAtLen=NA) # maturity at age (from age 0), inputted as values in data frame
sel_L50 <- c(250, 250) # females, males - Logistic age selectivity relationship parameters
sel_L95 <- c(300, 300) # females, males - Logistic age selectivity relationship parameters
ret_Pmax <- c(1.0, 1.0) # maximum retention, values lower than 1 imply discarding of fish above MLL
ret_L50 <- c(250, 250) # females, males - Logistic age fish retention at age parameters
ret_L95 <- c(300, 300) # females, males - Logistic age fish retention at age parameters
DiscMort <- 0.25 # discard mortality (e.g. 50\% released fish die = 0.5)
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
Steepness_sd <- 0.025
NatMort <- 0.2 # natural mortality  (year-1)
NatMort_sd <- 0.025
Current_F <- 0.1 # estimate of fishing mortality, e.g. from catch curve analysis
Current_F_sd <- 0.005
PlotOpt <- 1 # 1=females, 2=males, 3=combined sex
RefPointPlotOpt <- 1 # 0=don't plot, 1=plot defaults, 2=plot BMSY ref points
nReps = 10 # number of resampling trials. Set to low number to test, then much higher for final analysis.
FittedRes=GetPerRecruitResults_LB_with_err(MaxModelAge, TimeStep, lbnd, ubnd, midpt, nLenCl, GrowthCurveType, GrowthParams,
                                     RefnceAges, CVSizeAtAge, lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type,
                                     EstWtAtLen, ReprodScale, ReprodPattern, InitRatioFem, FinalSex_Pmax, FinalSex_L50,
                                     FinalSex_L95, mat_L50, mat_L95, EstMatAtLen, sel_L50, sel_L95, ret_Pmax,
                                      ret_L50, ret_L95, DiscMort, Steepness, Steepness_sd, SRrel_Type, NatMort, NatMort_sd,
                                      Current_F, Current_F_sd, nReps)
# Plot. Note, can skip above step and set FittedRes=NA (plot function will be slower
PlotPerRecruit_Biom_with_err_LB(MaxModelAge, TimeStep, Linf, vbK, tzero, EstLenAtAge,
                                            lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtAge, ReprodScale, ReprodPattern,
                                            InitRatioFem, FinalSex_Pmax, FinalSex_A50, FinalSex_A95, mat_A50, mat_A95,
                                            EstMatAtAge, sel_A50, sel_A95, EstSelAtAge, ret_Pmax, ret_A50, ret_A95,
                                            EstRetenAtAge, DiscMort, Steepness, Steepness_sd, SRrel_Type, NatMort, NatMort_sd,
                                            Current_F, Current_F_sd, PlotOpt, RefPointPlotOpt, FittedRes, nReps, MainLabel=NA,
                                            xaxis_lab=NA, yaxis_lab=NA, xmax=NA, xint=NA, ymax=NA, yint=NA)
}
