% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3Analyses.R
\name{SimLenAndAgeFreqData_DynMod_ALB}
\alias{SimLenAndAgeFreqData_DynMod_ALB}
\title{Simulate length composition data from a dynamic age-based model allowing for length and recruitment variation}
\usage{
SimLenAndAgeFreqData_DynMod_ALB(
  SimAnnSampSize,
  nYears,
  lnSigmaR,
  autocorr,
  InitRec,
  MaxModelAge,
  TimeStep,
  lbnd,
  ubnd,
  midpt,
  nLenCl,
  GrowthCurveType,
  GrowthParams,
  RefnceAges,
  CVSizeAtAge,
  lenwt_a,
  ln_lenwt_a,
  lenwt_b,
  WLrel_Type,
  EstWtAtLen,
  ReprodScale,
  ReprodPattern,
  InitRatioFem,
  FinalSex_L50,
  FinalSex_L95,
  EstSexRatioAtLen,
  EggFertParam,
  mat_L50,
  mat_L95,
  EstMatAtLen,
  sel_L50,
  sel_L95,
  EstGearSelAtLen,
  ret_Pmax,
  ret_L50,
  ret_L95,
  EstRetenAtLen,
  DiscMort,
  Steepness,
  SRrel_Type,
  NatMort,
  FMortByYear
)
}
\arguments{
\item{SimAnnSampSize}{specified sample size (for both sexes)}

\item{nYears}{number of years of deviations required}

\item{lnSigmaR}{natural logarithm of standard deviation of annual recruitment deviations}

\item{autocorr}{level of autocorrelation of annual recruitment deviations}

\item{InitRec}{initial recruitment (R0)}

\item{MaxModelAge}{maximum age considered by model}

\item{TimeStep}{model time step (in y)}

\item{lbnd}{lower bounds of length classes}

\item{ubnd}{upper bounds of length classes}

\item{midpt}{mid points bounds of length classes}

\item{nLenCl}{number of length classes}

\item{GrowthCurveType}{1=von Bertalanffy, 2=Schnute}

\item{GrowthParams}{growth parameters of either von Bertalanffy or Schnute model}

\item{RefnceAges}{reference ages for Schnute model, set to NA if using von Bertalanffy model}

\item{CVSizeAtAge}{coefficient of variation for size at age}

\item{lenwt_a}{weight-length parameter}

\item{ln_lenwt_a}{weight-length parameter}

\item{lenwt_b}{weight-length parameter}

\item{WLrel_Type}{1=power, 2=log-log}

\item{EstWtAtLen}{user-specified weights at lengths}

\item{ReprodScale}{<- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)}

\item{ReprodPattern}{reproductive pattern, 1=gonochoristic, 2=protogynous (female to male sex change) hermaphroditism,
3=protandrous hermaphroditism (male to female sex change)}

\item{InitRatioFem}{proportion of fish that are females at hatching}

\item{FinalSex_L50}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{FinalSex_L95}{logistic sex change parameter for hermaphroditic species (set to NA for gonochoristic species)}

\item{EstSexRatioAtLen}{NA  # sex ratio at length, inputted as vector}

\item{EggFertParam}{<- NA # (NA or from ~0.2-1) NA = no effect, 0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects}

\item{mat_L50}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{mat_L95}{logistic length at maturity parameter (set to NA if directly inputting proportion mature at age)}

\item{EstMatAtLen}{vector of proportion mature at length (set to NA if using age at maturity parameters)}

\item{sel_L50}{logistic parameter for gear selectivity curve}

\item{sel_L95}{logistic parameter for gear selectivity curve}

\item{EstGearSelAtLen}{gear selectivity curve inputted as vector}

\item{ret_Pmax}{logistic parameter for fish retention curve}

\item{ret_L50}{logistic parameter for fish retention curve}

\item{ret_L95}{logistic parameter for fish retention curve}

\item{EstRetenAtLen}{retention curve inputted as vector}

\item{DiscMort}{proportion of fish that die following capture and release}

\item{Steepness}{steepness parameter of Beverton-Holt or Ricker stock-recruitment relationship}

\item{SRrel_Type}{1 = Beverton-Holt, 2=Ricker stock-recruitment relationship}

\item{NatMort}{natural mortality}

\item{FMortByYear}{annual fully-selected fishing mortality}
}
\value{
ModelDiag (including FMortByYear, random_dev, FemSpBiom, MalSpBiom, CombSexSpBiom, AnnRecruit,
RetCatch_Fem_yal, RetCatch_Mal_yal, DiscCatch_Fem_yal, DiscCatch_Mal_yal, SampleSize_Fem, SampleSize_Mal,
SampleSize, DiscSampleSize_Fem, DiscSampleSize_Mal, DiscSampleSize), ResultsSummary (including ObsRetCatchFreqAtLen_Fem,
ObsRetCatchFreqAtLen_Mal, ObsDiscCatchFreqAtLen_Fem, ObsDiscCatchFreqAtLen_Mal, ObsRetCatchFreqAtDecAge_Fem,
ObsRetCatchFreqAtDecAge_Mal, ObsRetCatchFreqAtDecAge, ObsDiscCatchFreqAtDecAge_Fem, ObsDiscCatchFreqAtDecAge_Mal,
ObsDiscCatchFreqAtDecAge, ObsRetCatchFreqAtLengthAndDecAge_Fem, ObsRetCatchFreqAtLengthAndDecAge_Mal,
ObsRetCatchFreqAtLengthAndDecAge, ObsDiscCatchFreqAtLengthAndDecAge_Fem, ObsDiscCatchFreqAtLengthAndDecAge_Mal,
ObsDiscCatchFreqAtLengthAndDecAge, ObsDecAgeRetCatch_Fem, ObsDecAgeRetCatch_Mal, ObsDecAgeRetCatch, ObsDecAgeDiscCatch_Fem,
ObsDecAgeDiscCatch_Mal, ObsDecAgeDiscCatch, ObsAgeClRetCatch_Fem, ObsAgeClRetCatch_Mal, ObsAgeClRetCatch,
ObsAgeClDiscCatch_Fem, ObsAgeClDiscCatch_Mal, ObsAgeClDiscCatch, ObsLenClRetCatchMidPt_Fem, ObsLenClRetCatchMidPt_Mal,
ObsLenClRetCatchMidPt, ObsLenClDiscCatchMidPt_Fem, ObsLenClDiscCatchMidPt_Mal, ObsLenClDiscCatchMidPt
}
\description{
This function provides simulated age, length, and conditional age-at-length data from a dynamic age-based model allowing
for length variation, with specified annual fishing and natural mortality, and annual recruitment with specified process
error and autocorrelation.
}
\examples{
# Simulate length data from dynamic age-length model, given specified biology and exploitation history
MaxModelAge <- 60 # maximum age considered by model, years
TimeStep <- 1 # Model time step (y) (for shorter-lived species, might be appropriate to use a smaller time step)
MaxLen <- 1400
LenInc <- 20
lbnd <- seq(0,MaxLen - LenInc, LenInc)
ubnd <- lbnd + LenInc
midpt <- lbnd + (LenInc/2)
nLenCl <- length(midpt)
GrowthCurveType <- 1 # 1 = von Bertalanffy, 2 = Schnute
Linf <- c(1000, 1000) # mm - von Bertalanffy growth model parameters - Females, males
vbK <- c(0.12, 0.12) # year-1 - von Bertalanffy growth model parameters - Females, males
tzero <- c(0, 0) # years - von Bertalanffy growth model parameters - Females, males
GrowthParams <- data.frame(Linf=Linf, vbK=vbK, tzero=tzero)
RefnceAges <- NA
CVSizeAtAge = c(0.06,0.06)
lenwt_a <- 0.000005 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
ln_lenwt_a <- NA # for log-log relationship
lenwt_b <- 3 # combined sexes - weight (g) vs length (mm, TL) relationship parameters
WLrel_Type <- 1 # 1=power, 2=log-log relationship
EstWtAtLen <- data.frame(EstFemWtAtLen=NA,
                         EstMalWtAtLen=NA) # weight at length, inputted as values in data frame
ReprodScale <- 1 # 1=default (standard calculations for spawning biomass), 2=hyperallometric reproductive scaling with female mass (i.e. BOFFF effects)
ReprodPattern <- 1 # 1 = gonochoristic (separate sexes), 2 = protogynous (female to male sex change), 3 = protandrous (male to female sex change)
InitRatioFem <- 0.5 # Ratio of females to males at recruitment age/length
FinalSex_L50 <- NA # Logistic sex change relationship parameters (inflection point)
FinalSex_L95 <- NA # Logistic sex change relationship parameters (95\% of max probability)
EstSexRatioAtLen <- NA  # sex ratio at length inputted as vector
EggFertParam <- NA # (NA or from ~0.2-1) NA = no effect, ~0.2 = direct effect of popn. sex ratio changes on egg fertilisation rates, 1 = no effects
mat_L50 <- c(350, 350) # females, males - Logistic length (mm) at maturity relationship parameters
mat_L95 <- c(400, 400) # females, males - Logistic length (mm) at maturity relationship parameters
EstMatAtLen <- data.frame(EstFemMatAtLen=NA,
                          EstMalMatAtLen=NA) # maturity at length, inputted as values in data frame
sel_L50 <- c(350, 350) # females, males - Logistic length selectivity relationship parameters
sel_L95 <- c(400, 400) # females, males - Logistic length selectivity relationship parameters
EstGearSelAtLen <- data.frame(EstFemGearSelAtLen=NA,
                              EstMalGearSelAtLen=NA)
ret_Pmax <- c(1.0, 1.0) # maximum retention, values lower than 1 imply discarding of fish at large sizes
ret_L50 <- c(490, 490) # females, males - Logistic fish retention at length parameters
ret_L95 <- c(510, 510) # females, males - Logistic fish retention at length parameters
EstRetenAtLen <- data.frame(EstFemRetenAtLen=NA,
                            EstMalRetenAtLen=NA)
DiscMort <- 0.5 # discard mortality (e.g. 50\% released fish die = 0.5)
Steepness <- 0.75 # steepness parameter of the Beverton and Holt stock-recruitment relationship
SRrel_Type <- 1 # 1 = Beverton-Holt, 2=Ricker
NatMort <- 4.22 / 41 # natural mortality  (year-1)
InitRec <- 1000 # initial recruitment level
lnSigmaR <- 0.0 # annual recruitment variation
autocorr <- 0.0 # annual recruit autocorrelation
# Specify exploitation history - initial in increase F, reduction in F then stable F period
nPeriods <- 3 # number of periods of mortality trends
InitYr <- 1970 # first year of mortality
PeriodEndYr <- c(1985,1995,2020) # periods when mortality changes
InitYr_FMort <- 0.33*NatMort # initial mortality, at beginning of first period
PeriodEndYr_FMort <- c(3*NatMort,0.67*NatMort,0.67*NatMort)
F_autocorr <- 0.0 # annual fishing mortality autocorrelation
F_randerr_sd <- 0.01 # annual fishing mortality variation
set.seed(123)
res=SimExploitationHistory_DynMod(nPeriods, InitYr, PeriodEndYr, InitYr_FMort,
                                  PeriodEndYr_FMort, F_autocorr, F_randerr_sd)
# Get random age and length data
nYears <- res$nYears
FMortByYear <- res$rand_FMort
SimAnnSampSize <- 5000
res=SimLenAndAgeFreqData_DynMod_ALB(SimAnnSampSize, nYears, lnSigmaR, autocorr, InitRec,
                                    MaxModelAge, TimeStep, lbnd, ubnd, midpt, nLenCl, GrowthCurveType, GrowthParams,
                                    RefnceAges, CVSizeAtAge, lenwt_a, ln_lenwt_a, lenwt_b, WLrel_Type, EstWtAtLen,
                                    ReprodScale, ReprodPattern, InitRatioFem, FinalSex_L50, FinalSex_L95, EstSexRatioAtLen,
                                    EggFertParam, mat_L50, mat_L95, EstMatAtLen, sel_L50, sel_L95, EstGearSelAtLen, ret_Pmax,
                                    ret_L50, ret_L95, EstRetenAtLen, DiscMort, Steepness, SRrel_Type, NatMort, FMortByYear)
# Generate some random cpue data
set.seed(123)
q <- 0.001
cv <- 0.2
ExpCPUE <- q * res$ModelDiag$CombSexVulnBiom
log_sd <- sqrt(log(1 + cv^2))
RandCPUE <- ExpCPUE * exp(rnorm(length(ExpCPUE), mean = 0, sd = log_sd))
RandCPUElw <- RandCPUE * exp(-1.96 * log_sd)
RandCPUEup <- RandCPUE * exp(1.96 * log_sd)
# plot randomly-generated cpue data
par(mfrow=c(1,1))
plot(1:nYears, RandCPUE, "p", pch = 16, ylim = c(0, 1.2*max(RandCPUEup)), bty='n',
     xlab = "Year", ylab = "CPUE", main = "Rand CPUE")
arrows(1:nYears, RandCPUElw, 1:nYears,RandCPUEup, angle = 90, code = 3, length = 0.05)
lines(1:nYears, ExpCPUE, col = "blue", lty = 2, lwd = 2)
}
